% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.acts.WHAMP.R
\name{acts_msm_whamp}
\alias{acts_msm_whamp}
\title{Sexual Acts Module}
\usage{
acts_msm_whamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with the updated discordant act
list (\code{dal}). Each element of \code{dal} is a data frame with the ids of the
discordant pair repeated the number of times they have AI.
}
\description{
Module function for setting the number of sexual acts on the
             discordant edgelist for the WHAMP model.
}
\details{
The number of acts at each time step is specified as a function of the expected values 
specific to each partnership type. For one-off partnerships, this is deterministically
set at 1, whereas for main and persistent partnerships it is a stochastic draw
from a Poisson distribution. The number of total acts may further be modified
by the level of HIV viral suppression in an infected person.
}
\keyword{module}
\keyword{msm}
