% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.acts.shamp.R
\name{acts_shamp}
\alias{acts_shamp}
\title{Sexual Acts Module for up to 5 race groups and 2 sexes.}
\usage{
acts_shamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with the updated discordant act
list (\code{dal}). Each element of \code{dal} is a data frame with the ids of the
discordant pair repeated the number of times they have AI.
}
\description{
Module function for setting the number of sexual acts on the
             discordant edgelist.
}
\details{
The number of acts at each time step is specified as a function of the race of
both members in a pair and whether the relationship is male male
or heterosexual.  Rates are averages of the reported rates by each race.
For one-off partnerships, this is deterministically
set at 1, whereas for main and causal partnerships it is a stochastic draw
from a Poisson distribution. The number of total acts may further be modified
by the level of HIV viral suppression in an infected person.
}
\keyword{SHAMP}
\keyword{het}
\keyword{module}
\keyword{msm}
