% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{add_nwstats_het}
\alias{add_nwstats_het}
\title{Add Heterosexual Target Statistics for Network Model Estimation}
\usage{
add_nwstats_het(nwstats, num.B.het = 9000, num.W.het = 9000,
  male.prop = 0.5, msmw.prop = 0.5, ...)
}
\arguments{
\item{...}{}
}
\description{
Calculates the target statistics for the formation and dissolution
             components of the network model to be estimated with \code{netest}.
}
