% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.aging.camplc.R
\name{aging_camplc}
\alias{aging_camplc}
\title{Aging Module}
\usage{
aging_camplc(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns \code{dat} after updating the nodal attribute
\code{age} and \code{sqrt.age}.
}
\description{
Module for aging over time for active nodes in the population.
}
\keyword{module}
