% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{assign_degree_msmhet}
\alias{assign_degree_msmhet}
\title{Assign Degree Vertex Attribute on Network Objects}
\usage{
assign_degree_msmhet(nw, deg.type, nwstats)
}
\arguments{
\item{nw}{Object of class \code{network} that is the target for the vertex
attribute.}

\item{deg.type}{Type of degree to assign to \code{nw}, with options of
\code{"pers"} to assign casual degree onto main network and
\code{"main"} to assign main degree to casual network.}

\item{nwstats}{Object of class \code{nwstats}.}
}
\description{
Assigns the degree vertex attributes on network objects
             conditional on their values from the other networks.
}
