% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimation.R
\name{base_nw.mard}
\alias{base_nw.mard}
\title{Make Base Population}
\usage{
base_nw.mard(nwstats)
}
\arguments{
\item{nwstats}{Output from \code{\link{calc_nwstats.mard}}.}
}
\description{
description
}
\details{
This function ...
}

