% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{base_nw_msm}
\alias{base_nw_msm}
\title{Construct Base Network for Model Estimation and Simulation}
\usage{
base_nw_msm(nwstats)
}
\arguments{
\item{nwstats}{An object of class \code{nwstats}, as output from
\code{\link{calc_nwstats_msm}}.}
}
\description{
Initializes the base network for model estimation within
             \code{netest}.
}
\details{
This function takes the output of \code{\link{calc_nwstats_msm}} and constructs
an empty network with the necessary attributes for race, square root of age,
and sexual role class. This base network is used for all three network
estimations.
}
\seealso{
The final vertex attributes on the network for cross-network degree are
calculated and set on the network with \code{\link{assign_degree}}.
}
\keyword{msm}
