% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{base_nw_msm_het}
\alias{base_nw_msm_het}
\title{Construct Base Network for Model Estimation and Simulation}
\usage{
base_nw_msm_het(nwstats)
}
\arguments{
\item{nwstats}{An object of class \code{nwstats}, as output from
\code{\link{calc_nwstats_msm}}.}
}
\description{
Initializes the base network for model estimation within
             \code{netest}.
}
