% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.births.R
\name{births_het}
\alias{births_het}
\title{Births Module}
\usage{
births_het(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module for simulating births/entries into the population, including
             initialization of attributes for incoming nodes.
}
\keyword{het}
\keyword{module}
