% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.births.R
\name{births_msm}
\alias{births_msm}
\title{Births Module}
\usage{
births_msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function updates the \code{attr} list with new attributes for each new
population member, and the \code{nw} objects with new vertices.
}
\description{
Module function for births or entries into the sexually active
             population.
}
\details{
New population members are added based on expected numbers of entries among
black and white MSM, stochastically determined with draws from Poisson
distributions. For each new entry, a set of attributes is added for that node,
and the nodes are added onto the network objects. Only attributes that are
a part of the network model formulae are updated as vertex attributes on the
network objects.
}
\keyword{module}
\keyword{msm}
