% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.births.WHAMP.R
\name{births_msm_whamp}
\alias{births_msm_whamp}
\title{Births Module}
\usage{
births_msm_whamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function updates the \code{attr} list with new attributes for each new
population member, and the \code{nw} objects with new vertices.
}
\description{
Module function for births or entries into the sexually active
             population for the WHAMP model
}
\details{
New population members are added based on expected numbers of entries,
stochastically determined with draws from Poisson distributions. The 
expected number of entries is set to balance exits due to aging out and  
background mortality, defined as the number of deaths among HIV-negative men plus
the expected number of deaths among HIV+ if they were subject to general pop
age-specific mortality (i.e. subtracting excess mortality due to disease).
For each new entry, a set of attributes is added for that node,
and the nodes are added onto the network objects. Only attributes that are
a part of the network model formulae are updated as vertex attributes on the
network objects.
}
\keyword{module}
\keyword{msm}
