% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{calc_nwstats_no_mix}
\alias{calc_nwstats_no_mix}
\title{Calculate Target Statistics for Network Model Estimation (No age mixing)}
\usage{
calc_nwstats_no_mix(time.unit = 7, method = 1, num.B, num.W, num.B.msm,
  num.W.msm, num.B.asmm, num.W.asmm, deg.mp.B, deg.mp.W, mdeg.inst.B,
  mdeg.inst.W, deg.asmm, cross.frac, qnts.B, qnts.W, prop.hom.mpi.B,
  prop.hom.mpi.W, balance = "mean", sqrt.adiff.BB, sqrt.adiff.WW,
  sqrt.adiff.BW, cubert.adiff.asmm, diss.main, diss.pers, diss.asmm, durs.main,
  durs.pers, rates.asmm, durs.asmm, ages, ages.asmm, ages.yamsm, ages.oamsm,
  ages.msm, birth.age, out.age.prob, debut.entry.prob, debut.prob, asmr.B,
  asmr.W, role.B.prob.msm, role.W.prob.msm, role.B.prob.asmm, role.W.prob.asmm,
  role.shift)
}
\description{
Calculates the target statistics for a no age mixng counterfactual \code{netest}.
}
\details{
Temporary function for counterfactual case.
}
\seealso{
Network statistics calculated here are entered into \code{\link{base_nw_msm}}
to construct the base network, and then into the parameters in
\code{\link{param_msm}}.
}
\keyword{msm}
