% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.cd4.R
\name{cd4_het}
\alias{cd4_het}
\title{CD4 Progression Module}
\usage{
cd4_het(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function for simulating progression of CD4 in natural
             disease dynamics and in the presence of ART.
}
\keyword{het}
\keyword{module}
