% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.cd4.R
\name{cd4_het_dcm}
\alias{cd4_het_dcm}
\title{CD4 Progression Module (DCM Version)}
\usage{
cd4_het_dcm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function for simulating progression of CD4 following broad
             compartment transitions.
}
\keyword{het}
\keyword{module}
