% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.ConcDurDx.shamp.R
\name{ConcDurDx_shamp}
\alias{ConcDurDx_shamp}
\title{Concurrency diagnostics.}
\usage{
ConcDurDx_shamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with the updated partnership list for diagnostics
on \code{dat$plist}.
}
\description{
Module function for analysing the duration of concurrencies of particular types.
}
\details{
At each time step sets of concurent relationships are identified, 
the race of the individual in the relationship is stored and the 
duration of the overlap is calculated.
}
\keyword{SHAMP}
\keyword{cocnurrency}
\keyword{diagnostics}
\keyword{module}
