% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.condoms.R
\name{condoms.mard}
\alias{condoms.mard}
\title{Condom Use Module}
\usage{
condoms.mard(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
Updates the discordant edgelist with a \code{uai} variable indicating whether
condoms were used in that act.
}
\description{
Module function stochastically simulates potential condom use
             for each act on the discordant edgelist.
}
\details{
For each act on the discordant edgelist, condom use is stochastically simulated
based on the partnership type and racial combination of the dyad. Other
modifiers for the probability of condom use in that pair are diagnosis of
disease, disclosure of status, and full or partial HIV viral suppression
given HIV anti-retroviral therapy.
}
\keyword{module}

