% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.cond.edu.R
\name{condoms_edu_campcl}
\alias{condoms_edu_campcl}
\title{Condom Intervention Module}
\usage{
condoms_edu_campcl(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
Updates attr cond.edu.timer to indicate who is exposed and the number of timesteps since exposure.
}
\description{
Module function selects those eligible for a Condom use intervention and updates the
             intervention effect timer.
}
\details{
ASMM may undergo a sex education intervention that increases condom use.  This module selects which ASMM 
are exposed to the intervention and set the duration of the intervention effect.
}
\keyword{module}
\keyword{msm}
