% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.condoms.WHAMP.R
\name{condoms_msm_whamp}
\alias{condoms_msm_whamp}
\title{Condom Use Module}
\usage{
condoms_msm_whamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
Updates the discordant edgelist with a \code{uai} variable indicating whether
condoms were used in that act.
}
\description{
Module function stochastically simulates potential condom use
             for each act on the discordant edgelist for the WHAMP model
}
\details{
For each act on the discordant edgelist, condom use is stochastically simulated
based on the partnership type and, for main partnerships, age combination. Other
modifiers for the probability of condom use in that pair are diagnosis of
disease, disclosure of status, full or partial HIV viral suppression
given HIV anti-retroviral therapy, and PrEP use. If \code{rcomp.discont == TRUE},
condom use stays at on-PrEP risk compensation levels after spontaneous 
discontinuation.
}
\keyword{module}
\keyword{msm}
