% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.condoms.shamp.R
\name{condoms_shamp}
\alias{condoms_shamp}
\title{Condom Use Module for 5 race groups and het/msm.}
\usage{
condoms_shamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
Updates the discordant edgelist with a \code{uai} variable indicating whether
condoms were used in that act for male-male dyads and a \code{uvi} variable 
indicating whether condoms were used in that act for heterosexual dyads.
}
\description{
Module function stochastically simulates potential condom use
             for each act on the discordant edgelist.
}
\details{
For each act on the discordant edgelist, condom use is stochastically simulated
based on the partnership type, racial combination of the dyad and whether the
relationship is male-male (ai) or hetersexual (vi). Other
modifiers for the probability of condom use in that pair are diagnosis of
disease, disclosure of status, and full or partial HIV viral suppression
given HIV anti-retroviral therapy.
}
\keyword{SHAMP}
\keyword{ai}
\keyword{het}
\keyword{module}
\keyword{msm}
\keyword{vi}
