% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/params.R
\name{control.mard}
\alias{control.mard}
\title{Mardham2 Control Settings}
\usage{
control.mard(simno = 1, nsims = 1, ncores = 1, nsteps = 52 * 50,
  start = 1, initialize.FUN = initialize.mard, aging.FUN = aging.mard,
  deaths.FUN = deaths.mard, births.FUN = births.mard,
  test.FUN = test.mard, tx.FUN = tx.mard, progress.FUN = progress.mard,
  vl.FUN = update_vl.mard, aiclass.FUN = update_aiclass.mard,
  roleclass.FUN = update_roleclass.mard, edgescorr.FUN = edges_correct.mard,
  resimnets.FUN = simnet.mard, disclose.FUN = disclose.mard,
  acts.FUN = acts.mard, condoms.FUN = condoms.mard,
  position.FUN = position.mard, trans.FUN = trans.mard,
  getprev.FUN = prevalence.mard, verbose.FUN = verbose.mard,
  delete.nodes = TRUE, save.dal = FALSE, save.nwstats = TRUE,
  save.network = FALSE, save.other = "attr", verbose = TRUE,
  verbose.int = 1, ...)
}
\arguments{
\item{simno}{Unique identifier for the simulation run.}

\item{nsims}{Number of simulations to run.}

\item{ncores}{Number of cores to use.}

\item{nsteps}{Number of time steps in each simulation.}

\item{start}{Time step at which to start simulation.}

\item{initialize.FUN}{Function to use for initialization.}

\item{aging.FUN}{Aging function.}

\item{deaths.FUN}{Deaths function.}

\item{births.FUN}{Births function.}

\item{test.FUN}{Disease testing function.}

\item{tx.FUN}{Treatment function.}

\item{progress.FUN}{Disease progress function.}

\item{vl.FUN}{Viral load function.}

\item{aiclass.FUN}{AI class function.}

\item{roleclass.FUN}{Role class function.}

\item{edgescorr.FUN}{Edges coefficient correction function.}

\item{resimnets.FUN}{Network resimulation function.}

\item{disclose.FUN}{Disclosure function.}

\item{acts.FUN}{Acts function.}

\item{condoms.FUN}{Condom use function.}

\item{position.FUN}{Position function.}

\item{trans.FUN}{Transmission function.}

\item{getprev.FUN}{Prevalence summary function.}

\item{verbose.FUN}{Verbose simulation progress function.}

\item{delete.nodes}{Logical. Should inactive nodes be deleted?}

\item{save.dal}{Logical. Should the discordant act list be saved at each time
step?}

\item{save.nwstats}{Logical. Should the network statistics be saved at each
time step?}

\item{save.network}{Logical. Should the entire networks be saved at each time
step?}

\item{save.other}{Other list elements of \code{dat} to save.}

\item{verbose}{Logical. Should simulation progress be printed from within the
time loop?}

\item{verbose.int}{Integer giving the interval between time steps at which
progress is printed.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list object of class \code{control.mard}, which can be passed to
        EpiModel function \code{netsim}.
}
\description{
Set control options for a stochastic epidemic network model.
}
\details{
Arguments that end in \code{.FUN} default to the corresponding
         \code{.mard} function of the same name.
}

