% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{control.msm}
\alias{control.msm}
\title{Epidemic Model Control Settings}
\usage{
control.msm(simno = 1, nsims = 1, ncores = 1, nsteps = 100, start = 1,
  initialize.FUN = initialize.msm, aging.FUN = aging.msm,
  deaths.FUN = deaths.msm, births.FUN = births.msm, test.FUN = test.msm,
  tx.FUN = tx.msm, prep.FUN = prep.msm, progress.FUN = progress.msm,
  vl.FUN = update_vl.msm, aiclass.FUN = NULL, roleclass.FUN = NULL,
  edgescorr.FUN = edges_correct.msm, resim_nets.FUN = simnet.msm,
  disclose.FUN = disclose.msm, acts.FUN = acts.msm,
  condoms.FUN = condoms.msm, riskhist.FUN = riskhist.msm,
  position.FUN = position.msm, trans.FUN = trans.msm,
  getprev.FUN = prevalence.msm, verbose.FUN = verbose.msm,
  prune.discl.list = TRUE, save.other = NULL, verbose = TRUE,
  verbose.int = 1, ...)
}
\arguments{
\item{simno}{Unique ID for the simulation run, used for file naming purposes
if used in conjunction with the \code{EpiModelHPC} package.}

\item{nsims}{Number of simulations.}

\item{ncores}{Number of cores per run, if parallelization is used within the
\code{EpiModelHPC} package.}

\item{nsteps}{Number of time steps per simulation.}

\item{start}{Starting time step for simulation, with default to 1 to run new
simulation. This may also be set to 1 greater than the final time
step of a previous simulation to resume the simulation with different
parameters.}

\item{initialize.FUN}{Module function to use for initialization of the epidemic
model.}

\item{aging.FUN}{Module function for aging.}

\item{deaths.FUN}{Module function for general and disease-realted deaths.}

\item{births.FUN}{Module function for births or entries into the population.}

\item{test.FUN}{Module function for diagnostic disease testing.}

\item{tx.FUN}{Module function for ART initiation and adherence.}

\item{prep.FUN}{Module function for PrEP initiation and utilization.}

\item{progress.FUN}{Module function for HIV disease progression.}

\item{vl.FUN}{Module function for HIV viral load evolution.}

\item{aiclass.FUN}{Module function for one-off AI risk class transitions.}

\item{roleclass.FUN}{Module function for transitions in sexual roles.}

\item{edgescorr.FUN}{Module function for the edges coefficient adjustment
to preserve mean degree under varying population sizes.}

\item{resim_nets.FUN}{Module function for network resimulation at each time
step.}

\item{disclose.FUN}{Module function for HIV status disclosure.}

\item{acts.FUN}{Module function to simulate the number of sexual acts within
partnerships.}

\item{condoms.FUN}{Module function to simulate condom use within acts.}

\item{riskhist.FUN}{Module function to calculate risk history for uninfected
persons in the population.}

\item{position.FUN}{Module function to simulate sexual position within acts.}

\item{trans.FUN}{Module function to stochastically simulate disease transmission
over acts given individual and dyadic attributes.}

\item{getprev.FUN}{Module function to calculate prevalence summary statistics.}

\item{verbose.FUN}{Module function to print model progress to the console or
external text files.}

\item{prune.discl.list}{If \code{TRUE}, remove any dissolved edges from the
disclosure list.
out at the end of simulation (necessary for restarting a simulation).}

\item{save.other}{Character vector containing other list elements of \code{dat}
to save.}

\item{verbose}{If \code{TRUE}, print out simulation progress to the console
if in interactive mode or text files if in batch mode.}

\item{verbose.int}{Integer specifying the interval between time steps at which
progress is printed.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list object of class \code{control.msm}, which can be passed to the
EpiModel function \code{netsim}.
}
\description{
Sets the controls for stochastic network models simulated with
             \code{\link{netsim}}.
}

