% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{control_het}
\alias{control_het}
\title{Control Settings for Stochastic Network Model of HIV-1 Infection in
       Sub-Saharan Africa}
\usage{
control_het(simno = 1, nsteps = 100, start = 1, nsims = 1, ncores = 1,
  par.type = "single", initialize.FUN = initialize_het,
  aging.FUN = aging_het, cd4.FUN = cd4_het, vl.FUN = vl_het,
  dx.FUN = dx_het, tx.FUN = tx_het, deaths.FUN = deaths_het,
  births.FUN = births_het, resim_nets.FUN = simnet_het,
  trans.FUN = trans_het, prev.FUN = prevalence_het,
  verbose.FUN = verbose_het, module.order = NULL, save.nwstats = FALSE,
  save.other = c("el", "attr"), verbose = TRUE, verbose.int = 1,
  skip.check = TRUE, ...)
}
\arguments{
\item{simno}{Simulation ID number.}

\item{nsteps}{Number of time steps to simulate the model over in whatever unit
implied by \code{time.unit}.}

\item{start}{Starting time step for simulation}

\item{nsims}{Number of simulations.}

\item{ncores}{Number of parallel cores to use for simulation jobs, if using
the \code{EpiModel.hpc} package.}

\item{par.type}{Parallelization type, either of \code{"single"} for multi-core
or \code{"mpi"} for multi-node MPI threads.}

\item{initialize.FUN}{Module to initialize the model at time 1.}

\item{aging.FUN}{Module to age active nodes.}

\item{cd4.FUN}{CD4 progression module.}

\item{vl.FUN}{HIV viral load progression module.}

\item{dx.FUN}{HIV diagnosis module.}

\item{tx.FUN}{HIV treatment module.}

\item{deaths.FUN}{Module to simulate death or exit.}

\item{births.FUN}{Module to simulate births or entries.}

\item{resim_nets.FUN}{Module to resimulate the network at each time step.}

\item{trans.FUN}{Module to simulate disease infection.}

\item{prev.FUN}{Module to calculate disease prevalence at each time step,
with the default function of \code{\link{prevalence_het}}.}

\item{verbose.FUN}{Module to print simulation progress to screen, with the
default function of \code{\link{verbose_het}}.}

\item{module.order}{A character vector of module names that lists modules the
order in which they should be evaluated within each time step. If
\code{NULL}, the modules will be evaluated as follows: first any
new modules supplied through \code{...} in the order in which they are
listed, then the built-in modules in their order of the function listing.
The \code{initialize.FUN} will always be run first and the
\code{verbose.FUN} always last.}

\item{save.nwstats}{Save out network statistics.}

\item{save.other}{Other list elements of dat to save out.}

\item{verbose}{If \code{TRUE}, print progress to console.}

\item{verbose.int}{Interval for printing progress to console.}

\item{skip.check}{If \code{TRUE}, skips the error check for parameter values,
initial conditions, and control settings before running the models.}

\item{...}{Additional arguments passed to the function.}
}
\description{
This function sets the control settings for the stochastic
             network models in the \code{epimethods} package.
}
\details{
This function sets the parameters for the models.
}
\keyword{het}
