% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{control_msm}
\alias{control_msm}
\title{Epidemic Model Control Settings}
\usage{
control_msm(simno = 1, nsims = 1, ncores = 1, nsteps = 100,
  start = 1, initialize.FUN = initialize_msm, aging.FUN = aging_msm,
  departure.FUN = departure_msm, arrival.FUN = arrival_msm,
  hivtest.FUN = hivtest_msm, hivtx.FUN = hivtx_msm,
  hivprogress.FUN = hivprogress_msm, hivvl.FUN = hivvl_msm,
  resim_nets.FUN = simnet_msm, acts.FUN = acts_msm,
  condoms.FUN = condoms_msm, position.FUN = position_msm,
  prep.FUN = prep_msm, hivtrans.FUN = hivtrans_msm,
  stitrans.FUN = stitrans_msm, stirecov.FUN = stirecov_msm,
  stitx.FUN = stitx_msm, prev.FUN = prevalence_msm,
  verbose.FUN = verbose.net, save.nwstats = FALSE,
  save.clin.hist = FALSE, truncate.plist = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{simno}{Unique ID for the simulation run, used for file naming purposes
if used in conjunction with the \code{EpiModelHPC} package.}

\item{nsims}{Number of simulations.}

\item{ncores}{Number of cores per run, if parallelization is used within the
\code{EpiModelHPC} package.}

\item{nsteps}{Number of time steps per simulation.}

\item{start}{Starting time step for simulation, with default to 1 to run new
simulation. This may also be set to 1 greater than the final time
step of a previous simulation to resume the simulation with different
parameters.}

\item{initialize.FUN}{Module function to use for initialization of the epidemic
model.}

\item{aging.FUN}{Module function for aging.}

\item{departure.FUN}{Module function for general and disease-realted depatures.}

\item{arrival.FUN}{Module function for entries into the sexually active population.}

\item{hivtest.FUN}{Module function for HIV diagnostic disease testing.}

\item{hivtx.FUN}{Module function for ART initiation and adherence.}

\item{hivprogress.FUN}{Module function for HIV disease progression.}

\item{hivvl.FUN}{Module function for HIV viral load evolution.}

\item{resim_nets.FUN}{Module function for network resimulation at each time
step.}

\item{acts.FUN}{Module function to simulate the number of sexual acts within
partnerships.}

\item{condoms.FUN}{Module function to simulate condom use within acts.}

\item{position.FUN}{Module function to simulate sexual position within acts.}

\item{prep.FUN}{Module function for PrEP initiation and utilization.}

\item{hivtrans.FUN}{Module function to stochastically simulate HIV transmission
over acts given individual and dyadic attributes.}

\item{stitrans.FUN}{Module function to simulate GC/CT transmission over current
edgelist.}

\item{stirecov.FUN}{Module function to simulate recovery from GC/CT, heterogeneous
by disease, site, symptoms, and treatment status.}

\item{stitx.FUN}{Module function to simulate treatment of GC/CT.}

\item{prev.FUN}{Module function to calculate prevalence summary statistics.}

\item{verbose.FUN}{Module function to print model progress to the console or
external text files.}

\item{save.nwstats}{Calculate and save network statistics as defined in the
\code{simnet} modules.}

\item{save.clin.hist}{Save individual-level clinical history matrices.}

\item{truncate.plist}{Truncate the cumulative partnership list to only include
active partnerships.}

\item{verbose}{If \code{TRUE}, print out simulation progress to the console
if in interactive mode or text files if in batch mode.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list object of class \code{control_msm}, which can be passed to the
EpiModel function \code{netsim}.
}
\description{
Sets the controls for stochastic network models simulated with
             \code{\link{netsim}}.
}
\keyword{msm}
