% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.death.R
\name{deaths_het}
\alias{deaths_het}
\title{Deaths Module}
\usage{
deaths_het(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module for simulating deaths among susceptible and infected
             persons within the population.
}
\keyword{het}
\keyword{module}
