% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.death.whamp.R
\name{deaths_msm_whamp}
\alias{deaths_msm_whamp}
\title{Death Module}
\usage{
deaths_msm_whamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the updated \code{dat} object accounting for deaths.
The deaths are deactivated from the main and persistent networks, as those are in
\code{networkDynamic} class objects; dead nodes are not deleted from the
instant network until the \code{\link{simnet_msm}} module for bookkeeping
purposes.
}
\description{
Module function for simulting both general and disease-related
             deaths, as well as aging out of the network, among population 
             members for the WHAMP model.
}
\details{
Deaths are divided into two categories: general deaths, for which demographic
data on age-specific mortality rates applies; and disease-related deaths,
which are calculated by multiplying general age-specific mortality rates by a 
scalar for persons living with HIV.
Which nodes have died is determined stochastically using draws from binomial 
distributions for both general and HIV-related deaths. This module also tracks 
nodes that age out of the network.
}
\keyword{module}
\keyword{msm}
