% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.death.shamp.R
\name{deaths_shamp}
\alias{deaths_shamp}
\title{Death Module for up to 5 race groups heterosexuals and MSM.}
\usage{
deaths_shamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the updated \code{dat} object accounting for deaths.
The deaths are deactivated from the main and casual networks, as those are in
\code{networkDynamic} class objects; dead nodes are not deleted from the
instant network until the \code{\link{simnet_shamp}} module for bookkeeping
purposes.
}
\description{
Module function for simulting both general and disease-related
             deaths among population members.
}
\details{
Deaths are divided into two categories: general deaths, for which demographic
data on age-specific mortality rates applies; and disease-related diseases,
for which the rate of death is a function of progression to end-stage AIDS.
Which nodes have died is determined stochastically for general deaths using
draws from a binomial distribution, and deterministically for disease-related
deaths after nodes have reach a maximum viral load value set in the
\code{vl.fatal} parameter.
}
\keyword{het}
\keyword{module}
\keyword{msm}
\keyword{shamp}
