% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.demog.shamp.R
\name{demogupdate_shamp}
\alias{demogupdate_shamp}
\title{Demography Check Module}
\usage{
demogupdate_shamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module updated demographic clasifications that change over time.
}
\keyword{het}
\keyword{module}
