% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.demog.shamp1875.R
\name{demogupdate_shamp_1875}
\alias{demogupdate_shamp_1875}
\title{Demography Check Module (age 18-75)}
\usage{
demogupdate_shamp_1875(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module updated demographic clasifications that change over time.
}
\keyword{het}
\keyword{module}
