% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mod.acts.R
\name{disc_check}
\alias{disc_check}
\title{Check for Discordant Pairs}
\usage{
disc_check(el, status)
}
\arguments{
\item{el}{Current edgelist.}

\item{status}{Disease status vector.}
}
\value{
This function returns an edgelist of only the discordant relations, with the
seropositive man in the first column.
}
\description{
Checks that the discordant edgelist is in a proper form before
             returning its values in a matrix.
}
\keyword{submodule}

