% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.disclose.WHAMP.R
\name{disclose_msm_whamp}
\alias{disclose_msm_whamp}
\title{Disclosure Module}
\usage{
disclose_msm_whamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with the updated disclosure list,
on \code{temp$discl.list}.
}
\description{
Module function for disclosure of HIV status to partners given
             non-disclosure in the past for the WHAMP model.
}
\details{
Persons who are infected may disclose their status to partners at three
distinct time points: at relationship onset for newly formed discordant
pairs; at diagnosis for pairs starting as both negative but with one newly
infected; or post diagnosis for ones recently infected. The rates of disclosure
vary at these three points, and also by the partnership type. In this model, we
do not stratify disclosure by race/ethnicity.
}
\keyword{module}
\keyword{msm}
