% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.disclose.shamp.R
\name{disclose_shamp}
\alias{disclose_shamp}
\title{Disclosure Module for up to 5 race groups and heterosexuals and MSM.}
\usage{
disclose_shamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with the updated disclosure list,
on \code{temp$discl.list}.
}
\description{
Module function for disclosure of HIV status to partners given
             non-disclosure in the past.
}
\details{
Persons who are infected may disclose their status to partners at three
distinct time points: at relationship onset for newly formed discordant
pairs; at diagnosis for pairs starting as both negative but with one newly
infected; or post diagnosis for one recently infected. The rates of disclosure
vary at these three points, and also by the partnership type race and sexual identity.
}
\keyword{SHAMP}
\keyword{het}
\keyword{module}
\keyword{msm}
