% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.simnet.shamp.R
\name{edges_correct_msm}
\alias{edges_correct_msm}
\title{Adjustment for the Edges Coefficient with Changing Network Size}
\usage{
edges_correct_msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
The network model parameters stored in \code{dat$nwparam} are updated for
each of the three network models.
}
\description{
Adjusts the edges coefficients in a dynamic network model
             to preserve the mean degree.
}
\details{
In HIV/STI modeling, there is typically an assumption that changes in
population size do not affect one's number of partners, specified as the
mean degree for network models. A person would not have 10 times the number
of partners should he move from a city 10 times as large. This module uses
the adjustment of Krivitsky et al. to adjust the edges coefficients on the
three network models to account for varying population size in order to
preserve that mean degree.
}
\references{
Krivitsky PN, Handcock MS, and Morris M. "Adjusting for network size and
composition effects in exponential-family random graph models." Statistical
Methodology. 2011; 8.4: 319-339.
}
\keyword{module}
\keyword{msm}
\keyword{shamp}
