% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ego.netest.KTM}
\alias{ego.netest.KTM}
\title{Ego to est for KenyaTM}
\usage{
ego.netest.KTM(eefit, disscoef, obs.adjust = FALSE, obs.offset = 52)
}
\arguments{
\item{path}{Directory of files to source.}

\item{verbose}{Print names of sourced files to console.}

\item{...}{Additional arguments passed to \code{source}.}
}
\description{
temporary version of ego.netest for imposing changes in the Kenya formation models
}
