% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelHIV-package.R
\docType{package}
\name{EpiModelHIV-package}
\alias{EpiModelHIV}
\alias{EpiModelHIV-package}
\title{HIV Transmission Dynamics among MSM}
\description{
\tabular{ll}{
   Package: \tab EpiModelHIV\cr
   Type: \tab Package\cr
   Version: \tab 1.0.0\cr
   Date: \tab 2016-06-25\cr
   License: \tab GPL (>= 2)\cr
   LazyLoad: \tab yes\cr
}
}
\details{
EpiModel version of stochastic network modeling project for
modeling HIV-1 transmission dynamics among men who have sex with men (MSM),
and heterosexual populations.
}
\keyword{package}

