% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelHIVmsm-package.R
\docType{package}
\name{EpiModelHIVmsm-package}
\alias{EpiModelHIVmsm}
\alias{EpiModelHIVmsm-package}
\title{HIV Transmission Dynamics among MSM}
\description{
\tabular{ll}{
   Package: \tab EpiModelHIVmsm\cr
   Type: \tab Package\cr
   Version: \tab 0.1.0\cr
   Date: \tab 2015-03-20\cr
   License: \tab GPL (>= 2)\cr
   LazyLoad: \tab yes\cr
}
}
\details{
EpiModel version of MARDHAM stochastic network modeling project for
modeling HIV-1 transmission dynamics among young men who have sex with men (MSM)
in Atlanta. This translates and updates the original MARDHAM software
into EpiModel modules for greater flexibility and efficiency, and then
adds additional features like pre-exposure prophylaxis (PrEP).
}
\keyword{package}

