% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.ept.R
\name{ept_msm}
\alias{ept_msm}
\title{EPT Module}
\usage{
ept_msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function for implementation and uptake of expedited
             partner therapy (EPT) to prevent STI infection.
}
\keyword{module}
\keyword{msm}
