% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ergm_prep}
\alias{ergm_prep}
\title{PrEP ergm}
\usage{
ergm_prep(
  nw,
  formation,
  coef,
  constraints,
  control = ergm::control.simulate.ergm()
)
}
\arguments{
\item{path}{Directory of files to source.}

\item{verbose}{Print names of sourced files to console.}

\item{...}{Additional arguments passed to \code{source}.}
}
\description{
set up the structure for the ergm function. No longer exported in tegmLite
}
