% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{est}
\alias{est}
\title{Fitted Network Models for 10000 Person Network}
\format{A list with three elements.}
\description{
A dataset containing a list of three outputs from \code{netest}, one for each
network. This would be used primarily for more efficient package testing.
}
\keyword{datasets}
