% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.FU.elements.R
\name{FU_elements_camplc}
\alias{FU_elements_camplc}
\title{Follow-up elements Module}
\usage{
FU_elements_camplc(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns \code{dat} after adding elements to dat that are not tracked during the burnin
\code{riskhist} , \code{age.inf.vec}, \code{counts}.
}
\description{
Module sets up adolcent risk history and infection age elements of dat.
}
\keyword{module}
