% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R, R/utilities.WHAMP.R
\name{get_args}
\alias{get_args}
\title{Get Arguments from EpiModel Parameterization Functions}
\usage{
get_args(formal.args, dot.args)

get_args(formal.args, dot.args)
}
\arguments{
\item{formal.args}{The output of \code{formals(sys.function())}.}

\item{dot.args}{The output of \code{list(...)}.}

\item{formal.args}{The output of \code{formals(sys.function())}.}

\item{dot.args}{The output of \code{list(...)}.}
}
\description{
Returns a list of argument names and values for use for parameter
             processing functions.

Returns a list of argument names and values for use for parameter
             processing functions.
}
