% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_sims2}
\alias{get_sims2}
\title{Select sim with minimum divergence}
\usage{
get_sims2(x, tar, var)
}
\arguments{
\item{...}{Additional arguments passed to \code{source}.}
}
\description{
Loops through a seletion of simulations and selects a simulation with the minimum divergence 
             from a vector of target outcomes among those that are at least 95% the value of the targets.
}
