% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.msmf.heatbath.R
\name{heatbath_msmf_shamp}
\alias{heatbath_msmf_shamp}
\title{Transmission Module for Heat Bath Exposure of MSMF to an MSM heatbath}
\usage{
heatbath_msmf_shamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
For each new infection, the disease status, infection time, and related
HIV attributes are updated for the infected node. Summary statistics for
disease incidence overall, and by race are calculated and
stored on \code{dat$epi}.
}
\description{
Stochastically simulates disease transmission from the heat bath given the current
             state of the active HIV negative population.
}
\details{
This function takes all active HIV negative MSMF nodes and calculates a
transmission probability for each individual based race specific exposure to the heatbath
which incorperates race specific prevalence estiamtes for determining risk.
After transmission events, individual-level attributes for the infected
persons are updated and summary statistics for incidence calculated.

The input values \code{dat$param$msm.aq.prob.'R'} 
where R is a race specific probability taking values (B,BI,H,HI,W).
}
\keyword{SHAMP}
\keyword{heatbath}
\keyword{module}
\keyword{msm}
\keyword{msmf}
