% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.hivtest.R
\name{hivtest_msm}
\alias{hivtest_msm}
\alias{dx_het}
\title{HIV Testing Module}
\usage{
hivtest_msm(dat, at)

dx_het(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with updated \code{last.neg.test},
\code{diag.status} and \code{diag.time} attributes.
}
\description{
Module function for HIV diagnostic testing of infected persons.
}
\details{
This testing module supports memoryless HIV testing for stochastic and
geometrically-distributed waiting times to test (constant hazard).
}
\keyword{module}
\keyword{msm}
