% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.hivtx.R
\name{hivtx_msm}
\alias{hivtx_msm}
\alias{tx_het}
\title{Treatment Module}
\usage{
hivtx_msm(dat, at)

tx_het(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with updated \code{tx.status},
\code{tx.init.time}, \code{cuml.time.on.tx}, \code{cuml.time.off.tx} attributes.
}
\description{
Module function for anti-retroviral treatment initiation and
             adherence over time.
}
\details{
Persons enter into the simulation with one of four ART "patterns": never
tested, tested but never treated, treated and achieving partial HIV viral
suppression, or treated with full viral suppression (these types are stored
as individual-level attributes in \code{tt.traj}). This module initiates ART
for treatment naive persons in the latter two types, and then cycles them on
and off treatment conditional on empirical race-specific adherence rates. ART
initiation, non-adherence, and restarting are all stochastically simulated
based on binomial statistical models.
}
\keyword{module}
\keyword{msm}
