% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.trans.R
\name{infect_het}
\alias{infect_het}
\title{Infection Module}
\usage{
infect_het(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function to simulate transmission over an active discordant
             edgelist.
}

