% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/params.R
\name{init.mard}
\alias{init.mard}
\title{Mardham2 Initial Conditions}
\usage{
init.mard(nwstats, prev.B = 0.275, prev.W = 0.275,
  init.prev.age.slope.B = 0.05/12, init.prev.age.slope.W = 0.05/12, ...)
}
\arguments{
\item{nwstats}{An list of target statistics created by
\code{\link{calc_nwstats.mard}}.}

\item{prev.B}{Disease prevalence among Black MSM.}

\item{prev.W}{Disease prevalence among White MSM.}

\item{init.prev.age.slope.B}{Slope of initial prevalence by age for Black MSM.}

\item{init.prev.age.slope.W}{Slope of initial prevalence by age for White MSM.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list object of class \code{init.mard}, which can be passed to
        EpiModel function \code{netsim}.
}
\description{
Set the initial conditions for a stochastic epidemic
             network simulation.
}
\details{
\code{nwstats} must be supplied by the user. Prevalence defaults
         to 27.5 percent for both black and white men.
}

