% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{init.msm}
\alias{init.msm}
\title{Epidemic Model Initial Conditions}
\usage{
init.msm(nwstats, prev.B = 0.15, prev.W = 0.15, ...)
}
\arguments{
\item{nwstats}{Target statistics for the network model. An object of class
\code{nwstats} output from \code{\link{calc_nwstats.msm}}.}

\item{prev.B}{Initial disease prevalence among black MSM.}

\item{prev.W}{Initial disease prevalence among white MSM.}

\item{...}{Additional arguments passed to function.}
}
\value{
A list object of class \code{init.msm}, which can be passed to EpiModel
function \code{\link{netsim}}.
}
\description{
Sets the initial conditions for a stochastic epidemic models
             simulated with \code{\link{netsim}}.
}

