% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mod.init.R
\name{init_ccr5}
\alias{init_ccr5}
\title{Initialization of CCR5 Gene}
\usage{
init_ccr5(dat)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}
}
\description{
Initializes CCR5 gene that confers disease protection.
}
\keyword{submodule}

