% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.initialize.R
\name{init_ccr5_msm}
\alias{init_ccr5_msm}
\title{Sets the CCR5 genetic status of persons}
\usage{
init_ccr5_msm(dat)
}
\arguments{
\item{dat}{Data object created in initialization module.}
}
\description{
Initializes the CCR5-delta-32 genetic allele of the men in the
             population, based on parameters defining the probability
             distribution.
}
\keyword{initiation}
\keyword{msm}
\keyword{utility}
