% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{init_het}
\alias{init_het}
\title{Initial Conditions for Stochastic Network Model of HIV-1 Infection in
       Sub-Saharan Africa}
\usage{
init_het(i.prev.male = 0.05, i.prev.feml = 0.05, ages.male = seq(18,
  55, 7/365), ages.feml = seq(18, 55, 7/365),
  inf.time.dist = "geometric", max.inf.time = 5 * 365, ...)
}
\arguments{
\item{i.prev.male}{Prevalence of initially infected males.}

\item{i.prev.feml}{Prevalence of initially infected females.}

\item{ages.male}{initial ages of males in the population.}

\item{ages.feml}{initial ages of females in the population.}

\item{inf.time.dist}{Probability distribution for setting time of infection
for nodes infected at T1, with options of \code{"geometric"} for
randomly distributed on a geometric distribution with a probability
of the reciprocal of the average length of infection, \code{"uniform"}
for a uniformly distributed time over that same interval, or
\code{"allacute"} for placing all infections in the acute stage at
the start.}

\item{max.inf.time}{Maximum infection time in days for infection at
initialization, used when \code{inf.time.dist} is \code{"geometric"}
or \code{"uniform"}.}

\item{...}{additional arguments to be passed into model.}
}
\description{
This function sets the initial conditions for the stochastic
             network models in the \code{epimethods} package.
}
\details{
This function sets the initial conditions for the models.
}
\keyword{het}
