% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.KTM.R
\name{init_KTM}
\alias{init_KTM}
\title{Epidemic Model Initial Conditions for SHAMP}
\usage{
init_KTM(prev.B.f = 0.02, prev.B.msf = 0.02, ...)
}
\arguments{
\item{...}{Additional arguments passed to function.}

\item{prev.R.S}{Initial disease prevalence among persons of group R,
where R is a charater (B, BI, H, HI or W), and sex group S
      where s is a charter (f, msf, msm, msmf)..}
}
\value{
A list object of class \code{init_shamp}, which can be passed to EpiModel
function \code{\link{netsim}}.
}
\description{
Sets the initial conditions for a stochastic epidemic models
             simulated with \code{\link{netsim}}.
}
\keyword{SHAMP}
