% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.WHAMP.R
\name{init_msm_whamp}
\alias{init_msm_whamp}
\title{Epidemic Model Initial Conditions for the WHAMP model}
\usage{
init_msm_whamp(nwstats, prev.H..wa = 0.08, prev.B..wa = 0.12,
  prev.O..wa = 0.05, prev.ugc = 0, prev.rgc = 0, prev.uct = 0,
  prev.rct = 0, ...)
}
\arguments{
\item{nwstats}{Target statistics for the network model. An object of class
\code{nwstats} output from \code{\link{calc_nwstats_msm}}.}

\item{prev.H..wa}{Initial disease prevalence among Hispanic MSM.}

\item{prev.B..wa}{Initial disease prevalence among non-Hispanic black MSM.}

\item{prev.O..wa}{Initial disease prevalence among non-Hispanic other MSM.}

\item{prev.ugc}{Initial prevalence of urethral gonorrhea.}

\item{prev.rgc}{Initial prevalence of rectal gonorrhea.}

\item{prev.uct}{Initial prevalence of urethral chlamydia.}

\item{prev.rct}{Initial prevalence of rectal chlamydia.}

\item{...}{Additional arguments passed to function.}
}
\value{
A list object of class \code{init_msm}, which can be passed to EpiModel
function \code{\link{netsim}}.
}
\description{
Sets the initial conditions for a stochastic epidemic models
             simulated with \code{\link{netsim}}.
}
\keyword{msm}
