% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.shamp.R
\name{init_shamp}
\alias{init_shamp}
\title{Epidemic Model Initial Conditions for SHAMP}
\usage{
init_shamp(prev.B.f = 0.02, prev.BI.f = 0.02, prev.H.f = 0.02,
  prev.HI.f = 0.02, prev.W.f = 0.02, prev.B.msf = 0.02,
  prev.BI.msf = 0.02, prev.H.msf = 0.02, prev.HI.msf = 0.02,
  prev.W.msf = 0.02, prev.B.msm = 0.02, prev.BI.msm = 0.02,
  prev.H.msm = 0.02, prev.HI.msm = 0.02, prev.W.msm = 0.02,
  prev.B.msmf = 0.02, prev.BI.msmf = 0.02, prev.H.msmf = 0.02,
  prev.HI.msmf = 0.02, prev.W.msmf = 0.02, ...)
}
\arguments{
\item{...}{Additional arguments passed to function.}

\item{prev.R.S}{Initial disease prevalence among persons of group R,
where R is a charater (B, BI, H, HI or W), and sex group S
      where s is a charter (f, msf, msm, msmf)..}
}
\value{
A list object of class \code{init_shamp}, which can be passed to EpiModel
function \code{\link{netsim}}.
}
\description{
Sets the initial conditions for a stochastic epidemic models
             simulated with \code{\link{netsim}}.
}
\keyword{SHAMP}
