% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mod.init.R
\name{init_status.mard}
\alias{init_status.mard}
\title{Initialization of Disease-Related Attributes}
\usage{
init_status.mard(dat)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}
}
\description{
Initializes disease-related attributes such as disease status,
             time of infection, and diagnosis and treatment histories.
}
\keyword{submodule}

