% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.initialize.KTM.R
\name{init_status_shamp}
\alias{init_status_shamp}
\title{Initialize the HIV status of persons in the network}
\usage{
init_status_shamp(dat)
}
\arguments{
\item{dat}{Data object created in initialization module.}
}
\description{
Sets the initial individual-level disease status of persons
             in the network, as well as disease-related attributes for
             infected persons.
}
\keyword{initiation}
\keyword{shamp}
\keyword{utility}
