% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.initialize.shamp.R
\name{initialize_shamp}
\alias{initialize_shamp}
\title{Initialization Module for up to 5 race groups heterosexuals and MSM.}
\usage{
initialize_shamp(x, param, init, control, s)
}
\arguments{
\item{x}{An \code{EpiModel} object of class \code{\link{netest}}.}

\item{param}{An \code{EpiModel} object of class \code{\link{param_shamp}}.}

\item{init}{An \code{EpiModel} object of class \code{\link{init_shamp}}.}

\item{control}{An \code{EpiModel} object of class \code{\link{control_shamp}}.}

\item{s}{Simulation number, used for restarting dependent simulations.}
}
\value{
This function returns the updated \code{dat} object with the initialized values
for demographics and disease-related variables.
}
\description{
This function initializes the master \code{dat} object on which
             data are stored, initial state of the network is simulated using ergm.ego, and
             simulates disease status and other attributes are initialized.
}
\keyword{HET}
\keyword{MSM}
\keyword{ego}
\keyword{module}
