% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.data.shamp.R
\name{input_shamp}
\alias{input_shamp}
\title{Import and Check Data}
\usage{
input_shamp(data, data.params, immigration, msm.msmf)
}
\arguments{
\item{...}{Additional arguments passed to the function.}
}
\value{
A list object of class \code{data.params}, which can be passed to
EpiModel function \code{params}
EpiModel function \code{setup}
.
}
\description{
Imports the data for simulations using ergm.ego and SHAMP function.  Ego and alter data are checked to verify that all 
             of the required fields are present.  Egos and alters with missing data are dropped.
}
\keyword{shamp}
