% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ltGhana}
\alias{ltGhana}
\title{Life Table for Ghana}
\format{A data frame with 132 rows and 5 variables.}
\description{
A dataset containing age-sex-period specific mortality rates.
}
\details{
\itemize{
  \item \strong{year}: calendar year for rates, with options of 1990, 2000,
        and 2011.
  \item \strong{male}: binary indicator for male (vs female) rates.
  \item \strong{agStart}: lower age interval at which the rate applies.
  \item \strong{agEnd}: upper age interval at which the rate applies.
  \item \strong{mrate}: the mortality rate.
}
}
\keyword{datasets}
