% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{make_nw_het}
\alias{make_nw_het}
\title{Calculate Network Statistics}
\usage{
make_nw_het(n = 10000, meandeg = 0.8, prop.male = 0.5,
  start.prev = 0.05, part.dur = 1000, time.unit = 7)
}
\arguments{
\item{n}{Population size.}

\item{meandeg}{Mean degree.}

\item{prop.male}{Percent of the population that is male.}

\item{start.prev}{Starting HIV prevalence in the population.}

\item{part.dur}{Mean duration of partnerships.}

\item{time.unit}{Time unit used, relative to days.}
}
\description{
This function calculates the target statistics for the formation
             and dissolution models estimated in \code{netest}.
}
\keyword{het}
