% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.small.data.shamp.R
\name{make_small}
\alias{make_small}
\title{Cut population size for testing}
\usage{
make_small(data)
}
\arguments{
\item{...}{Additional arguments passed to the function.}
}
\value{
data.
}
\description{
Imports the data for simulations using ergm.ego and SHAMP function.  Ego and alter data are cut to a limited age range to 
reduce object size.
}
\keyword{shamp}
