% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdx.alt.R
\name{netdx.alt}
\alias{netdx.alt}
\title{netdx - shamp alternative}
\usage{
netdx.alt(x, nsims = 1, dynamic = TRUE, nsteps,
  nwstats.formula = "formation", set.control.ergm, set.control.stergm,
  keep.tedgelist = FALSE, verbose = TRUE, ncores = 1)
}
\arguments{
\item{x}{is a fit object.}
}
\value{
This function returns \code{netdx}.
}
\description{
Runs dynamic diagnostics on an ERGM/STERGM estimated through ergm.ego and altered for EpiModelHIV.
}
\keyword{diagnostics}
\keyword{module}
\keyword{netwrok}
