% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.out.debut.R
\name{out_debut_camplc}
\alias{out_debut_camplc}
\title{Out and Debut Module}
\usage{
out_debut_camplc(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns \code{dat} after updating the nodal attribute
\code{asmm} , \code{yamsm} ,\code{oamsm} ,\code{out} ,\code{debuted}.
}
\description{
Module coming out (entering the population) and debut as sexually available.
}
\keyword{module}
