% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/params.R
\name{param.mard}
\alias{param.mard}
\title{Mardham2 Parameters}
\usage{
param.mard(nwstats, last.neg.test.time.range.B = round(301/7),
  mean.test.interv.B = round(301/7),
  last.neg.test.time.range.W = round(315/7),
  mean.test.interv.W = round(315/7), testing.pattern = "interval",
  test.window.period = 3, tt.traj.freq.B = c(0.077, 0, 0.356, 0.567),
  tt.traj.freq.W = c(0.052, 0, 0.331, 0.617), prob.tx.init.B = 0.092,
  prob.tx.init.W = 0.127, prob.tx.halt.B = 0.0102,
  prob.tx.halt.W = 0.0071, prob.tx.reinit.B = 0.00066,
  prob.tx.reinit.W = 0.00291, max.time.off.tx.full = 520,
  max.time.on.tx.part = 52 * 15, max.time.off.tx.part = 520,
  vl.acute.rise.dur = 3, vl.acute.peak = 6.886, vl.acute.fall.dur = 3,
  vl.set.point = 4.5, vl.aids.onset = 520, vl.aids.dur = 52 * 2,
  vl.fatal = 7, vl.full.supp = 1.5, vl.part.supp = 3.5,
  full.supp.down.slope = 0.25, full.supp.up.slope = 0.25,
  part.supp.down.slope = 0.25, part.supp.up.slope = 0.25,
  b.rate.B = 0.001, b.rate.W = 0.001, birth.age = 18,
  betabase.URAI = 0.0082 * 1.09, betabase.UIAI = 0.0031 * 1.09,
  betamult.acute = 4, betamult.circ = 0.4, betamult.condom = 0.25,
  disc.main.outset.B = (0.685 + 0.889)/2, disc.main.outset.W = (0.685 +
  0.889)/2, disc.main.at.diag.B = 1, disc.main.at.diag.W = 1,
  disc.main.post.diag.B = 0, disc.main.post.diag.W = 0,
  disc.pers.outset.B = (0.527 + 0.828)/2, disc.pers.outset.W = (0.527 +
  0.828)/2, disc.pers.at.diag.B = 1, disc.pers.at.diag.W = 1,
  disc.pers.post.diag.B = 0, disc.pers.post.diag.W = 0,
  disc.inst.B = (0.445 + 0.691)/2, disc.inst.W = (0.445 + 0.691)/2,
  circ.prev.B = 0.874, circ.prev.W = 0.918, ccr5.freq.B = c(0.0105,
  0.105), ccr5.freq.W = c(0.0105, 0.105), ccr5.heteroz.rr = 0.3,
  num.inst.ai.classes = 1, base.exp.ai.main.BB = 1.19,
  base.exp.ai.main.BW = 1.79, base.exp.ai.main.WW = 1.56,
  base.exp.ai.pers.BB = 0.75, base.exp.ai.pers.BW = 1.13,
  base.exp.ai.pers.WW = 0.98, incr.exp.ai.full.supp.main = 0,
  incr.exp.ai.part.supp.main = 0, redux.exp.ai.diag.main = 0,
  redux.exp.ai.discl.main = 0, incr.exp.ai.full.supp.pers = 0,
  incr.exp.ai.part.supp.pers = 0, redux.exp.ai.diag.pers = 0,
  redux.exp.ai.discl.pers = 0, cprob.main.BB = 0.38, cprob.main.BW = 0.1,
  cprob.main.WW = 0.15, cprob.pers.BB = 0.39, cprob.pers.BW = 0.11,
  cprob.pers.WW = 0.16, cprob.inst.BB = 0.49, cprob.inst.BW = 0.15,
  cprob.inst.WW = 0.22, beta.cond.fsupp.main = 0,
  beta.cond.psupp.main = 0, beta.cond.diag.main = -0.67,
  beta.cond.discl.main = -0.85, beta.cond.fsupp.pers = 0,
  beta.cond.psupp.pers = 0, beta.cond.diag.pers = -0.67,
  beta.cond.discl.pers = -0.85, beta.cond.fsupp.inst = 0,
  beta.cond.psupp.inst = 0, beta.cond.diag.inst = -0.67,
  beta.cond.discl.inst = -0.85, vv.prob.iev.BB = 0.42,
  vv.prob.iev.BW = 0.56, vv.prob.iev.WW = 0.49, ...)
}
\arguments{
\item{nwstats}{Target statistics for the network model. An object of class
"nwstats", which emerges from \code{\link{calc_nwstats.mard}}.}

\item{last.neg.test.time.range.B}{Time range for last negative test for
black men.}

\item{mean.test.interv.B}{The mean intertest interval for Black MSM who test.}

\item{last.neg.test.time.range.W}{Time range for last negative test for
white men.}

\item{mean.test.interv.W}{The mean intertest interval for White MSM who test}

\item{testing.pattern}{Whether HIV testing is memoryless (occurs with fixed
probabilty, without regard to time since previous test) or occurs in
fixed intervals.  Options are "memoryless" or "interval".}

\item{test.window.period}{The length of the HIV test's window period.}

\item{tt.traj.freq.B}{The proportion of Black MSM who enter into the four
testing/treatment trajectories.  Values represent, in order, the
propotions who are "NN" (never test or treat), "YN" (test, but then
never initiate treatment), "YP" (test, and then when on treatment
achieve partial suppression), "YF" (test, and then when on treatment
achieve full suppression).}

\item{tt.traj.freq.W}{The proportion of White MSM who enter into the four
testing/treatment trajectories.  Values represent, in order, the
propotions who are "NN" (never test or treat), "YN" (test, but then
never initiate treatment), "YP" (test, and then when on treatment
achieve partial suppression), "YF" (test, and then when on treatment
achieve full suppression).}

\item{prob.tx.init.B}{The probability per time step that a Black MSM who has
tested positive will initiate treatment for the very first time.}

\item{prob.tx.init.W}{The probability per time step that a White MSM who has
tested positive will initiate treatment for the very first time.}

\item{prob.tx.halt.B}{The probability per time step that a Black MSM who is
currently on treatment will halt treatment.}

\item{prob.tx.halt.W}{The probability per time step that a White MSM who is
currently on treatment will halt treatment.}

\item{prob.tx.reinit.B}{The probability per time step that a Black MSM who is
not currently on treatment but who has been in the past will
re-initiate treatment.}

\item{prob.tx.reinit.W}{The probability per time step that a White MSM who is
not currently on treatment but who has been in the past will
re-initiate treatment.}

\item{max.time.off.tx.full}{The number of time steps off treatment that a man
who is a full suppressor while on treatment wil last before the onset
of the AIDS stage.  Note that this includes the time before diagnosis.}

\item{max.time.on.tx.part}{The number of time steps on treatment that a man
who is a partial suppressor wil last before the onset of the AIDS
stage.}

\item{max.time.off.tx.part}{The number of time steps off treatment that a man
who is a partial suppressor while on treatment wil last before the
onset of the AIDS stage.  Note that this includes the time before
diagnosis.}

\item{vl.acute.rise.dur}{The number of time steps over which viral load rises
during the initial phase of acute infection.}

\item{vl.acute.peak}{Peak viral load (in log10 units) at the height of acute
infection.}

\item{vl.acute.fall.dur}{The number of time steps over which viral load falls
during the latter phase of acute infection.}

\item{vl.set.point}{Set point viral load for all men (in log10 units).}

\item{vl.aids.onset}{The number of time steps of infection after which the
AIDS stage will initiate for a man who has never been on treatment.}

\item{vl.aids.dur}{The duration of the AIDS stage.}

\item{vl.fatal}{The viral load obtained at the end of the AIDS stage, at
which death from AIDS occurs.  Note that the code can handle
param$vl.fatal being lower than param$vl.acute.peak, as it tests for
both a sufficiently high viral load and presence in the AIDS stage to
deterine AIDS deaths.}

\item{vl.full.supp}{Log10 viral load when fully suppressed.}

\item{vl.part.supp}{Log10 viral load when partially suppressed}

\item{full.supp.down.slope}{For men who are full suppressors, the number of
log10 units that viral load falls per time step from treatment
initiation or re-initiation until the new level is hit.}

\item{full.supp.up.slope}{For men who are full suppressors, the number of
log10 units that viral load rises per time step from treatment halting
until the new level is hit.}

\item{part.supp.down.slope}{For men who are partial suppressors, the number
of log10 units that viral load falls per time step from treatment
initiation or re-initiation until the new level is hit.}

\item{part.supp.up.slope}{For men who are partial suppressors, the number of
log10 units that viral load rises per time step from treatment halting
until the new level is hit.}

\item{b.rate.B}{The rate at which Black MSM enter the population.}

\item{b.rate.W}{The rate at which White MSM enter the population.}

\item{birth.age}{The age (in years) of new arrivals.}

\item{betabase.URAI}{Transmissibility for a man having unprotected receptive
anal intercourse with a positive man at set point viral load.}

\item{betabase.UIAI}{Transmissibility for an uncircumcised man having
unprotected insertive anal intercourse with a positive man at set
point viral load.}

\item{betamult.acute}{The factor by which acute infection is
extra-infectious, above and beyond that predicted by elevated viral
load.}

\item{betamult.circ}{The factor by which to multiply infectivity from
insertive anal sex when the negative insertive partner is circumcised.}

\item{betamult.condom}{The factor by which to multiply infectivity for anal
sex when a condom is used.}

\item{disc.main.outset.B}{The probability that an HIV positive Black MSM will
disclose his status at the start of a main partnership.}

\item{disc.main.outset.W}{The probability that an HIV positive White MSM will
disclose his status at the start of a main partnership.}

\item{disc.main.at.diag.B}{The probability that a Black MSM already in a main
partnership will disclose at the time of diagnosis.}

\item{disc.main.at.diag.W}{The probability that a White MSM already in a main
partnership will disclose at the time of diagnosis.}

\item{disc.main.post.diag.B}{The probability that an HIV positive Black MSM
in a main partnership will disclose his status, assuming he didn't
at the start of the partnership or at diagnosis.}

\item{disc.main.post.diag.W}{The probability that an HIV positive White MSM
in a main partnership will disclose his status, assuming he didn't
at the start of the partnership or at diagnosis.}

\item{disc.pers.outset.B}{The probability that an HIV positive Black MSM will
disclose his status at the start of a casual partnership.}

\item{disc.pers.outset.W}{The probability that an HIV positive White MSM will
disclose his status at the start of a casual partnership.}

\item{disc.pers.at.diag.B}{The probability that a Black MSM already in a
casual partnership will disclose at the time of diagnosis.}

\item{disc.pers.at.diag.W}{The probability that a White MSM already in a
casual partnership will disclose at the time of diagnosis.}

\item{disc.pers.post.diag.B}{The probability that an HIV positive Black MSM
in a casual partnership will disclose his status, assuming he
didn't at the start of the partnership or at diagnosis.}

\item{disc.pers.post.diag.W}{The probability that an HIV positive White MSM
in a casual partnership will disclose his status, assuming he
didn't at the start of the partnership or at diagnosis.}

\item{disc.inst.B}{The probability that an HIV positive Black MSM will
disclose his status to a one-off partner.}

\item{disc.inst.W}{The probability that an HIV positive White MSM will
disclose his status to a one-off partner.}

\item{circ.prev.B}{The probablity that a Black new arrival in the population
will be circumcised.}

\item{circ.prev.W}{The probablity that a White new arrival in the population
will be circumcised.}

\item{ccr5.freq.B}{A vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respecitively) in the CCR5 gene
among Black MSM.}

\item{ccr5.freq.W}{A vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respecitively) in the CCR5 gene
among White MSM.}

\item{ccr5.heteroz.rr}{Relative risk of infection for men who are heterzygous
in the CCR5 mutation.}

\item{num.inst.ai.classes}{The number of quantiles into which men should be
divided in determining their levels of one-off AI.}

\item{base.exp.ai.main.BB}{Expected coital frequency in Black-Black main
partnerships (acts per week).}

\item{base.exp.ai.main.BW}{Expected coital frequency in Black-White main
partnerships (acts per week).}

\item{base.exp.ai.main.WW}{Expected coital frequency in White-White main
partnerships (acts per week).}

\item{base.exp.ai.pers.BB}{Expected coital frequency in Black-Black casual
partnerships (acts per week).}

\item{base.exp.ai.pers.BW}{Expected coital frequency in Black-White casual
partnerships (acts per week).}

\item{base.exp.ai.pers.WW}{Expected coital frequency in White-White casual
partnerships (acts per week).}

\item{incr.exp.ai.full.supp.main}{Percent increase in expected coital
frequency in main partnerships associated with full suppression of
the disease.}

\item{incr.exp.ai.part.supp.main}{Percent increase in expected coital
frequency in main partnerships associated with partial suppression of
the disease.}

\item{redux.exp.ai.diag.main}{Percent reduction in expected coital frequency
in main partnerships associated with diagnosis.}

\item{redux.exp.ai.discl.main}{Percent reduction in expected coital frequency
in main partnerships associated with disclosure of diagnosis.}

\item{incr.exp.ai.full.supp.pers}{Percent increase in expected coital
frequency in casual partnerships associated with full suppression of
the disease.}

\item{incr.exp.ai.part.supp.pers}{Percent increase in expected coital
frequency in casual partnerships associated with partial suppression of
the disease.}

\item{redux.exp.ai.diag.pers}{Percent reduction in expected coital frequency
in casual partnerships associated with diagnosis.}

\item{redux.exp.ai.discl.pers}{Percent reduction in expected coital frequency
in casual partnerships associated with disclosure of diagnosis.}

\item{cprob.main.BB}{The probability of condom use in a Black-Black main
partnership.}

\item{cprob.main.BW}{The probability of condom use in a Black-White main
partnership.}

\item{cprob.main.WW}{The probability of condom use in a White-White main
partnership.}

\item{cprob.pers.BB}{The probability of condom use in a Black-Black casual
partnership.}

\item{cprob.pers.BW}{The probability of condom use in a Black-White casual
partnership.}

\item{cprob.pers.WW}{The probability of condom use in a White-White casual
partnership.}

\item{cprob.inst.BB}{The probability of condom use in a Black-Black one-off
partnership.}

\item{cprob.inst.BW}{The probability of condom use in a Black-White one-off
partnership.}

\item{cprob.inst.WW}{The probability of condom use in a White-White one-off
partnership.}

\item{beta.cond.fsupp.main}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV positive man is fully
suppressed.}

\item{beta.cond.psupp.main}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV positive man is partially
suppressed.}

\item{beta.cond.diag.main}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV positive man has been
diagnosed.}

\item{beta.cond.discl.main}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV positive man has disclosed
his status.}

\item{beta.cond.fsupp.pers}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV positive man is fully
suppressed.}

\item{beta.cond.psupp.pers}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV positive man is partially
suppressed.}

\item{beta.cond.diag.pers}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV positive man has been
diagnosed.}

\item{beta.cond.discl.pers}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV positive man has disclosed
his status.}

\item{beta.cond.fsupp.inst}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV positive man is fully
suppressed.}

\item{beta.cond.psupp.inst}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV positive man is partially
suppressed.}

\item{beta.cond.diag.inst}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV positive man has been
diagnosed.}

\item{beta.cond.discl.inst}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV positive man has disclosed
his status.}

\item{vv.prob.iev.BB}{The probability that in a Black-Black partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{vv.prob.iev.BW}{The probability that in a Black-White partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{vv.prob.iev.WW}{The probability that in a White-White partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list object of class \code{param.mard}, which can be passed to
EpiModel function \code{netsim}.
}
\description{
Gather transmission related parameters before simulating from a
             stochastic epidemic model.
}

