% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.param_change.R
\name{param_changer}
\alias{param_changer}
\title{Module to modify the parameters at specified time steps during the simulation}
\usage{
param_changer(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module to modify the parameters at specified time steps during the simulation
}
\details{
if a list `dat$param$param_updaters` is present, this module will update the
`dat$param` list with new values at given timesteps.
`dat$param$param_updaters` is a list containing `updaters`. An updater is a
list containing an `at` element telling when the changes will happen and a
`param` element which is a named list of the parameters to change.
}
\examples{
 ## example of a `param_updaters` list
 list(
   list(
     at = 4860,
     params = list(
       hiv.test.rate = rep(0.0128, 3),
       trans.scale = c(1.61, 0.836, 0.622)
     )
   ),
   list(
     at = 5160,
     params = list(tx.init.prob = c(0.125, 0.158, 0.164))
   )
 )

}
