% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_het}
\alias{param_het}
\title{Parameters for Stochastic Network Model of HIV-1 Infection in
       Sub-Saharan Africa}
\usage{
param_het(time.unit = 7, acute.stage.mult = 5, aids.stage.mult = 1,
  vl.acute.topeak = 14, vl.acute.toset = 107, vl.acute.peak = 6.7,
  vl.setpoint = 4.5, vl.aidsmax = 7, cond.prob = 0.09,
  cond.eff = 0.78, act.rate.early = 0.362, act.rate.late = 0.197,
  act.rate.cd4 = 50, acts.rand = TRUE, circ.prob.birth = 0.9,
  circ.eff = 0.53, tx.elig.cd4 = 350, tx.init.cd4.mean = 120,
  tx.init.cd4.sd = 40, tx.adhere.full = 0.76, tx.adhere.part = 0.5,
  tx.vlsupp.time = 365/3, tx.vlsupp.level = 1.5,
  tx.cd4.recrat.feml = 11.6/30, tx.cd4.recrat.male = 9.75/30,
  tx.cd4.decrat.feml = 11.6/30, tx.cd4.decrat.male = 9.75/30,
  tx.coverage = 0.3, tx.prev.eff = 0.96, b.rate = 0.03/365,
  b.rate.method = "totpop", b.propmale = NULL, ds.exit.age = 55,
  ds.rate.mult = 1, di.cd4.aids = 50, di.cd4.rate = 2/365, ...)
}
\arguments{
\item{time.unit}{Unit of time relative to one day.}

\item{acute.stage.mult}{Acute stage multiplier for increased infectiousness
above impact of heightened viral load.}

\item{aids.stage.mult}{AIDS stage multiplier for increased infectiousness in
AIDS above impact of heightened viral load.}

\item{vl.acute.topeak}{Time in days to peak viremia during acute infection.}

\item{vl.acute.toset}{Time in days to viral set point following peak viremia.}

\item{vl.acute.peak}{Log 10 viral load at acute peak.}

\item{vl.setpoint}{Log 10 viral load at set point.}

\item{vl.aidsmax}{Maximum log 10 viral load during AIDS.}

\item{cond.prob}{Probability of condoms per act with partners.}

\item{cond.eff}{Efficacy of condoms per act in HIV prevention.}

\item{act.rate.early}{Daily per-partnership act rate in early disease.}

\item{act.rate.late}{Daily per-partnership act rate in late disease.}

\item{act.rate.cd4}{CD4 count at which the \code{act.rate.late} applies.}

\item{acts.rand}{If \code{TRUE}, will draw number of total and unprotected
acts from a binomial distribution parameterized by the
\code{act.rate}.}

\item{circ.prob.birth}{Proportion of men circumcised at birth.}

\item{circ.eff}{Efficacy of circumcision per act in HIV prevention.}

\item{tx.elig.cd4}{CD4 count at which a person becomes eligible for
treatment.}

\item{tx.init.cd4.mean}{Mean CD4 count at which person presents for care.}

\item{tx.init.cd4.sd}{SD of CD4 count at which person presents for care.}

\item{tx.adhere.full}{Proportion of people who start treatment who are fully
adherent.}

\item{tx.adhere.part}{Of the not fully adherent proportion, the percent of
time they are on medication.}

\item{tx.vlsupp.time}{Time in days from treatment initiation to
viral suppression.}

\item{tx.vlsupp.level}{Log 10 viral load level at suppression.}

\item{tx.cd4.recrat.feml}{Rate of CD4 recovery under treatment for males.}

\item{tx.cd4.recrat.male}{Rate of CD4 recovery under treatment for females.}

\item{tx.cd4.decrat.feml}{Rate of CD4 decline under periods of non-adherence
for females.}

\item{tx.cd4.decrat.male}{Rate of CD4 decline under periods of non-adherence
for males.}

\item{tx.coverage}{Proportion of treatment-eligible persons who have
initiated treatment.}

\item{tx.prev.eff}{Proportional amount by which treatment reduces infectivity
of infected partner.}

\item{b.rate}{General entry rate per day for males and females specified.}

\item{b.rate.method}{Method for assigning birth rates, with options of
"totpop" for births as a function of the total population size,
"fpop" for births as a function of the female population size, and
"stgrowth" for a constant stable growth rate.}

\item{b.propmale}{Proportion of entries assigned as male. If NULL, then set
adaptively based on the proportion at time 1.}

\item{ds.exit.age}{Age at which the age-specific ds.rate is set to 1, with NA
value indicating no censoring.}

\item{ds.rate.mult}{Simple multiplier for background death rates.}

\item{di.cd4.aids}{CD4 count at which late-stage AIDS occurs and the risk of
mortality is governed by \code{di.cd4.rate}.}

\item{di.cd4.rate}{Mortality in late-stage AIDS after hitting a nadir CD4 of
\code{di.cd4.aids}.}

\item{...}{additional arguments to be passed into model.}
}
\description{
Sets the simulation parameters for the stochastic
             network model of HIV-1 Infection among Heterosexuals in
             Sub-Saharan Africa for the \code{EpiModelHIV} package.
}
\details{
This function sets the parameters for the models.
}
\keyword{het}
