% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.KTM.R
\name{param_KTM}
\alias{param_KTM}
\title{Epidemic Model Parameters for SHAMP}
\usage{
param_KTM(
  race.method = 1,
  time.unit = 7,
  method = 1,
  age.unit = 52,
  last.neg.test.B.f.int = 1042,
  last.neg.test.B.msf.int = 1446,
  mean.test.B.f.int = 1042,
  mean.test.B.msf.int = 1446,
  test.prob.nevtest.m = c(0.00465691, 0.012052342, 0.006761833),
  test.prob.nevtest.f = c(0.006597747, 0.035640496, 0.0225),
  test.prob.tested.m = c(0.014126, 0.019322, 0.01983, 0.01352),
  test.prob.tested.f = c(0.015517, 0.021565, 0.020283, 0.015),
  test.window.int.ab = 3 * 7,
  test.window.int.rna = 1 * 7,
  test.window.int.rna.alt = 2 * 7,
  tt.traj.B.f.prob = c(0, 0, 0.23, 0.77),
  tt.traj.B.msf.prob = c(0, 0, 0.23, 0.77),
  intervention_TM = "NONE",
  seek.hc.AHI.prob = 0.325,
  sym.seek.prob = 0.002606,
  sym.test.prob.bl = 0.256,
  sym.test.prob.tm = 0.949,
  partner.test.prob.bl = 0.023,
  partner.test.prob.tm = 0.169,
  p.prev.follow.win = 52,
  p.acute.follow.win = 12,
  PS.time = 6,
  tx.init.PS.prob = 0.159,
  ab.test.sens = 0.999 * 1,
  ab.test.spec = (0.998 + 0.994) - (0.998 * 0.994),
  rna.test.sens = 0.989,
  rna.test.spec = 1,
  rna.alt.test.acute.sens = 0.989,
  rna.alt.test.acute.spec = 1,
  rna.alt.test.prev.sens = 0.989,
  rna.alt.test.prev.spec = 1,
  tx.init.TMP.prob = 0.31,
  tx.init.B.f.prob = 0.15912,
  tx.init.BI.f.prob = 0.15912,
  tx.init.H.f.prob = 0.15912,
  tx.init.HI.f.prob = 0.15912,
  tx.init.W.f.prob = 0.15912,
  tx.init.B.msf.prob = 0.15912,
  tx.init.BI.msf.prob = 0.15912,
  tx.init.H.msf.prob = 0.15912,
  tx.init.HI.msf.prob = 0.15912,
  tx.init.W.msf.prob = 0.15912,
  tx.init.B.msm.prob = 0.5,
  tx.init.BI.msm.prob = 0.5,
  tx.init.H.msm.prob = 0.5,
  tx.init.HI.msm.prob = 0.5,
  tx.init.W.msm.prob = 0.5,
  tx.init.B.msmf.prob = 0.5,
  tx.init.BI.msmf.prob = 0.5,
  tx.init.H.msmf.prob = 0.5,
  tx.init.HI.msmf.prob = 0.5,
  tx.init.W.msmf.prob = 0.5,
  tx.halt.B.f.prob = 0.0020245,
  tx.halt.BI.f.prob = 0.0020245,
  tx.halt.H.f.prob = 0.0020245,
  tx.halt.HI.f.prob = 0.0020245,
  tx.halt.W.f.prob = 0.0020245,
  tx.halt.B.msf.prob = 0.0020245,
  tx.halt.BI.msf.prob = 0.0020245,
  tx.halt.H.msf.prob = 0.0020245,
  tx.halt.HI.msf.prob = 0.0020245,
  tx.halt.W.msf.prob = 0.0020245,
  tx.halt.B.msm.prob = 0.5,
  tx.halt.BI.msm.prob = 0.5,
  tx.halt.H.msm.prob = 0.5,
  tx.halt.HI.msm.prob = 0.5,
  tx.halt.W.msm.prob = 0.5,
  tx.halt.B.msmf.prob = 0.5,
  tx.halt.BI.msmf.prob = 0.5,
  tx.halt.H.msmf.prob = 0.5,
  tx.halt.HI.msmf.prob = 0.5,
  tx.halt.W.msmf.prob = 0.5,
  tx.reinit.B.f.prob = 0.01,
  tx.reinit.BI.f.prob = 0.01,
  tx.reinit.H.f.prob = 0.01,
  tx.reinit.HI.f.prob = 0.01,
  tx.reinit.W.f.prob = 0.01,
  tx.reinit.B.msf.prob = 0.01,
  tx.reinit.BI.msf.prob = 0.01,
  tx.reinit.H.msf.prob = 0.01,
  tx.reinit.HI.msf.prob = 0.01,
  tx.reinit.W.msf.prob = 0.01,
  tx.reinit.B.msm.prob = 1,
  tx.reinit.BI.msm.prob = 1,
  tx.reinit.H.msm.prob = 1,
  tx.reinit.HI.msm.prob = 1,
  tx.reinit.W.msm.prob = 1,
  tx.reinit.B.msmf.prob = 1,
  tx.reinit.BI.msmf.prob = 1,
  tx.reinit.H.msmf.prob = 1,
  tx.reinit.HI.msmf.prob = 1,
  tx.reinit.W.msmf.prob = 1,
  max.time.off.tx.full.int = 3297,
  max.time.on.tx.part.int = 52 * 15 * 7,
  max.time.off.tx.part.int = 3297,
  vl.acute.rise.int = 12,
  vl.acute.peak = 6.76,
  vl.acute.fall.int = 20,
  vl.set.point = 4,
  vl.aids.onset.int = 3297,
  vl.aids.int = 280,
  vl.fatal = 7,
  vl.full.supp = 2.75,
  vl.part.supp = 3.5,
  full.supp.down.slope = 0.11,
  full.supp.up.slope = 0.11,
  part.supp.down.slope = 0.11,
  part.supp.up.slope = 0.11,
  birth.age = 18,
  exit.age = 76,
  age.adj = 7.925357,
  b.method = "fixed",
  msm.frac = 0,
  msmf.frac.B = 0,
  msmf.frac.BI = 0,
  msmf.frac.H = 0,
  msmf.frac.HI = 0,
  msmf.frac.W = 0,
  percent.male = 0.4868,
  URAI.prob = 138/10000,
  UIAI.prob = 11/10000,
  URVI.prob = 0.0182,
  UIVI.prob = 0.0182,
  VI.foi.scale = 1,
  AI.foi.scale = 1,
  acute.rr = 4,
  circ.rr = 0.4,
  condom.rr = 0.25,
  disc.outset.main.prob = 0.03,
  disc.at.diag.main.prob = 0.03,
  disc.post.diag.main.prob = 0,
  disc.outset.pers.prob = 0.02,
  disc.at.diag.pers.prob = 0.02,
  disc.post.diag.pers.prob = 0,
  disc.inst.prob = 0.01,
  circ.B.prob = 0.925,
  circ.BI.prob = 0.5,
  circ.H.prob = 0.5,
  circ.HI.prob = 0.5,
  circ.W.prob = 0.5,
  ccr5.B.f.prob = c(0, 0),
  ccr5.BI.f.prob = c(0, 0),
  ccr5.H.f.prob = c(0, 0),
  ccr5.HI.f.prob = c(0, 0),
  ccr5.W.f.prob = c(0, 0),
  ccr5.B.m.prob = c(0, 0),
  ccr5.BI.m.prob = c(0, 0),
  ccr5.H.m.prob = c(0, 0),
  ccr5.HI.m.prob = c(0, 0),
  ccr5.W.m.prob = c(0, 0),
  ccr5.heteroz.rr = 0,
  num.inst.ai.classes = 1,
  base.ai.main.B.rate = 1.379/7,
  base.ai.main.BI.rate = 1.379/7,
  base.ai.main.H.rate = 1.379/7,
  base.ai.main.HI.rate = 1.379/7,
  base.ai.main.W.rate = 1.379/7,
  base.ai.pers.B.rate = 0.883/7,
  base.ai.pers.BI.rate = 0.883/7,
  base.ai.pers.H.rate = 0.883/7,
  base.ai.pers.HI.rate = 0.883/7,
  base.ai.pers.W.rate = 0.883/7,
  base.vi.main.B.rate = 0.4818,
  base.vi.main.BI.rate = 0,
  base.vi.main.H.rate = 0,
  base.vi.main.HI.rate = 0,
  base.vi.main.W.rate = 0,
  base.vi.pers.B.rate = 0.2311,
  base.vi.pers.BI.rate = 0,
  base.vi.pers.H.rate = 0,
  base.vi.pers.HI.rate = 0,
  base.vi.pers.W.rate = 0,
  ai.scale = 1,
  vi.scale = 1,
  cond.main.B.prob.msm = 0.5,
  cond.main.BI.prob.msm = 0.5,
  cond.main.H.prob.msm = 0.5,
  cond.main.HI.prob.msm = 0.5,
  cond.main.W.prob.msm = 0.5,
  cond.pers.always.prob.msm = 1e-04,
  cond.pers.B.prob.msm = 0.5,
  cond.pers.BI.prob.msm = 0.5,
  cond.pers.H.prob.msm = 0.5,
  cond.pers.HI.prob.msm = 0.5,
  cond.pers.W.prob.msm = 0.5,
  cond.inst.always.prob.msm = 1e-04,
  cond.inst.B.prob.msm = 0.5,
  cond.inst.BI.prob.msm = 0.5,
  cond.inst.H.prob.msm = 0.5,
  cond.inst.HI.prob.msm = 0.5,
  cond.inst.W.prob.msm = 0.5,
  cond.always.prob.corr.msm = 0.001,
  cond.main.always.prob.het = 0.01,
  cond.main.B.prob.het = 0.04,
  cond.pers.always.prob.het = 0.2,
  cond.pers.B.prob.het = 0.09,
  cond.inst.always.prob.het = 0.46,
  cond.inst.B.prob.het = 0.11,
  cond.always.prob.corr.het = 0.5,
  cond.diag.main.beta.msm = -0.67,
  cond.discl.main.beta.msm = -0.85,
  cond.diag.pers.beta.msm = -0.67,
  cond.discl.pers.beta.msm = -0.85,
  cond.diag.inst.beta.msm = -0.67,
  cond.discl.inst.beta.msm = -0.85,
  cond.diag.main.beta.het = -0.67,
  cond.discl.main.beta.het = -0.85,
  cond.diag.pers.beta.het = -0.67,
  cond.discl.pers.beta.het = -0.85,
  cond.diag.inst.beta.het = -0.67,
  cond.discl.inst.beta.het = -0.85,
  role.B.msm.prob = c(0.242, 0.321, 0.437),
  role.BI.msm.prob = c(0.242, 0.321, 0.437),
  role.H.msm.prob = c(0.242, 0.321, 0.437),
  role.HI.msm.prob = c(0.242, 0.321, 0.437),
  role.W.msm.prob = c(0.242, 0.321, 0.437),
  role.B.msmf.prob = c(0.242, 0.321, 0.437),
  role.BI.msmf.prob = c(0.242, 0.321, 0.437),
  role.H.msmf.prob = c(0.242, 0.321, 0.437),
  role.HI.msmf.prob = c(0.242, 0.321, 0.437),
  role.W.msmf.prob = c(0.242, 0.321, 0.437),
  vv.iev.B.prob = 0.47874,
  vv.iev.BI.prob = 0.47874,
  vv.iev.H.prob = 0.47874,
  vv.iev.HI.prob = 0.47874,
  vv.iev.W.prob = 0.47874,
  riskh.start = Inf,
  prep.start = Inf,
  prep.ps = FALSE,
  prep.disc = FALSE,
  prep.risk = "NONE",
  prep.risk.int = 12,
  prep.class.prob = c(0, 0, 0, 1),
  prep.class.hr = c(0, 0, 0, 0.29),
  prep.tst.int = 90,
  prep.start.prob = 0.236,
  prep.stop.prob = 0.0092,
  conc_dur_dx = TRUE,
  cel.complete.trim = TRUE,
  cel.complete.lag = 52,
  kill.poi.rels = TRUE,
  death_stats = FALSE,
  p.growth = FALSE,
  p.growth.size = 0,
  ...
)
}
\arguments{
\item{test.window.int.ab}{The test window of an antibody test}

\item{test.window.int.rna}{The test window of an RNA test}

\item{intervention_TM}{The intervention state of KTM plus}

\item{seek.hc.AHI.prob}{The probability of seeking health care while experiencing Acute HIV symptoms}

\item{sym.seek.prob}{the joint probability of experiencing acute HIV LIKE symptoms and seeking treatment}

\item{sym.test.prob.bl}{The probability of PITC for those presenting with symptoms with no intervention}

\item{sym.test.prob.tm}{The probability of PITC for those presenting with symtoms with the KTM plus intervention}

\item{partner.test.prob.bl}{Probability of finding and testing a partner of a positive case without the KTM plus intervention}

\item{partner.test.prob.tm}{Probability of finding and testing a partner of a positive case with the KTM plus intervention}

\item{p.prev.follow.win}{The number of weeks back partner services will look for partners of prevalent cases}

\item{p.acute.follow.win}{The number of weeks back partner services will look for partners of acute cases}

\item{PS.time}{How long partner services will continue to look for partners in weeks}

\item{tx.init.PS.prob}{Probability someone identified as positive via partner services will initiate treatment}

\item{ab.test.sens}{The sensitivity of the antibody test}

\item{ab.test.spec}{The specificity of the antibody test}

\item{rna.test.sens}{The sensitivity of the RNA test}

\item{rna.test.spec}{The specificity of the RNA test}

\item{max.time.off.tx.full.int}{Number of days off treatment for a full
suppressor before onset of AIDS, including time before diagnosis.}

\item{max.time.on.tx.part.int}{Number of days on treatment for a
partial suppressor before onset of AIDS.}

\item{max.time.off.tx.part.int}{Nnumber of days off treatment for a
partial suppressor before onset of AIDS, including time before
diagnosis.}

\item{vl.acute.rise.int}{Number of days to peak viremia during acute infection.}

\item{vl.acute.peak}{Peak viral load (in log10 units) at the height of acute infection.}

\item{vl.acute.fall.int}{Number of days from peak viremia to set-point
viral load during the acute infection period.}

\item{vl.set.point}{Set point viral load (in log10 units).}

\item{vl.aids.onset.int}{Number of days to AIDS for a treatment-naive patient.}

\item{vl.aids.int}{Duration of AIDS stage infection in days.}

\item{vl.fatal}{Viral load in AIDS at which death occurs.}

\item{vl.full.supp}{Log10 viral load at full suppression on ART.}

\item{vl.part.supp}{Log10 viral load at partial suppression on ART.}

\item{full.supp.down.slope}{For full suppressors, number of log10 units that
viral load falls per time step from treatment initiation or re-initiation
until the level in \code{vl.full.supp}.}

\item{full.supp.up.slope}{For full suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected value.}

\item{part.supp.down.slope}{For partial suppressors, number of log10 units
that viral load falls per time step from treatment initiation or
re-initiation until the level in \code{vl.part.supp}.}

\item{part.supp.up.slope}{For partial suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected value.}

\item{birth.age}{Age (in years) of new arrivals.}

\item{age.adj}{This value is added to the age for females in age.adj
so that sex age asymmetry can be captured with an absdiff term}

\item{b.method}{Method for calculating the number of expected births at each
time step, with \code{"fixed"} based on the number of persons at the
initial time step and \code{"varying"} based on the current time step.}

\item{msm.frac}{The fraction of males entering the population at each time step 
as a male that has intercourse exclusivly with other males "msm" - 
males not designated as msm or msmf are considered "msf".}

\item{URAI.prob}{Probability of transmission for a man having unprotected
receptive anal intercourse with an infected man at set point viral
load.}

\item{UIAI.prob}{Probability of transmission for an uncircumcised man having
unprotected insertive anal intercourse with an infected man at set
point viral load.}

\item{URVI.prob}{Probability of transmission for a woman having unprotected
receptive vaginal intercourse with an infected man at set point viral
load.}

\item{UIVI.prob}{Probability of transmission for an uncircumcised man having
unprotected insertive vaginal intercourse with an infected women at set
point viral load.}

\item{acute.rr}{Relative risk of infection (compared to that predicted by
elevated viral load) when positive partner is in the acute stage.}

\item{circ.rr}{Relative risk of infection from insertive anal sex when the
negative insertive partner is circumcised.}

\item{condom.rr}{Relative risk of infection from anal sex when a condom is
used.}

\item{ccr5.heteroz.rr}{Relative risk of infection for people who are heterozygous
in the CCR5 mutation.}

\item{num.inst.ai.classes}{Number of quantiles into which men should be
divided in determining their levels of one-off anal intercourse.}

\item{ai.scale}{General relative scaler for all ai act rates for model
calibration.}

\item{vi.scale}{General relative scaler for all vi act rates for model
calibration.}

\item{cond.pers.always.prob.msm}{Fraction of men in male-male casual partnerships who always
use condoms in those partnerships.}

\item{cond.inst.always.prob.msm}{Fraction of men in male-male one-time partnerships who always
use condoms in those partnerships.}

\item{cond.always.prob.corr.msm}{Correlation coefficient for probability of always
using condoms in both casual and one-off male-male partnerships}

\item{cond.pers.always.prob.het}{Fraction of people in heterosexual casual partnerships who always
use condoms in those partnerships.}

\item{cond.inst.always.prob.het}{Fraction of people in heterosexual one-time partnerships who always
use condoms in those partnerships.}

\item{cond.always.prob.corr.het}{Correlation coefficient for probability of always
using condoms in both casual and one-off heterosexual partnerships}

\item{cond.diag.main.beta.msm}{Beta multiplier for the log odds of using a
condom in a male-male main partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.main.beta.msm}{Beta multiplier for the log odds of using a
condom in a male-male main partnership if the HIV-infected man has disclosed.}

\item{cond.diag.pers.beta.msm}{Beta multiplier for the log odds of using a
condom in a male-male casual partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.pers.beta.msm}{Beta multiplier for the log odds of using a
condom in a male-male casual partnership if the HIV-infected man has disclosed
his status.}

\item{cond.diag.inst.beta.msm}{Beta multiplier for the log odds of using a
condom in a male-male one-off partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.inst.beta.msm}{Beta multiplier for the log odds of using a
condom in a male-male one-off partnership if the HIV-infected man has disclosed
his status.}

\item{cond.diag.main.beta.het}{Beta multiplier for the log odds of using a
condom in a heterosexual main partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.main.beta.het}{Beta multiplier for the log odds of using a
condom in a heterosexual main partnership if the HIV-infected man has disclosed.}

\item{cond.diag.pers.beta.het}{Beta multiplier for the log odds of using a
condom in a heterosexual casual partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.pers.beta.het}{Beta multiplier for the log odds of using a
condom in a heterosexual casual partnership if the HIV-infected man has disclosed
his status.}

\item{cond.diag.inst.beta.het}{Beta multiplier for the log odds of using a
condom in a heterosexual one-off partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.inst.beta.het}{Beta multiplier for the log odds of using a
condom in a heterosexual one-off partnership if the HIV-infected man has disclosed
his status.}

\item{prep.start}{Time step at which the PrEP intervention should start.}

\item{prep.risk.int}{Time window for assessment of risk eligibility for PrEP
in days.}

\item{prep.class.prob}{The probability of adherence class in non-adherent,
low adherence, medium adherence, or high adherence groups.}

\item{prep.class.hr}{The hazard ratio for infection per act associated with each
level of adherence (from Patel et all and Lasry et al both based on studies by Baetan et al and Thigpen et al).}

\item{prep.tst.int}{Testing interval for those who are actively on PrEP. This
overrides the mean testing interval parameters.}

\item{conc_dur_dx}{If TRUE the simulation will store cel.temp (a current edgelist with partnership start time) 
and cel.complete (a complete list of all partnerships including start and end times.  These are used for the analysis of concurrency
and multiplexity.}

\item{death_stats}{IF TRUE the data.frame death.stats stores all care continuum information for individuals at the time of death.
HIV related death data are store in death.stats, care continuum information for those that age out positive is store in death.stats$age.out.}

\item{p.growth}{If TRUE the number of additional births per time step can be set with 'p.growth.size' and the number of time steps with 'p.growth.nsteps'.}

\item{p.growth.size}{The number of additioan entries made at each time step 2 to p.growth.nsteps if p.growth =TRUE.}

\item{...}{Additional arguments passed to the function.}

\item{last.neg.test.R.S.int}{Time range in days for last negative test for
race group R where R is a charater (B, BI, H, HI or W) and sex group S
where s is a charter (f, msf, msm, msmf).}

\item{mean.test.B.int}{Mean intertest interval in days for race group R where R 
is a charater (B, BI, H, HI or W) and sex group S
where s is a charter (f, msf, msm, msmf) who test.}

\item{tt.traj.S.prob}{Proportion of race group R, where R is a charater (B, BI, H, HI or W), 
and sex group S, where s is a charter (f, msf, msm, msmf), who enter one of four
testing/treatment trajectories: never test or treat, test and never
initiate treatment, test and treated with partial viral suppression,
and test and treated with full suppression.}

\item{tx.init.R.S.prob}{Probability per time step that a person in race group R, where R is a 
charater (B, BI, H, HI or W), and sex group S, where s is a charter (f, msf, msm, msmf),
who has tested positive will initiate treatment for the first 8 weeks following diagnosis.}

\item{tx.halt.R.S.prob}{Probability per time step that a person in race group R, where R is a 
charater (B, BI, H, HI or W), and sex group S, where s is a charter (f, msf, msm, msmf),
who is currently on treatment will halt treatment.}

\item{tx.reinit.R.S.prob}{Probability per time step that a person in race group R, where R is a 
charater (B, BI, H, HI or W), and sex group S, where s is a charter (f, msf, msm, msmf),
who is not currently on treatment but who has been in the past will
re-initiate treatment or that someone who did not initiate treatment within 8 weeks of diagnosis will initiate treatment.}

\item{b.R.S.rate}{Rate at which a person in race group R, where R is a 
charater (B, BI, H, HI or W), and sex group S, where s is a charter (f, msf, msm, msmf),
enters the population.}

\item{msmf.frac}{The fraction of males entering the population at each time step as
as a male that has intercourse with both males and females "msmf" - 
males not designated as msm or msmf are considered "msf".}

\item{disc.outset.main.R.S.prob}{Probability that an HIV-infected person in race group R,
where R is a charater (B, BI, H, HI or W), and sex group S, where s is a charter 
(f, msf, msm, msmf), will disclose their status at the start of a main partnership.}

\item{disc.at.diag.main.R.S.prob}{Probability that a a person in race group R, where R is a 
charater (B, BI, H, HI or W), and sex group S, where s is a charter (f, msf, msm, msmf),
already in a main partnership will disclose at the time of diagnosis.}

\item{disc.post.diag.main.R.S.prob}{Probability that an HIV-infected a person in race group R, 
where R is a charater (B, BI, H, HI or W), and sex group S, where s is a charter 
(f, msf, msm, msmf), in a main partnership will disclose their status, assuming they didn't
at the start of the partnership or at diagnosis.}

\item{disc.outset.pers.R.S.prob}{Probability that an HIV-infected a person in race group R, 
where R is a charater (B, BI, H, HI or W), and sex group S, where s is a charter
(f, msf, msm, msmf), will disclose their status at the start of a casual partnership.}

\item{disc.at.diag.pers.R.S.prob}{Probability that a person in race group R, where R is a
charater (B, BI, H, HI or W), and sex group S, where s is a charter (f, msf, msm, msmf),
already in a casual partnership will disclose at the time of diagnosis.}

\item{disc.post.diag.pers.R.S.prob}{Probability that an HIV-infected person in race group R,
where R is a charater (B, BI, H, HI or W), and sex group S, where s is a charter 
(f, msf, msm, msmf), in a casual partnership will disclose their status, assuming they
didn't at the start of the partnership or at diagnosis.}

\item{disc.inst.R.S.prob}{Probability that an HIV-infected person in race group R, 
where R is a charater (B, BI, H, HI or W), and sex group S, where s is a charter
(f, msf, msm, msmf), will disclose ther status to a one-off partner.}

\item{circ.R.prob}{Probablity that a male in race group R, where R is a 
charater (B, BI, H, HI or W), newly arriving in the population
will be circumcised.}

\item{ccr5.R.S.prob}{Vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respectively) in the CCR5 gene
among people in race group R, where R is a charater (B, BI, H, HI or W), 
and sex group S, where s is a charter (f or m).}

\item{base.ai.main.R.rate}{Expected coital frequency (for anal intercourse) for males 
(msm and msmf) in race group R, where R is a charater (B, BI, H, HI or W), 
in main partnerships (acts per day).}

\item{base.vi.main.R.rate}{Expected coital frequency (for vaginal intercourse) for 
males and females in race group R, where R is a charater (B, BI, H, HI or W), 
in main partnerships (acts per day).}

\item{base.ai.pers.R.rate}{Expected coital frequency (for anal intercourse) for males 
(msm and msmf) in race group R, where R is a charater (B, BI, H, HI or W), 
in casual partnerships (acts per day).}

\item{base.vi.pers.R.rate}{Expected coital frequency (for vaginal intercourse) for 
males and females in race group R, where R is a charater (B, BI, H, HI or W), 
in casual partnerships (acts per day).}

\item{cond.main.R.prob.msm}{Probability of condom use by an male of race group R, 
where R is a charater (B, BI, H, HI or W), in a male-male main partnership.}

\item{cond.pers.R.prob.msm}{Of men who are not consistent condom users, per-act
probability of condom use in a casual male-male partnership by 
males of race group R, where R is a charater (B, BI, H, HI or W).}

\item{cond.inst.R.prob.msm}{Of men who are not consistent condom users, per-act
probability of condom use in a one-time male-male partnership by 
males of race group R, where R is a charater (B, BI, H, HI or W).}

\item{cond.main.R.prob.het}{Probability of condom use by an person of race group R, 
where R is a charater (B, BI, H, HI or W), in a heterosexual main partnership.}

\item{cond.pers.R.prob.het}{Of people who are not consistent condom users, per-act
probability of condom use in a casual heterosexual partnership by 
people of race group R, where R is a charater (B, BI, H, HI or W).}

\item{cond.inst.R.prob.het}{Of people who are not consistent condom users, per-act
probability of condom use in a one-time heterosexual partnership by 
males of race group R, where R is a charater (B, BI, H, HI or W).}

\item{vv.iev.R.prob}{Probability that a person of group R, where R is a charater (B, BI, H, HI or W),
in a male male relationship will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{prep.elig.model}{Modeling approach for determining who is eligible for
PrEP. Current options are limited to: \code{"all"} for all persons who
have never been on PrEP and are disease-susceptible.}

\item{prep.coverage}{The proportion of the eligible population who are start
PrEP once they become eligible.}

\item{prep.cov.method}{The method for calculating PrEP coverage, with options
of \code{"curr"} to base the numerator on the number of people currently
on PrEP and \code{"ever"} to base it on the number of people ever on
PrEP.}

\item{prep.cov.rate}{The rate at which persons initiate PrEP conditional on
their eligibility, with 1 equal to instant start.}

\item{prep.risk.reassess}{If \code{TRUE}, reassess eligibility for PrEP at
each testing visit.}

\item{immig.depart.R.S}{The probability per time step that a person of group R,
where R is a charater (BI or HI), and sex S, were S is a charater (f or m), leaves the population
being simulated by returning to their home country.}

\item{immig.return.R.S}{The probability per time step that a person of group R,
where R is a charater (BI or HI), and sex S, were S is a charater (f or m), who has left
the population being simulated by returning to their home country returns to the population
being simulated.}

\item{immig.aq.prob.R.S}{The probability per time step that a person of group R,
where R is a charater (BI or HI), and sex S, were S is a charater (f or m), who has left
the population being simulated aquires HIV while in their home country.}

\item{pos.entry.BI}{The probability that a BI entering into the population that is etering at an
age older than the birth age (replacing deaths from mortality proportional to the demographic
distribution and approximating in-migration) is HIV positive}

\item{pos.entry.HI}{The probability that a HI entering into the population that is etering at an
age older than the birth age (replacing deaths from mortality proportional to the demographic
distribution and approximating in-migration) is HIV positive}

\item{Ecohab.window}{The number of time steps individuals in a cohab relationship will have the nodal attribute Ecohab=1.}

\item{p.growth.nsteps}{The number of time steps for which there are 'p.growth.size' additional entries into the population if p.growth = TRUE.}

\item{add.demog.groups}{if TRUE will include the calcualtion of additional attributes that are conmbinations of attributes as well as
attributes that change based on concurrency status. These were used as attributes during the estimation process but no included in our final model. additioanl attribute combinations as an options}

\item{test.prob.nevtest.S}{The age group specific vector of probabilities of testing if the node has never tested by sex=S}

\item{test.prob.tested.S}{The age group specific vector of probabilities of testing if the node has previously tested by sex=S}
}
\value{
A list object of class \code{param_KTM}, which can be passed to
EpiModel function \code{netsim}.
}
\description{
Sets the epidemic parameters for stochastic network models
             simulated with \code{\link{netsim}} for EpiModelHIV
}
\keyword{shamp}
