% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_msm}
\alias{param_msm}
\title{Epidemic Model Parameters}
\usage{
param_msm(
  netstats,
  hiv.test.rate = c(0.01325, 0.0125, 0.0124),
  hiv.test.late.prob = c(0.25, 0.25, 0.25),
  test.window.int = 21/7,
  tt.part.supp = c(0.2, 0.2, 0.2),
  tt.full.supp = c(0.4, 0.4, 0.4),
  tt.dur.supp = c(0.4, 0.4, 0.4),
  tx.init.prob = c(0.092, 0.092, 0.127),
  tx.halt.part.prob = c(0.0102, 0.0102, 0.0071),
  tx.halt.full.rr = c(0.9, 0.9, 0.9),
  tx.halt.dur.rr = c(0.5, 0.5, 0.5),
  tx.reinit.part.prob = c(0.00066, 0.00066, 0.00291),
  tx.reinit.full.rr = c(1, 1, 1),
  tx.reinit.dur.rr = c(1, 1, 1),
  max.time.off.tx.full.int = 52 * 15,
  max.time.on.tx.part.int = 52 * 10,
  max.time.off.tx.part.int = 52 * 10,
  vl.acute.rise.int = 6.4,
  vl.acute.peak = 6.886,
  vl.acute.fall.int = 6.4,
  vl.set.point = 4.5,
  vl.aids.onset.int = 520,
  vl.aids.int = 104,
  vl.aids.peak = 7,
  vl.full.supp = 1.5,
  vl.part.supp = 3.5,
  vl.tx.down.slope = 0.25,
  vl.tx.aids.down.slope = 0.25,
  vl.tx.up.slope = 0.25,
  aids.mr = 1/104,
  a.rate = 0.00052,
  arrival.age = 15,
  URAI.prob = 0.008938,
  UIAI.prob = 0.003379,
  trans.scale = c(1, 1, 1),
  acute.rr = 6,
  circ.rr = 0.4,
  cond.eff = 0.95,
  cond.fail = c(0.25, 0.25, 0.25),
  circ.prob = c(0.874, 0.874, 0.918),
  epistats,
  acts.aids.vl = 5.75,
  acts.scale = 1,
  cond.scale = 1,
  rgc.tprob = 0.35,
  ugc.tprob = 0.25,
  rct.tprob = 0.2,
  uct.tprob = 0.16,
  rgc.sympt.prob = 0.16,
  ugc.sympt.prob = 0.8,
  rct.sympt.prob = 0.14,
  uct.sympt.prob = 0.58,
  rgc.ntx.int = 16.8,
  ugc.ntx.int = 16.8,
  gc.tx.int = 1.4,
  rct.ntx.int = 32,
  uct.ntx.int = 32,
  ct.tx.int = 1.4,
  gc.sympt.prob.tx = c(0.95, 0.95, 0.95),
  ct.sympt.prob.tx = c(0.9, 0.9, 0.9),
  gc.asympt.prob.tx = c(0.15, 0.15, 0.15),
  ct.asympt.prob.tx = c(0.15, 0.15, 0.15),
  sti.cond.eff = 0.9,
  sti.cond.fail = c(0.2, 0.2, 0.2),
  hiv.rgc.rr = 2.78,
  hiv.ugc.rr = 1.73,
  hiv.rct.rr = 2.78,
  hiv.uct.rr = 1.73,
  hiv.dual.rr = 0.2,
  riskh.start = Inf,
  prep.start = Inf,
  prep.start.prob = 0.2,
  prep.adhr.dist = c(0.089, 0.127, 0.784),
  prep.adhr.hr = c(0.69, 0.19, 0.01),
  prep.discont.rate = 1 - (2^(-1/(224.4237/7))),
  prep.tst.int = 90/7,
  prep.risk.int = 182/7,
  prep.sti.screen.int = 182/7,
  prep.sti.prob.tx = 1,
  prep.risk.reassess.method = "year",
  prep.require.lnt = TRUE,
  ...
)
}
\arguments{
\item{netstats}{Target statistics and related network initialization data from
the standard ARTnet workflow.}

\item{hiv.test.rate}{Mean probability of HIV testing per week for
black/hispanic/white MSM (vector of length 3).}

\item{hiv.test.late.prob}{Proportion of black/hispanic/white MSM who test only
during AIDS stage infection (vector of length 3).}

\item{test.window.int}{Length of the HIV test window period in weeks.}

\item{tt.part.supp}{Proportion of black/hispanic/white MSM who enter partial viral
suppression category after ART initiation (vector of length 3).}

\item{tt.full.supp}{Proportion of black/hispanic/white MSM who enter full viral
suppression category after ART initiation (vector of length 3).}

\item{tt.dur.supp}{Proportion of black/hispanic/white MSM who enter durable viral
suppression category after ART initiation (vector of length 3).}

\item{tx.init.prob}{Probability per time step that a black/hispanic/white MSM who has
tested positive will initiate treatment (vector of length 3).}

\item{tx.halt.part.prob}{Probability per time step that black/hispanic/white
MSM who have started treatment and assigned to the partial VL suppression
category will stop treatment (vector of length 3).}

\item{tx.halt.full.rr}{Relative reduction in \code{tx.halt.part.prob} for
black/hispanic/white MSM in the full VL suppression category (vector of length 3).}

\item{tx.halt.dur.rr}{Relative reduction in \code{tx.halt.part.prob} for
black/hispanic/white MSM in the durable VL suppression category (vector of length 3).}

\item{tx.reinit.part.prob}{Probability per time step that a black/hispanic/white
MSM who has stopped treatment and assigned to the partial VL suppression
category will restart treatment (vector of length 3).}

\item{tx.reinit.full.rr}{Relative reduction in \code{tx.reinit.part.prob} for
black/hispanic/white MSM in the full VL suppression category (vector of length 3).}

\item{tx.reinit.dur.rr}{Relative reduction in \code{tx.reinit.part.prob} for
black/hispanic/white MSM in the durable VL suppression category (vector of length 3).}

\item{max.time.off.tx.full.int}{Number of weeks off treatment for a full
suppressor before onset of AIDS, including time before diagnosis.}

\item{max.time.on.tx.part.int}{Number of weeks on treatment for a
partial suppressor beofre onset of AIDS.}

\item{max.time.off.tx.part.int}{Nnumber of weeks off treatment for a
partial suppressor before onset of AIDS, including time before
diagnosis.}

\item{vl.acute.rise.int}{Number of weeks to peak viremia during acute
infection.}

\item{vl.acute.peak}{Peak viral load (in log10 units) at the height of acute
infection.}

\item{vl.acute.fall.int}{Number of weeks from peak viremia to set-point
viral load during the acute infection period.}

\item{vl.set.point}{Set point viral load (in log10 units).}

\item{vl.aids.onset.int}{Number of weeks to AIDS for a treatment-naive
patient.}

\item{vl.aids.int}{Duration of AIDS stage infection in weeks.}

\item{vl.aids.peak}{Maximum viral load during AIDS stage.}

\item{vl.full.supp}{Log10 viral load at full suppression on ART.}

\item{vl.part.supp}{Log10 viral load at partial suppression on ART.}

\item{vl.tx.down.slope}{Number of log10 units that viral load falls per time
step from treatment initiation or re-initiation until the suppression
level is reached (pre-AIDS stages).}

\item{vl.tx.aids.down.slope}{Number of log10 units that viral load falls per time
step from treatment initiation or re-initiation until the suppression
level is reached (AIDS stage).}

\item{vl.tx.up.slope}{Number of log10 units that viral load rises per time
step from treatment halting until expected value.}

\item{aids.mr}{Mortality rate of persons in the AIDS stage who are currently
off ART.}

\item{a.rate}{Rate at which MSM enter the population.}

\item{arrival.age}{Age (in years) of new arrivals.}

\item{URAI.prob}{Probability of transmission for a man having unprotected
receptive anal intercourse with an infected man at set point viral
load.}

\item{UIAI.prob}{Probability of transmission for an uncircumcised man having
unprotected insertive anal intercourse with an infected man at set
point viral load.}

\item{trans.scale}{Relative scalar on base infection probabilities for model
calibration for black/hispanic/white men (vector of length 3).}

\item{acute.rr}{Relative risk of infection (compared to that predicted by
elevated viral load) when positive partner is in the acute stage.}

\item{circ.rr}{Relative risk of infection from insertive anal sex when the
negative insertive partner is circumcised.}

\item{cond.eff}{Relative risk of HIV infection from anal sex when a condom is
used properly (biological efficacy).}

\item{cond.fail}{Condom failure rates for HIV for black/hispanic/white MSM, as a reduction
in the cond.eff parameter (vector of length 3).}

\item{circ.prob}{Probablity that a black/hispanic/white new arrival in the population
will be circumcised (vector of length 3).}

\item{epistats}{GLMs for epidemiological parameter from the standard ARTnet workflow.}

\item{acts.aids.vl}{Viral load level after which sexual act rate goes to zero.}

\item{acts.scale}{Scalar for main/casual act rate for model calibration.}

\item{cond.scale}{Scalar for condom use probability for model calibration.}

\item{rgc.tprob}{Probability of rectal gonorrhea infection per act.}

\item{ugc.tprob}{Probability of urethral gonorrhea infection per act.}

\item{rct.tprob}{Probability of rectal chlamydia infection per act.}

\item{uct.tprob}{Probability of urethral chlamydia infection per act.}

\item{rgc.sympt.prob}{Probability of symptoms given infection with rectal
gonorrhea.}

\item{ugc.sympt.prob}{Probability of symptoms given infection with urethral
gonorrhea.}

\item{rct.sympt.prob}{Probability of symptoms given infection with rectal
chlamydia.}

\item{uct.sympt.prob}{Probability of symptoms given infection with urethral
chlamydia.}

\item{rgc.ntx.int}{Average duration in weeks of untreated rectal gonorrhea.}

\item{ugc.ntx.int}{Average duration in weeks of untreated urethral gonorrhea.}

\item{gc.tx.int}{Average duration in weeks of treated gonorrhea (both sites).}

\item{rct.ntx.int}{Average in weeks duration of untreated rectal chlamydia.}

\item{uct.ntx.int}{Average in weeks duration of untreated urethral chlamydia.}

\item{ct.tx.int}{Average in weeks duration of treated chlamydia (both sites).}

\item{gc.sympt.prob.tx}{Probability of treatment for symptomatic gonorrhea
for black/hispanic/white men (vector of length 3).}

\item{ct.sympt.prob.tx}{Probability of treatment for symptomatic chlamydia
for black/hispanic/white men (vector of length 3).}

\item{gc.asympt.prob.tx}{Probability of treatment for asymptomatic gonorrhea
for black/hispanic/white men (vector of length 3).}

\item{ct.asympt.prob.tx}{Probability of treatment for asymptomatic chlamydia
for black/hispanic/white men (vector of length 3).}

\item{sti.cond.eff}{Relative risk of STI infection from anal sex when a condom is
used properly (biological efficacy).}

\item{sti.cond.fail}{Condom failure rates for STI for black/hispanic/white MSM, as
a reduction in the cond.eff parameter (vector of length 3).}

\item{hiv.rgc.rr}{Relative risk of HIV infection given current rectal gonorrhea.}

\item{hiv.ugc.rr}{Relative risk of HIV infection given current urethral gonorrhea.}

\item{hiv.rct.rr}{Relative risk of HIV infection given current rectal chlamydia.}

\item{hiv.uct.rr}{Relative risk of HIV infection given current urethral chlamydia.}

\item{hiv.dual.rr}{Additive proportional risk, from 0 to 1, for HIV infection
given dual infection with both gonorrhea and chlamydia.}

\item{riskh.start}{Time step at which behavioral risk history assessment occurs.}

\item{prep.start}{Time step at which the PrEP intervention should start.}

\item{prep.start.prob}{Probability of starting PrEP given current indications.}

\item{prep.adhr.dist}{Proportion of men who are low, medium, and high
adherent to PrEP.}

\item{prep.adhr.hr}{The hazard ratio for infection per act associated with each
level of adherence (from Grant).}

\item{prep.discont.rate}{Rate of random discontinuation from PrEP.}

\item{prep.tst.int}{Testing interval for those who are actively on PrEP. This
overrides the mean testing interval parameters.}

\item{prep.risk.int}{Time window for assessment of risk eligibility for PrEP
in weeks.}

\item{prep.sti.screen.int}{Interval in weeks between STI screening at PrEP visits.}

\item{prep.sti.prob.tx}{Probability of treatment given positive screening during
PrEP visit.}

\item{prep.risk.reassess.method}{Interval for reassessment of risk indications
of active PrEP users, either \code{"none"} for no reassessment,
\code{"inst"} for weekly, or \code{"year"} for year.}

\item{prep.require.lnt}{If \code{TRUE}, only start on PrEP if current time step is
equal to the last negative test.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list object of class \code{param_msm}, which can be passed to
EpiModel function \code{netsim}.
}
\description{
Sets the epidemic parameters for stochastic network models
             simulated with \code{\link{netsim}} for EpiModelHIV
}
\keyword{msm}
