% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.WHAMP.R
\name{param_msm_whamp}
\alias{param_msm_whamp}
\title{Epidemic Model Parameters for the WHAMP model}
\usage{
param_msm_whamp(nwstats, change.size = FALSE, new.size = 90000,
  racedist = sumto1(c(0.0549, 0.0421, 0.4739, 0.0309, 0.0166, 0.2807,
  0.0222, 0.0021, 0.0767)), agedist = sumto1(c(0.1594, 0.1319, 0.1292,
  0.1173, 0.1183, 0.1148, 0.1071, 0.122)), nw_track = TRUE,
  test.int = 436, testing.pattern = "interval", test.window.int = 21,
  tt.traj.KC.B.prob = c(0.08 * (1 - 0.839), 0.08 * (0.839), (1 - 0.08) *
  (1 - 0.839), (1 - 0.08) * (0.839)), tt.traj.KC.H.prob = c(0.064 * (1 -
  0.902), 0.064 * (0.902), (1 - 0.064) * (1 - 0.902), (1 - 0.064) *
  (0.902)), tt.traj.KC.O.prob = c(0.104 * (1 - 0.92), 0.104 * (0.92), (1
  - 0.104) * (1 - 0.92), (1 - 0.104) * (0.92)),
  tt.traj.OW.B.prob = c(0.12 * (1 - 0.811), 0.12 * (0.811), (1 - 0.12) *
  (1 - 0.811), (1 - 0.12) * (0.811)), tt.traj.OW.H.prob = c(0.102 * (1 -
  0.885), 0.102 * (0.885), (1 - 0.102) * (1 - 0.885), (1 - 0.102) *
  (0.885)), tt.traj.OW.O.prob = c(0.148 * (1 - 0.905), 0.148 * (0.905),
  (1 - 0.148) * (1 - 0.905), (1 - 0.148) * (0.905)),
  tt.traj.EW.B.prob = c(0.161 * (1 - 0.759), 0.161 * (0.759), (1 - 0.161)
  * (1 - 0.759), (1 - 0.161) * (0.759)), tt.traj.EW.H.prob = c(0.14 * (1
  - 0.847), 0.14 * (0.847), (1 - 0.14) * (1 - 0.847), (1 - 0.14) *
  (0.847)), tt.traj.EW.O.prob = c(0.192 * (1 - 0.874), 0.192 * (0.874),
  (1 - 0.192) * (1 - 0.874), (1 - 0.192) * (0.874)),
  tx.init.int.KC.B = 46, tx.init.int.KC.H = 43,
  tx.init.int.KC.O = 47, tx.init.int.OW.B = 56,
  tx.init.int.OW.H = 53, tx.init.int.OW.O = 57,
  tx.init.int.EW.B = 53, tx.init.int.EW.H = 50,
  tx.init.int.EW.O = 53, tx.halt.full = 0.002226782,
  tx.halt.part.rr = 2, tx.reinit.full.KC.B = 0.0244,
  tx.reinit.full.KC.H = 0.0239, tx.reinit.full.KC.O = 0.0275,
  tx.reinit.full.OW.B = 0.02, tx.reinit.full.OW.H = 0.0195,
  tx.reinit.full.OW.O = 0.0223, tx.reinit.full.EW.B = 0.0219,
  tx.reinit.full.EW.H = 0.0209, tx.reinit.full.EW.O = 0.0237,
  tx.reinit.part.rr = 0.5, max.time.off.tx.int = 520 * 7,
  sympt.onset.int = 2737.5, vl.acute.rise.int = 45,
  vl.acute.peak = 6.886, vl.acute.fall.int = 45, vl.set.point = 4.5,
  vl.aids.onset.int = 520 * 7, vl.aids.int = 52 * 2 * 7,
  vl.aids.peak = 7, vl.full.supp = 1.5, vl.part.supp = 3.5,
  full.supp.down.slope = 0.75, full.supp.up.slope = 0.75,
  part.supp.down.slope = 0.25, part.supp.up.slope = 0.75,
  growth.rate = 1.0001441/7, birth.age = 18, exit.age = 60,
  asmr.rr.pos = c(3.791, 2.974, 1.984), URAI.prob = 0.0082 * 1.09,
  UIAI.prob = 0.0031 * 1.09, acute.rr = 6, circ.rr = 0.4,
  condom.rr = 0.295, tprob.scalar = 1, disc.outset.main.prob = 1,
  disc.at.diag.main.prob = 1, disc.post.diag.main.prob = 0,
  disc.outset.pers.prob = 0.5671, disc.at.diag.pers.prob = 1,
  disc.post.diag.pers.prob = 0, disc.inst.prob = 0.4918,
  circ.B.prob = 0.6449, circ.H.prob = 0.4897, circ.O.prob = 0.86,
  ccr5.B.prob = c(0, 0.034), ccr5.H.prob = c(0.003, 0.05),
  ccr5.O.prob = c(0.017, 0.164), ccr5.heteroz.rr = 0.3,
  base.ai.main.rate = 0.1864, base.ai.pers.rate = 0.118,
  ai.scale = 1, cond.main.YY.prob = 0.3391,
  cond.main.other.prob = 0.1399, cond.pers.always.prob = 0.0924,
  cond.pers.prob = 0.4001, cond.inst.always.prob = 0.1982,
  cond.inst.prob = 0.4138, cond.always.prob.corr = 0.6009,
  cond.rr = 1, cond.diag.main.beta = -0.67,
  cond.discl.main.beta = -0.85, cond.diag.pers.beta = -0.67,
  cond.discl.pers.beta = -0.85, cond.diag.inst.beta = -0.67,
  cond.discl.inst.beta = -0.85, vv.iev.prob = 0.42, prep.start = Inf,
  prep.adh = "low", prep.class.prob.low = c(0.05, 0.3, 0.65),
  prep.class.prob.high = c(0.03, 0.07, 0.9), prep.class.hr = c(0.69,
  0.19, 0.04), prep.uptake.scen = "stable",
  prep.coverage.init = 0.3331, prep.coverage.init.low = 0.25,
  prep.cov.max = 0.6618, prep.cov.method = "curr",
  prep.init.rate = 1, prep.tst.int = 90, prep.risk.int = 365,
  prep.risk.reassess.method = "inst", prep.discont = 0.3,
  prep.discont.prob = 0.017724, rcomp.prob = 0,
  rcomp.adh.groups = "all", rcomp.main.only = FALSE,
  rcomp.discl.only = FALSE, rcomp.discont = FALSE,
  rgc.tprob = 0.357698, ugc.tprob = 0.248095, rct.tprob = 0.321597,
  uct.tprob = 0.212965, rgc.sympt.prob = 0.076975,
  ugc.sympt.prob = 0.824368, rct.sympt.prob = 0.103517,
  uct.sympt.prob = 0.885045, rgc.asympt.int = 35.11851 * 7,
  ugc.asympt.int = 35.11851 * 7, gc.tx.int = 2 * 7, gc.ntx.int = NA,
  rct.asympt.int = 44.24538 * 7, uct.asympt.int = 44.24538 * 7,
  ct.tx.int = 2 * 7, ct.ntx.int = NA, gc.prob.cease = 0,
  ct.prob.cease = 0, gc.sympt.prob.tx = 0.9, ct.sympt.prob.tx = 0.85,
  gc.asympt.prob.tx = 0, ct.asympt.prob.tx = 0,
  prep.sti.screen.int = 182, prep.sti.prob.tx = 1,
  prep.continue.stand.tx = TRUE, sti.cond.rr = 0.3,
  hiv.rgc.rr = 2.780673, hiv.ugc.rr = 1.732363,
  hiv.rct.rr = 2.780673, hiv.uct.rr = 1.732363, hiv.dual.rr = 0.2,
  ...)
}
\arguments{
\item{nwstats}{Target statistics for the network model. An object of class
\code{nwstats} output from \code{\link{calc_nwstats_msm}}.}

\item{change.size}{If \code{TRUE}, change the size of the network for epidemic simulations}

\item{new.size}{Size to which to scale the population for epidemic simulations}

\item{racedist}{Population distribution by race/ethnicity and region (King County Hispanic,
black, other; other western WA Hispanic, black, and other, eastern WA Hispanic,
black, and other)}

\item{agedist}{Target population age structure (proportion in each group 18-24, 25-29... 55-59)}

\item{nw_track}{Sets whether to run the script in the nwfeatures_msm_whamp function to
track features of the sexual network (degree distribution, mixing, partnership age)
over time, with options \code{TRUE} or \code{FALSE}}

\item{test.int}{Average interest interval in days for men who test regularly}

\item{testing.pattern}{Method for HIV testing, with options \code{"memoryless"}
for constant hazard without regard to time since previous test, or
\code{"interval"} deterministic fixed intervals.}

\item{test.window.int}{Length of the HIV test window period in days.}

\item{tt.traj.KC.B.prob}{Proportion of black MSM in King County who enter one of four
testing/treatment trajectories: non-screener treated with partial viral suppression,
non-screener treated with full viral suppression, regular screener treated with partial
suppression, and regular screener treated with full suppression}

\item{tt.traj.KC.H.prob}{Proportion of Hispanic MSM in King County who enter into the four
testing/treatment trajectories, as defined above.}

\item{tt.traj.KC.O.prob}{Proportion of other MSM in King County who enter into the four
testing/treatment trajectories, as defined above.}

\item{tt.traj.OW.B.prob}{Proportion of black MSM in Western WA who enter one of four
testing/treatment trajectories, as defined above}

\item{tt.traj.OW.H.prob}{Proportion of Hispanic MSM in Western WA who enter into the four
testing/treatment trajectories, as defined above.}

\item{tt.traj.OW.O.prob}{Proportion of other MSM in Western WA who enter into the four
testing/treatment trajectories, as defined above.}

\item{tt.traj.EW.B.prob}{Proportion of black MSM in Eastern WA who enter one of four
testing/treatment trajectories, as defined above}

\item{tt.traj.EW.H.prob}{Proportion of Hispanic MSM in Eastern WA who enter into the four
testing/treatment trajectories, as defined above.}

\item{tt.traj.EW.O.prob}{Proportion of other MSM in Eastern WA who enter into the four
testing/treatment trajectories, as defined above.}

\item{tx.init.int.KC.B}{Days from diagnosis to treatment initiation for Black MSM in King County}

\item{tx.init.int.KC.H}{Days from diagnosis to treatment initiation for Hispanic MSM in King County}

\item{tx.init.int.KC.O}{Days from diagnosis to treatment initiation for Other MSM in King County}

\item{tx.init.int.OW.B}{Days from diagnosis to treatment initiation for Black MSM in other western WA}

\item{tx.init.int.OW.H}{Days from diagnosis to treatment initiation for Hispanic MSM in other western WA}

\item{tx.init.int.OW.O}{Days from diagnosis to treatment initiation for Other MSM in other western WA}

\item{tx.init.int.EW.B}{Days from diagnosis to treatment initiation for Black MSM in eastern WA}

\item{tx.init.int.EW.H}{Days from diagnosis to treatment initiation for Hispanic MSM in eastern WA}

\item{tx.init.int.EW.O}{Days from diagnosis to treatment initiation for Other MSM in eastern WA}

\item{tx.halt.full}{Probability per time step that a full suppressor who is
currently on treatment will halt treatment.}

\item{tx.halt.part.rr}{Relative risk of halting treatment per time step for a partial suppressor
currently on treatment, as compared to a full suppressor}

\item{tx.reinit.full.KC.B}{Probability per time step that black MSM in KC who are full suppressors and
not currently on treatment but who have been in the past will re-initiate}

\item{tx.reinit.full.KC.H}{Probability per time step that Hispanic MSM in KC who are full suppressors and
not currently on treatment but who have been in the past will re-initiate}

\item{tx.reinit.full.KC.O}{Probability per time step that other MSM in KC who are full suppressors and
not currently on treatment but who have been in the past will re-initiate}

\item{tx.reinit.full.OW.B}{Probability per time step that black MSM in OW who are full suppressors and
not currently on treatment but who have been in the past will re-initiate}

\item{tx.reinit.full.OW.H}{Probability per time step that Hispanic MSM in OW who are full suppressors and
not currently on treatment but who have been in the past will re-initiate}

\item{tx.reinit.full.OW.O}{Probability per time step that other MSM in OW who are full suppressors and
not currently on treatment but who have been in the past will re-initiate}

\item{tx.reinit.full.EW.B}{Probability per time step that black MSM in EW who are full suppressors and
not currently on treatment but who have been in the past will re-initiate}

\item{tx.reinit.full.EW.H}{Probability per time step that Hispanic MSM in EW who are full suppressors and
not currently on treatment but who have been in the past will re-initiate}

\item{tx.reinit.full.EW.O}{Probability per time step that other MSM in EW who are full suppressors and
not currently on treatment but who have been in the past will re-initiate}

\item{tx.reinit.part.rr}{Relative risk of reinitiating treatment per time step for a partial suppressor
currently on treatment, as compared to a full suppressor}

\item{max.time.off.tx.int}{Number of days off treatment for a before onset of AIDS, 
including time before diagnosis.}

\item{sympt.onset.int}{Number of days from infection to symptom onset in the absence of ART}

\item{vl.acute.rise.int}{Number of days to peak viremia during acute infection.}

\item{vl.acute.peak}{Peak viral load (in log10 units) at the height of acute infection.}

\item{vl.acute.fall.int}{Number of days from peak viremia to set-point
viral load during the acute infection period.}

\item{vl.set.point}{Set point viral load (in log10 units).}

\item{vl.aids.onset.int}{Number of days to AIDS for a treatment-naive
patient.}

\item{vl.aids.int}{Time from AIDS onset to peak AIDS VL in untreated infection in days.}

\item{vl.aids.peak}{Peak VL in AIDS.}

\item{vl.full.supp}{Log10 viral load at full suppression on ART.}

\item{vl.part.supp}{Log10 viral load at partial suppression on ART.}

\item{full.supp.down.slope}{For full suppressors, number of log10 units that
viral load falls per time step from treatment initiation or re-initiation
until the level in \code{vl.full.supp}.}

\item{full.supp.up.slope}{For full suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected
value.}

\item{part.supp.down.slope}{For partial suppressors, number of log10 units
that viral load falls per time step from treatment initiation or
re-initiation until the level in \code{vl.part.supp}.}

\item{part.supp.up.slope}{For partial suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected value.}

\item{growth.rate}{Rate of population growth per day.}

\item{birth.age}{Age (in years) of new arrivals.}

\item{exit.age}{Age (in years) at which individuals age out of the network}

\item{asmr.rr.pos}{Relative increase in age-specific mortality for HIV-positive men
(for age groups 18-44, 45-54, 55-59)}

\item{URAI.prob}{Probability of transmission for a man having unprotected
receptive anal intercourse with an infected man at set point viral
load.}

\item{UIAI.prob}{Probability of transmission for an uncircumcised man having
unprotected insertive anal intercourse with an infected man at set
point viral load.}

\item{acute.rr}{Relative risk of infection (compared to that predicted by
elevated viral load) when positive partner is in the acute stage.}

\item{circ.rr}{Relative risk of infection from insertive anal sex when the
negative insertive partner is circumcised.}

\item{condom.rr}{Relative risk of infection from anal sex when a condom is
used.}

\item{tprob.scalar}{General relative scaler for all per-act probability of 
transmission for model calibration.}

\item{disc.outset.main.prob}{Probability that an HIV-infected MSM will
disclose his status at the start of a main partnership.}

\item{disc.at.diag.main.prob}{Probability that an MSM already in a main
partnership will disclose at the time of diagnosis.}

\item{disc.post.diag.main.prob}{Probability that an HIV-infected MSM
in a main partnership will disclose his status, assuming he didn't
at the start of the partnership or at diagnosis.}

\item{disc.outset.pers.prob}{Probability that an HIV-infected MSM will
disclose his status at the start of a casual partnership.}

\item{disc.at.diag.pers.prob}{Probability that an MSM already in a
casual partnership will disclose at the time of diagnosis.}

\item{disc.post.diag.pers.prob}{Probability that an HIV-infected MSM
in a casual partnership will disclose his status, assuming he
didn't at the start of the partnership or at diagnosis.}

\item{disc.inst.prob}{Probability that an HIV-infected black MSM will
disclose his status to a one-off partner.}

\item{circ.B.prob}{Probablity that a black new arrival in the population
will be circumcised.}

\item{circ.H.prob}{Probablity that a Hispanic new arrival in the population
will be circumcised.}

\item{circ.O.prob}{Probablity that an other race/ethnicity new arrival in 
the population will be circumcised.}

\item{ccr5.B.prob}{Vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respectively) in the CCR5 gene
among black MSM.}

\item{ccr5.H.prob}{Vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respectively) in the CCR5 gene
among hispanic MSM.}

\item{ccr5.O.prob}{Vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respectively) in the CCR5 gene
among other race/ethnicity MSM.}

\item{ccr5.heteroz.rr}{Relative risk of infection for men who are heterozygous
in the CCR5 mutation.}

\item{base.ai.main.rate}{Expected coital frequency in main partnerships
(acts per day).}

\item{base.ai.pers.rate}{Expected coital frequency in persistent partnerships
(acts per day).}

\item{ai.scale}{General relative scaler for all act rates for model
calibration.}

\item{cond.main.YY.prob}{Probability of condom use in a Young-Young (18-34) main
partnership.}

\item{cond.main.other.prob}{Probability of condom use in main partnerships with
other age combinations.}

\item{cond.pers.always.prob}{Fraction of men in persistent partnerships who always
use condoms in those partnerships.}

\item{cond.pers.prob}{Of men who are not consistent condom users, per-act
probability of condom use in persistent partnerships.}

\item{cond.inst.always.prob}{Fraction of men in instant partnerships who always
use condoms in those partnerships.}

\item{cond.inst.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a instantaneous partnerships.}

\item{cond.always.prob.corr}{Correlation coefficient for probability of always
using condoms in both persistent and instantaneous partnerships}

\item{cond.rr}{Condom probability scaler for model calibration purposes.}

\item{cond.diag.main.beta}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.main.beta}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected man has disclosed.}

\item{cond.diag.pers.beta}{Beta multiplier for the log odds of using a
condom in a persistent partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.pers.beta}{Beta multiplier for the log odds of using a
condom in a persistent partnership if the HIV-infected man has disclosed
his status.}

\item{cond.diag.inst.beta}{Beta multiplier for the log odds of using a
condom in an instantaneous partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.inst.beta}{Beta multiplier for the log odds of using a
condom in an instantaneous partnership if the HIV-infected man has disclosed
his status.}

\item{vv.iev.prob}{Probability that in a partnership of two versatile men,
they will engage in intra-event versatility ("flipping") given that 
they're having AI.}

\item{prep.start}{Time step at which the PrEP intervention should start.}

\item{prep.adh}{PrEP adherence scenario, with options \code{"low"} and \code{"high"}}

\item{prep.class.prob.low}{The proportion of men in low, medium, and high adherence
groups if \code{prep.adh} == "low".}

\item{prep.class.prob.high}{The proportion of men in low, medium, and high adherence
groups if \code{prep.adh} == "high.}

\item{prep.class.hr}{The hazard ratio for infection per act associated with each
level of adherence.}

\item{prep.uptake.scen}{PrEP uptake scenarios, with options \code{"stable"} for
stable coverage at levels defined by \code{prep.coverage.init}, 
\code{"fifty"} for increased uptake to 50 percent overall by 2020 (from 2017), 
\code{"max"} for increased uptake to levels defined by \code{"prep.cov.max"}, 
after hitting 50 percent overall in 2020, and \code{"lower"} for stable coverage
at levels defined by \code{prep.coverage.init.low}.}

\item{prep.coverage.init}{The proportion of eligible men (eligibility defined by CAI and 
discordant partnership status in line with WA State guidelines) who are assigned PrEP 
in the time step in which PrEP is initiated (as specified in \code{prep.start}).}

\item{prep.coverage.init.low}{Low-range estimate for the proportion of eligible men 
(eligibility defined by CAI and discordant partnership status in line with WA 
State guidelines) who are assigned PrEP in the time step in which PrEP is initiated 
(as specified in \code{prep.start}).}

\item{prep.cov.max}{Maximum attainable PrEP coverage if \code{prep.uptake.scen} is set to "max".
Upon reaching these values, PrEP uptake will stabilize.}

\item{prep.cov.method}{The method for calculating PrEP coverage, with options
of \code{"curr"} to base the numerator on the number of people currently
on PrEP and \code{"ever"} to base it on the number of people ever on
PrEP.}

\item{prep.init.rate}{The rate at which persons initiate PrEP conditional on
their eligibility, with 1 equal to instant start.}

\item{prep.tst.int}{Testing interval for those who are actively on PrEP. This
overrides the mean testing interval parameters.}

\item{prep.risk.int}{Time window for assessment of risk eligibility for PrEP
in days.}

\item{prep.risk.reassess.method}{Method for determining risk-based discontinuation
of PrEP, with \code{"none"} for no risk-based discontinuation, \code{"inst"} 
for reassessment every time step, and \code{"year"} for reassessment at yearly
HIV screening visits.}

\item{prep.discont}{<- Proportion of PrEP users who will discontinue while still at risk.
This will be used to assign an attribute to indicate men who will discontinue
while they are still candidates for PrEP. The remainder will discontinue only
if their risk changes such that they are no longer eligilbe for the intervention.}

\item{prep.discont.prob}{Probability per time step of halting PrEP for men who 
discontinue.}

\item{rcomp.prob}{Level of risk compensation from 0 to 1, where 0 is no risk
compensation, 0.5 is a 50% reduction in the probability of condom use
per act, and 1 is a complete cessation of condom use following PrEP
initiation.}

\item{rcomp.adh.groups}{PrEP adherence groups for whom risk compensation
occurs, with options \code{"all"} for everyone, \code{"top2"} for risk compensation
only in those with medium and high adherence, and \code{"top1"} for risk compensation
only in those with high adherence to PrEP.}

\item{rcomp.main.only}{Logical, if risk compensation is limited to main
partnerships only, versus all partnerships.}

\item{rcomp.discl.only}{Logical, if risk compensation is limited known-discordant
partnerships only, versus all partnerships.}

\item{rcomp.discont}{Logical, if condom use stays at the level at which PrEP was 
discontinued versus returning to pre-PrEP levels}

\item{rgc.tprob}{Probability of rectal gonorrhea infection per act.}

\item{ugc.tprob}{Probability of urethral gonorrhea infection per act.}

\item{rct.tprob}{Probability of rectal chlamydia infection per act.}

\item{uct.tprob}{Probability of urethral chlamydia infection per act.}

\item{rgc.sympt.prob}{Probability of symptoms given infection with rectal
gonorrhea.}

\item{ugc.sympt.prob}{Probability of symptoms given infection with urethral
gonorrhea.}

\item{rct.sympt.prob}{Probability of symptoms given infection with rectal
chlamydia.}

\item{uct.sympt.prob}{Probability of symptoms given infection with urethral
chlamydia.}

\item{rgc.asympt.int}{Average duration in days of asymptomatic rectal gonorrhea.}

\item{ugc.asympt.int}{Average duration in days of asymptomatic urethral gonorrhea.}

\item{gc.tx.int}{Average duration in days of treated gonorrhea (both sites).}

\item{gc.ntx.int}{Average duration in days of untreated, symptomatic gonorrhea (both sites).
If \code{NA}, uses site-specific durations for asymptomatic infections.}

\item{rct.asympt.int}{Average in days duration of asymptomatic rectal chlamydia.}

\item{uct.asympt.int}{Average in days duration of asymptomatic urethral chlamydia.}

\item{ct.tx.int}{Average in days duration of treated chlamydia (both sites).}

\item{ct.ntx.int}{Average in days duration of untreated, symptomatic chlamydia (both sites).
If \code{NA}, uses site-specific durations for asymptomatic infections.}

\item{gc.prob.cease}{Probability of ceasing sexual activity during symptomatic
infection with gonorrhea.}

\item{ct.prob.cease}{Probability of ceasing sexual activity during symptomatic
infection with chlamydia.}

\item{gc.sympt.prob.tx}{Probability of treatment for symptomatic gonorrhea.}

\item{ct.sympt.prob.tx}{Probability of treatment for symptomatic chlamydia.}

\item{gc.asympt.prob.tx}{Probability of treatment for asymptomatic gonorrhea.}

\item{ct.asympt.prob.tx}{Probability of treatment for asymptomatic chlamydia.}

\item{prep.sti.screen.int}{Interval in days between STI screening at PrEP visits.}

\item{prep.sti.prob.tx}{Probability of treatment given positive screening during
PrEP visit.}

\item{prep.continue.stand.tx}{Logical, if \code{TRUE} will continue standard
STI treatment of symptomatic cases even after PrEP initiation.}

\item{sti.cond.rr}{Relative risk of STI infection (in either direction) given
a condom used by the insertive partner.}

\item{hiv.rgc.rr}{Relative risk of HIV infection given current rectal gonorrhea.}

\item{hiv.ugc.rr}{Relative risk of HIV infection given current urethral gonorrhea.}

\item{hiv.rct.rr}{Relative risk of HIV infection given current rectal chlamydia.}

\item{hiv.uct.rr}{Relative risk of HIV infection given current urethral chlamydia.}

\item{hiv.dual.rr}{Additive proportional risk, from 0 to 1, for HIV infection
given dual infection with both gonorrhea and chlamydia.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list object of class \code{param_msm}, which can be passed to
EpiModel function \code{netsim}.
}
\description{
Sets the epidemic parameters for stochastic network models
             simulated with \code{\link{netsim}} for EpiModelHIV
}
\keyword{msm}
