% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.shamp.R
\name{param_shamp}
\alias{param_shamp}
\title{Epidemic Model Parameters for SHAMP}
\usage{
param_shamp(race.method = 1, time.unit = 7, method = 1,
  last.neg.test.B.f.int = 1464, last.neg.test.BI.f.int = 1464,
  last.neg.test.H.f.int = 1464, last.neg.test.HI.f.int = 1464,
  last.neg.test.W.f.int = 1464, last.neg.test.B.msf.int = 1708,
  last.neg.test.BI.msf.int = 1708, last.neg.test.H.msf.int = 1708,
  last.neg.test.HI.msf.int = 1708, last.neg.test.W.msf.int = 1708,
  last.neg.test.B.msm.int = 1037, last.neg.test.BI.msm.int = 1037,
  last.neg.test.H.msm.int = 1037, last.neg.test.HI.msm.int = 1037,
  last.neg.test.W.msm.int = 1037, last.neg.test.B.msmf.int = 1129,
  last.neg.test.BI.msmf.int = 1129, last.neg.test.H.msmf.int = 1129,
  last.neg.test.HI.msmf.int = 1129, last.neg.test.W.msmf.int = 1129,
  mean.test.B.f.int = 1464, mean.test.BI.f.int = 1464,
  mean.test.H.f.int = 1464, mean.test.HI.f.int = 1464,
  mean.test.W.f.int = 1464, mean.test.B.msf.int = 1708,
  mean.test.BI.msf.int = 1708, mean.test.H.msf.int = 1708,
  mean.test.HI.msf.int = 1708, mean.test.W.msf.int = 1708,
  mean.test.B.msm.int = 1037, mean.test.BI.msm.int = 1037,
  mean.test.H.msm.int = 1037, mean.test.HI.msm.int = 1037,
  mean.test.W.msm.int = 1037, mean.test.B.msmf.int = 1129,
  mean.test.BI.msmf.int = 1129, mean.test.H.msmf.int = 1129,
  mean.test.HI.msmf.int = 1129, mean.test.W.msmf.int = 1129,
  testing.pattern = "memoryless", test.window.int = 15,
  tt.traj.B.f.prob = c(0.741, 0, 0.053, 0.206), tt.traj.BI.f.prob = c(0.741,
  0, 0.053, 0.206), tt.traj.H.f.prob = c(0.741, 0, 0.041, 0.218),
  tt.traj.HI.f.prob = c(0.741, 0, 0.041, 0.218), tt.traj.W.f.prob = c(0.741,
  0, 0.057, 0.202), tt.traj.B.msf.prob = c(0.741, 0, 0.066, 0.193),
  tt.traj.BI.msf.prob = c(0.741, 0, 0.066, 0.193),
  tt.traj.H.msf.prob = c(0.741, 0, 0.067, 0.192),
  tt.traj.HI.msf.prob = c(0.741, 0, 0.067, 0.192),
  tt.traj.W.msf.prob = c(0.741, 0, 0.052, 0.207),
  tt.traj.B.msm.prob = c(0.025, 0, 0.065, 0.911),
  tt.traj.BI.msm.prob = c(0.025, 0, 0.065, 0.911),
  tt.traj.H.msm.prob = c(0.025, 0, 0.065, 0.911),
  tt.traj.HI.msm.prob = c(0.025, 0, 0.065, 0.911),
  tt.traj.W.msm.prob = c(0.025, 0, 0.065, 0.911),
  tt.traj.B.msmf.prob = c(0.741, 0, 0.066, 0.193),
  tt.traj.BI.msmf.prob = c(0.741, 0, 0.066, 0.193),
  tt.traj.H.msmf.prob = c(0.741, 0, 0.041, 0.218),
  tt.traj.HI.msmf.prob = c(0.741, 0, 0.041, 0.218),
  tt.traj.W.msmf.prob = c(0.741, 0, 0.052, 0.207),
  tx.init.B.f.prob = 0.2331, tx.init.BI.f.prob = 0.2331,
  tx.init.H.f.prob = 0.2331, tx.init.HI.f.prob = 0.2331,
  tx.init.W.f.prob = 0.2331, tx.init.B.msf.prob = 0.2331,
  tx.init.BI.msf.prob = 0.2331, tx.init.H.msf.prob = 0.2331,
  tx.init.HI.msf.prob = 0.2331, tx.init.W.msf.prob = 0.2331,
  tx.init.B.msm.prob = 0.1467, tx.init.BI.msm.prob = 0.1467,
  tx.init.H.msm.prob = 0.1467, tx.init.HI.msm.prob = 0.1467,
  tx.init.W.msm.prob = 0.1467, tx.init.B.msmf.prob = 0.2331,
  tx.init.BI.msmf.prob = 0.2331, tx.init.H.msmf.prob = 0.2331,
  tx.init.HI.msmf.prob = 0.2331, tx.init.W.msmf.prob = 0.2331,
  tx.halt.B.f.prob = 1e-04, tx.halt.BI.f.prob = 1e-04,
  tx.halt.H.f.prob = 1e-04, tx.halt.HI.f.prob = 1e-04,
  tx.halt.W.f.prob = 1e-04, tx.halt.B.msf.prob = 1e-04,
  tx.halt.BI.msf.prob = 1e-04, tx.halt.H.msf.prob = 1e-04,
  tx.halt.HI.msf.prob = 1e-04, tx.halt.W.msf.prob = 1e-04,
  tx.halt.B.msm.prob = 1e-04, tx.halt.BI.msm.prob = 1e-04,
  tx.halt.H.msm.prob = 1e-04, tx.halt.HI.msm.prob = 1e-04,
  tx.halt.W.msm.prob = 1e-04, tx.halt.B.msmf.prob = 1e-04,
  tx.halt.BI.msmf.prob = 1e-04, tx.halt.H.msmf.prob = 1e-04,
  tx.halt.HI.msmf.prob = 1e-04, tx.halt.W.msmf.prob = 1e-04,
  tx.reinit.B.f.prob = 1, tx.reinit.BI.f.prob = 1, tx.reinit.H.f.prob = 1,
  tx.reinit.HI.f.prob = 1, tx.reinit.W.f.prob = 1,
  tx.reinit.B.msf.prob = 1, tx.reinit.BI.msf.prob = 1,
  tx.reinit.H.msf.prob = 1, tx.reinit.HI.msf.prob = 1,
  tx.reinit.W.msf.prob = 1, tx.reinit.B.msm.prob = 1,
  tx.reinit.BI.msm.prob = 1, tx.reinit.H.msm.prob = 1,
  tx.reinit.HI.msm.prob = 1, tx.reinit.W.msm.prob = 1,
  tx.reinit.B.msmf.prob = 1, tx.reinit.BI.msmf.prob = 1,
  tx.reinit.H.msmf.prob = 1, tx.reinit.HI.msmf.prob = 1,
  tx.reinit.W.msmf.prob = 1, max.time.off.tx.full.int = 520 * 7,
  max.time.on.tx.part.int = 52 * 15 * 7, max.time.off.tx.part.int = 520 * 7,
  vl.acute.rise.int = 45, vl.acute.peak = 6.886, vl.acute.fall.int = 21,
  vl.set.point = 4.5, vl.aids.onset.int = 520 * 7, vl.aids.int = 52 * 2 *
  7, vl.fatal = 7, vl.full.supp = 1.5, vl.part.supp = 3.5,
  full.supp.down.slope = 0.25, full.supp.up.slope = 0.25,
  part.supp.down.slope = 0.25, part.supp.up.slope = 0.25, birth.age = 18,
  exit.age = 60, b.method = "fixed", msm.frac = 0, msmf.frac = 0.012,
  URAI.prob = 138/10000, UIAI.prob = 11/10000, URVI.prob = 8/10000,
  UIVI.prob = 4/10000, acute.rr = 4, circ.rr = 0.4, condom.rr = 0.25,
  disc.outset.main.B.f.prob = 0, disc.outset.main.BI.f.prob = 0,
  disc.outset.main.H.f.prob = 0, disc.outset.main.HI.f.prob = 0,
  disc.outset.main.W.f.prob = 0, disc.outset.main.B.msf.prob = 0,
  disc.outset.main.BI.msf.prob = 0, disc.outset.main.H.msf.prob = 0,
  disc.outset.main.HI.msf.prob = 0, disc.outset.main.W.msf.prob = 0,
  disc.outset.main.B.msm.prob = 0, disc.outset.main.BI.msm.prob = 0,
  disc.outset.main.H.msm.prob = 0, disc.outset.main.HI.msm.prob = 0,
  disc.outset.main.W.msm.prob = 0, disc.outset.main.B.msmf.prob = 0,
  disc.outset.main.BI.msmf.prob = 0, disc.outset.main.H.msmf.prob = 0,
  disc.outset.main.HI.msmf.prob = 0, disc.outset.main.W.msmf.prob = 0,
  disc.at.diag.main.B.f.prob = 0, disc.at.diag.main.BI.f.prob = 0,
  disc.at.diag.main.H.f.prob = 0, disc.at.diag.main.HI.f.prob = 0,
  disc.at.diag.main.W.f.prob = 0, disc.at.diag.main.B.msf.prob = 0,
  disc.at.diag.main.BI.msf.prob = 0, disc.at.diag.main.H.msf.prob = 0,
  disc.at.diag.main.HI.msf.prob = 0, disc.at.diag.main.W.msf.prob = 0,
  disc.at.diag.main.B.msm.prob = 0, disc.at.diag.main.BI.msm.prob = 0,
  disc.at.diag.main.H.msm.prob = 0, disc.at.diag.main.HI.msm.prob = 0,
  disc.at.diag.main.W.msm.prob = 0, disc.at.diag.main.B.msmf.prob = 0,
  disc.at.diag.main.BI.msmf.prob = 0, disc.at.diag.main.H.msmf.prob = 0,
  disc.at.diag.main.HI.msmf.prob = 0, disc.at.diag.main.W.msmf.prob = 0,
  disc.post.diag.main.B.f.prob = 0, disc.post.diag.main.BI.f.prob = 0,
  disc.post.diag.main.H.f.prob = 0, disc.post.diag.main.HI.f.prob = 0,
  disc.post.diag.main.W.f.prob = 0, disc.post.diag.main.B.msf.prob = 0,
  disc.post.diag.main.BI.msf.prob = 0, disc.post.diag.main.H.msf.prob = 0,
  disc.post.diag.main.HI.msf.prob = 0, disc.post.diag.main.W.msf.prob = 0,
  disc.post.diag.main.B.msm.prob = 0, disc.post.diag.main.BI.msm.prob = 0,
  disc.post.diag.main.H.msm.prob = 0, disc.post.diag.main.HI.msm.prob = 0,
  disc.post.diag.main.W.msm.prob = 0, disc.post.diag.main.B.msmf.prob = 0,
  disc.post.diag.main.BI.msmf.prob = 0, disc.post.diag.main.H.msmf.prob = 0,
  disc.post.diag.main.HI.msmf.prob = 0, disc.post.diag.main.W.msmf.prob = 0,
  disc.outset.pers.B.f.prob = 0, disc.outset.pers.BI.f.prob = 0,
  disc.outset.pers.H.f.prob = 0, disc.outset.pers.HI.f.prob = 0,
  disc.outset.pers.W.f.prob = 0, disc.outset.pers.B.msf.prob = 0,
  disc.outset.pers.BI.msf.prob = 0, disc.outset.pers.H.msf.prob = 0,
  disc.outset.pers.HI.msf.prob = 0, disc.outset.pers.W.msf.prob = 0,
  disc.outset.pers.B.msm.prob = 0, disc.outset.pers.BI.msm.prob = 0,
  disc.outset.pers.H.msm.prob = 0, disc.outset.pers.HI.msm.prob = 0,
  disc.outset.pers.W.msm.prob = 0, disc.outset.pers.B.msmf.prob = 0,
  disc.outset.pers.BI.msmf.prob = 0, disc.outset.pers.H.msmf.prob = 0,
  disc.outset.pers.HI.msmf.prob = 0, disc.outset.pers.W.msmf.prob = 0,
  disc.at.diag.pers.B.f.prob = 0, disc.at.diag.pers.BI.f.prob = 0,
  disc.at.diag.pers.H.f.prob = 0, disc.at.diag.pers.HI.f.prob = 0,
  disc.at.diag.pers.W.f.prob = 0, disc.at.diag.pers.B.msf.prob = 0,
  disc.at.diag.pers.BI.msf.prob = 0, disc.at.diag.pers.H.msf.prob = 0,
  disc.at.diag.pers.HI.msf.prob = 0, disc.at.diag.pers.W.msf.prob = 0,
  disc.at.diag.pers.B.msm.prob = 0, disc.at.diag.pers.BI.msm.prob = 0,
  disc.at.diag.pers.H.msm.prob = 0, disc.at.diag.pers.HI.msm.prob = 0,
  disc.at.diag.pers.W.msm.prob = 0, disc.at.diag.pers.B.msmf.prob = 0,
  disc.at.diag.pers.BI.msmf.prob = 0, disc.at.diag.pers.H.msmf.prob = 0,
  disc.at.diag.pers.HI.msmf.prob = 0, disc.at.diag.pers.W.msmf.prob = 0,
  disc.post.diag.pers.B.f.prob = 0, disc.post.diag.pers.BI.f.prob = 0,
  disc.post.diag.pers.H.f.prob = 0, disc.post.diag.pers.HI.f.prob = 0,
  disc.post.diag.pers.W.f.prob = 0, disc.post.diag.pers.B.msf.prob = 0,
  disc.post.diag.pers.BI.msf.prob = 0, disc.post.diag.pers.H.msf.prob = 0,
  disc.post.diag.pers.HI.msf.prob = 0, disc.post.diag.pers.W.msf.prob = 0,
  disc.post.diag.pers.B.msm.prob = 0, disc.post.diag.pers.BI.msm.prob = 0,
  disc.post.diag.pers.H.msm.prob = 0, disc.post.diag.pers.HI.msm.prob = 0,
  disc.post.diag.pers.W.msm.prob = 0, disc.post.diag.pers.B.msmf.prob = 0,
  disc.post.diag.pers.BI.msmf.prob = 0, disc.post.diag.pers.H.msmf.prob = 0,
  disc.post.diag.pers.HI.msmf.prob = 0, disc.post.diag.pers.W.msmf.prob = 0,
  disc.inst.B.f.prob = 0, disc.inst.BI.f.prob = 0, disc.inst.H.f.prob = 0,
  disc.inst.HI.f.prob = 0, disc.inst.W.f.prob = 0,
  disc.inst.B.msf.prob = 0, disc.inst.BI.msf.prob = 0,
  disc.inst.H.msf.prob = 0, disc.inst.HI.msf.prob = 0,
  disc.inst.W.msf.prob = 0, disc.inst.B.msm.prob = 0,
  disc.inst.BI.msm.prob = 0, disc.inst.H.msm.prob = 0,
  disc.inst.HI.msm.prob = 0, disc.inst.W.msm.prob = 0,
  disc.inst.B.msmf.prob = 0, disc.inst.BI.msmf.prob = 0,
  disc.inst.H.msmf.prob = 0, disc.inst.HI.msmf.prob = 0,
  disc.inst.W.msmf.prob = 0, circ.B.prob = 0.77, circ.BI.prob = 0.45,
  circ.H.prob = 0.48, circ.HI.prob = 0.36, circ.W.prob = 0.91,
  ccr5.B.f.prob = c(0, 0), ccr5.BI.f.prob = c(0, 0), ccr5.H.f.prob = c(0,
  0), ccr5.HI.f.prob = c(0, 0), ccr5.W.f.prob = c(0.017, 0),
  ccr5.B.m.prob = c(0, 0), ccr5.BI.m.prob = c(0, 0), ccr5.H.m.prob = c(0,
  0), ccr5.HI.m.prob = c(0, 0), ccr5.W.m.prob = c(0.017, 0),
  ccr5.heteroz.rr = 0, num.inst.ai.classes = 1,
  base.ai.main.B.rate = 1.379/7, base.ai.main.BI.rate = 1.379/7,
  base.ai.main.H.rate = 1.379/7, base.ai.main.HI.rate = 1.379/7,
  base.ai.main.W.rate = 1.379/7, base.ai.pers.B.rate = 0.883/7,
  base.ai.pers.BI.rate = 0.883/7, base.ai.pers.H.rate = 0.883/7,
  base.ai.pers.HI.rate = 0.883/7, base.ai.pers.W.rate = 0.883/7,
  base.vi.main.B.rate = 0.25, base.vi.main.BI.rate = 0.19,
  base.vi.main.H.rate = 0.26, base.vi.main.HI.rate = 0.24,
  base.vi.main.W.rate = 0.22, base.vi.pers.B.rate = 0.12,
  base.vi.pers.BI.rate = 0.09, base.vi.pers.H.rate = 0.12,
  base.vi.pers.HI.rate = 0.09, base.vi.pers.W.rate = 0.12, ai.scale = 1,
  vi.scale = 1, cond.main.B.prob.msm = 0.062,
  cond.main.BI.prob.msm = 0.062, cond.main.H.prob.msm = 0.062,
  cond.main.HI.prob.msm = 0.062, cond.main.W.prob.msm = 0.062,
  cond.pers.always.prob.msm = 0.01, cond.pers.B.prob.msm = 0.0304,
  cond.pers.BI.prob.msm = 0.0304, cond.pers.H.prob.msm = 0.0304,
  cond.pers.HI.prob.msm = 0.0304, cond.pers.W.prob.msm = 0.0304,
  cond.inst.always.prob.msm = 0.01, cond.inst.B.prob.msm = 0.413,
  cond.inst.BI.prob.msm = 0.413, cond.inst.H.prob.msm = 0.413,
  cond.inst.HI.prob.msm = 0.413, cond.inst.W.prob.msm = 0.413,
  cond.always.prob.corr.msm = 0.01, cond.main.B.prob.het = 0.12,
  cond.main.BI.prob.het = 0.17, cond.main.H.prob.het = 0.17,
  cond.main.HI.prob.het = 0.19, cond.main.W.prob.het = 0.16,
  cond.pers.always.prob.het = 1e-04, cond.pers.B.prob.het = 0.54,
  cond.pers.BI.prob.het = 0.65, cond.pers.H.prob.het = 0.5,
  cond.pers.HI.prob.het = 0.49, cond.pers.W.prob.het = 0.48,
  cond.inst.always.prob.het = 1e-04, cond.inst.B.prob.het = 0.85,
  cond.inst.BI.prob.het = 0.73, cond.inst.H.prob.het = 0.78,
  cond.inst.HI.prob.het = 0.76, cond.inst.W.prob.het = 0.72,
  cond.always.prob.corr.het = 1e-04, cond.diag.main.beta.msm = -0.67,
  cond.discl.main.beta.msm = -0.85, cond.diag.pers.beta.msm = -0.67,
  cond.discl.pers.beta.msm = -0.85, cond.diag.inst.beta.msm = -0.67,
  cond.discl.inst.beta.msm = -0.85, cond.diag.main.beta.het = 0,
  cond.discl.main.beta.het = 0, cond.diag.pers.beta.het = 0,
  cond.discl.pers.beta.het = 0, cond.diag.inst.beta.het = 0,
  cond.discl.inst.beta.het = 0, role.B.msm.prob = c(0.242, 0.321, 0.437),
  role.BI.msm.prob = c(0.242, 0.321, 0.437), role.H.msm.prob = c(0.242,
  0.321, 0.437), role.HI.msm.prob = c(0.242, 0.321, 0.437),
  role.W.msm.prob = c(0.242, 0.321, 0.437), role.B.msmf.prob = c(0.242,
  0.321, 0.437), role.BI.msmf.prob = c(0.242, 0.321, 0.437),
  role.H.msmf.prob = c(0.242, 0.321, 0.437), role.HI.msmf.prob = c(0.242,
  0.321, 0.437), role.W.msmf.prob = c(0.242, 0.321, 0.437),
  vv.iev.B.prob = 0.47874, vv.iev.BI.prob = 0.47874,
  vv.iev.H.prob = 0.47874, vv.iev.HI.prob = 0.47874,
  vv.iev.W.prob = 0.47874, prep.start = Inf, prep.elig.model = "base",
  prep.class.prob = c(0, 0, 0, 0), prep.class.hr = c(1, 0.69, 0.19, 0.05),
  prep.coverage = 0, prep.cov.method = "curr", prep.cov.rate = 1,
  prep.tst.int = 90, prep.risk.int = 182, prep.risk.reassess = TRUE,
  msm.temp.adjust = 1, fa.temp.adjust = 1, depart.adjust = 1,
  return.adjust = 1, immig.depart.BI.f = (1/(42 * 52 * 4)) * depart.adjust,
  immig.depart.HI.f = (1/(42 * 52 * 4)) * depart.adjust,
  immig.depart.BI.m = (1/(42 * 52 * 4)) * depart.adjust,
  immig.depart.HI.m = (1/(42 * 52 * 4)) * depart.adjust,
  immig.return.BI.f = (1/4) * return.adjust, immig.return.HI.f = (1/4) *
  return.adjust, immig.return.BI.m = (1/4) * return.adjust,
  immig.return.HI.m = (1/4) * return.adjust, immig.aq.prob.BI.f = 0.3 * 0.06
  * ((base.vi.main.BI.rate + base.vi.pers.BI.rate)/2) * URVI.prob *
  fa.temp.adjust, immig.aq.prob.HI.f = 0.3 * 0.005 * ((base.vi.main.HI.rate +
  base.vi.pers.HI.rate)/2) * URVI.prob * fa.temp.adjust,
  immig.aq.prob.BI.m = 0.3 * 0.06 * ((base.vi.main.BI.rate +
  base.vi.pers.BI.rate)/2) * UIVI.prob * fa.temp.adjust,
  immig.aq.prob.HI.m = 0.3 * 0.005 * ((base.vi.main.HI.rate +
  base.vi.pers.HI.rate)/2) * UIVI.prob * fa.temp.adjust,
  msm.aq.prob.B = 7.183e-06 * msm.temp.adjust, msm.aq.prob.BI = 7.183e-06 *
  msm.temp.adjust, msm.aq.prob.H = 7.183e-06 * msm.temp.adjust,
  msm.aq.prob.HI = 7.183e-06 * msm.temp.adjust, msm.aq.prob.W = 7.183e-06 *
  msm.temp.adjust, demog.list = data.params[[1]]$demog.list,
  demog.dist = data.params[[1]]$demog.dist,
  sex.groups = data.params[[1]]$sex.groups,
  race.groups = data.params[[1]]$race.groups,
  age.groups = data.params[[1]]$age.groups,
  age.adj = data.params[[1]]$age.adj, ...)
}
\arguments{
\item{testing.pattern}{Method for HIV testing, with options \code{"memoryless"}
for constant hazard without regard to time since previous test, or
\code{"interval"} deterministic fixed intervals.}

\item{test.window.int}{Length of the HIV test window period in days.}

\item{max.time.off.tx.full.int}{Number of days off treatment for a full
suppressor before onset of AIDS, including time before diagnosis.}

\item{max.time.on.tx.part.int}{Number of days on treatment for a
partial suppressor beofre onset of AIDS.}

\item{max.time.off.tx.part.int}{Nnumber of days off treatment for a
partial suppressor before onset of AIDS, including time before
diagnosis.}

\item{vl.acute.rise.int}{Number of days to peak viremia during acute
infection.}

\item{vl.acute.peak}{Peak viral load (in log10 units) at the height of acute
infection.}

\item{vl.acute.fall.int}{Number of days from peak viremia to set-point
viral load during the acute infection period.}

\item{vl.set.point}{Set point viral load (in log10 units).}

\item{vl.aids.onset.int}{Number of days to AIDS for a treatment-naive
patient.}

\item{vl.aids.int}{Duration of AIDS stage infection in days.}

\item{vl.fatal}{Viral load in AIDS at which death occurs.}

\item{vl.full.supp}{Log10 viral load at full suppression on ART.}

\item{vl.part.supp}{Log10 viral load at partial suppression on ART.}

\item{full.supp.down.slope}{For full suppressors, number of log10 units that
viral load falls per time step from treatment initiation or re-initiation
until the level in \code{vl.full.supp}.}

\item{full.supp.up.slope}{For full suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected
value.}

\item{part.supp.down.slope}{For partial suppressors, number of log10 units
that viral load falls per time step from treatment initiation or
re-initiation until the level in \code{vl.part.supp}.}

\item{part.supp.up.slope}{For partial suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected value.}

\item{birth.age}{Age (in years) of new arrivals.}

\item{b.method}{Method for calculating the number of expected births at each
time step, with \code{"fixed"} based on the number of persons at the
initial time step and \code{"varying"} based on the current time step.}

\item{msm.frac}{The fraction of males entering the population at each time step 
as a male that has intercourse exclusivly with other males "msm" - 
males not designated as msm or msmf are considered "msf".}

\item{msmf.frac}{The fraction of males entering the population at each time step as
as a male that has intercourse with both males and females "msmf" - 
males not designated as msm or msmf are considered "msf".}

\item{URAI.prob}{Probability of transmission for a man having unprotected
receptive anal intercourse with an infected man at set point viral
load.}

\item{UIAI.prob}{Probability of transmission for an uncircumcised man having
unprotected insertive anal intercourse with an infected man at set
point viral load.}

\item{URVI.prob}{Probability of transmission for a woman having unprotected
receptive vaginal intercourse with an infected man at set point viral
load.}

\item{UIVI.prob}{Probability of transmission for an uncircumcised man having
unprotected insertive vaginal intercourse with an infected women at set
point viral load.}

\item{acute.rr}{Relative risk of infection (compared to that predicted by
elevated viral load) when positive partner is in the acute stage.}

\item{circ.rr}{Relative risk of infection from insertive anal sex when the
negative insertive partner is circumcised.}

\item{condom.rr}{Relative risk of infection from anal sex when a condom is
used.}

\item{ccr5.heteroz.rr}{Relative risk of infection for people who are heterozygous
in the CCR5 mutation.}

\item{num.inst.ai.classes}{Number of quantiles into which men should be
divided in determining their levels of one-off anal intercourse.}

\item{ai.scale}{General relative scaler for all ai act rates for model
calibration.}

\item{vi.scale}{General relative scaler for all vi act rates for model
calibration.}

\item{cond.pers.always.prob.msm}{Fraction of men in male-male casual partnerships who always
use condoms in those partnerships.}

\item{cond.inst.always.prob.msm}{Fraction of men in male-male one-time partnerships who always
use condoms in those partnerships.}

\item{cond.always.prob.corr.msm}{Correlation coefficient for probability of always
using condoms in both casual and one-off male-male partnerships}

\item{cond.pers.always.prob.het}{Fraction of people in heterosexual casual partnerships who always
use condoms in those partnerships.}

\item{cond.inst.always.prob.het}{Fraction of people in heterosexual one-time partnerships who always
use condoms in those partnerships.}

\item{cond.always.prob.corr.het}{Correlation coefficient for probability of always
using condoms in both casual and one-off heterosexual partnerships}

\item{cond.diag.main.beta.msm}{Beta multiplier for the log odds of using a
condom in a male-male main partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.main.beta.msm}{Beta multiplier for the log odds of using a
condom in a male-male main partnership if the HIV-infected man has disclosed.}

\item{cond.diag.pers.beta.msm}{Beta multiplier for the log odds of using a
condom in a male-male casual partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.pers.beta.msm}{Beta multiplier for the log odds of using a
condom in a male-male casual partnership if the HIV-infected man has disclosed
his status.}

\item{cond.diag.inst.beta.msm}{Beta multiplier for the log odds of using a
condom in a male-male one-off partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.inst.beta.msm}{Beta multiplier for the log odds of using a
condom in a male-male one-off partnership if the HIV-infected man has disclosed
his status.}

\item{cond.diag.main.beta.het}{Beta multiplier for the log odds of using a
condom in a heterosexual main partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.main.beta.het}{Beta multiplier for the log odds of using a
condom in a heterosexual main partnership if the HIV-infected man has disclosed.}

\item{cond.diag.pers.beta.het}{Beta multiplier for the log odds of using a
condom in a heterosexual casual partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.pers.beta.het}{Beta multiplier for the log odds of using a
condom in a heterosexual casual partnership if the HIV-infected man has disclosed
his status.}

\item{cond.diag.inst.beta.het}{Beta multiplier for the log odds of using a
condom in a heterosexual one-off partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.inst.beta.het}{Beta multiplier for the log odds of using a
condom in a heterosexual one-off partnership if the HIV-infected man has disclosed
his status.}

\item{prep.start}{Time step at which the PrEP intervention should start.}

\item{prep.elig.model}{Modeling approach for determining who is eligible for
PrEP. Current options are limited to: \code{"all"} for all persons who
have never been on PrEP and are disease-susceptible.}

\item{prep.class.prob}{The probability of adherence class in non-adherent,
low adherence, medium adherence, or high adherence groups (from Liu).}

\item{prep.class.hr}{The hazard ratio for infection per act associated with each
level of adherence (from Grant).}

\item{prep.coverage}{The proportion of the eligible population who are start
PrEP once they become eligible.}

\item{prep.cov.method}{The method for calculating PrEP coverage, with options
of \code{"curr"} to base the numerator on the number of people currently
on PrEP and \code{"ever"} to base it on the number of people ever on
PrEP.}

\item{prep.cov.rate}{The rate at which persons initiate PrEP conditional on
their eligibility, with 1 equal to instant start.}

\item{prep.tst.int}{Testing interval for those who are actively on PrEP. This
overrides the mean testing interval parameters.}

\item{prep.risk.int}{Time window for assessment of risk eligibility for PrEP
in days.}

\item{prep.risk.reassess}{If \code{TRUE}, reassess eligibility for PrEP at
each testing visit.}

\item{age.adj}{This value is added to the squareroot of age for females in sqrt.age.adj
so that sex age asymmetry can be captured with an absdiff term}

\item{...}{Additional arguments passed to the function.}

\item{last.neg.test.R.S.int}{Time range in days for last negative test for
race group R where R is a charater (B, BI, H, HI or W) and sex group S
where s is a charter (f, msf, msm, msmf).}

\item{mean.test.B.int}{Mean intertest interval in days for race group R where R 
is a charater (B, BI, H, HI or W) and sex group S
where s is a charter (f, msf, msm, msmf) who test.}

\item{tt.traj.R.S.prob}{Proportion of race group R, where R is a charater (B, BI, H, HI or W), 
and sex group S, where s is a charter (f, msf, msm, msmf), who enter one of four
testing/treatment trajectories: never test or treat, test and never
initiate treatment, test and treated with partial viral suppression,
and test and treated with full suppression.}

\item{tx.init.R.S.prob}{Probability per time step that a person in race group R, where R is a 
charater (B, BI, H, HI or W), and sex group S, where s is a charter (f, msf, msm, msmf),
who has tested positive will initiate treatment.}

\item{tx.halt.R.S.prob}{Probability per time step that a person in race group R, where R is a 
charater (B, BI, H, HI or W), and sex group S, where s is a charter (f, msf, msm, msmf),
who is currently on treatment will halt treatment.}

\item{tx.reinit.R.S.prob}{Probability per time step that a person in race group R, where R is a 
charater (B, BI, H, HI or W), and sex group S, where s is a charter (f, msf, msm, msmf),
who is not currently on treatment but who has been in the past will
re-initiate treatment.}

\item{b.R.S.rate}{Rate at which a person in race group R, where R is a 
charater (B, BI, H, HI or W), and sex group S, where s is a charter (f, msf, msm, msmf),
enters the population.}

\item{disc.outset.main.R.S.prob}{Probability that an HIV-infected person in race group R,
where R is a charater (B, BI, H, HI or W), and sex group S, where s is a charter 
(f, msf, msm, msmf), will disclose their status at the start of a main partnership.}

\item{disc.at.diag.main.R.S.prob}{Probability that a a person in race group R, where R is a 
charater (B, BI, H, HI or W), and sex group S, where s is a charter (f, msf, msm, msmf),
already in a main partnership will disclose at the time of diagnosis.}

\item{disc.post.diag.main.R.S.prob}{Probability that an HIV-infected a person in race group R, 
where R is a charater (B, BI, H, HI or W), and sex group S, where s is a charter 
(f, msf, msm, msmf), in a main partnership will disclose their status, assuming they didn't
at the start of the partnership or at diagnosis.}

\item{disc.outset.pers.R.S.prob}{Probability that an HIV-infected a person in race group R, 
where R is a charater (B, BI, H, HI or W), and sex group S, where s is a charter
(f, msf, msm, msmf), will disclose their status at the start of a casual partnership.}

\item{disc.at.diag.pers.R.S.prob}{Probability that a person in race group R, where R is a
charater (B, BI, H, HI or W), and sex group S, where s is a charter (f, msf, msm, msmf),
already in a casual partnership will disclose at the time of diagnosis.}

\item{disc.post.diag.pers.R.S.prob}{Probability that an HIV-infected person in race group R,
where R is a charater (B, BI, H, HI or W), and sex group S, where s is a charter 
(f, msf, msm, msmf), in a casual partnership will disclose their status, assuming they
didn't at the start of the partnership or at diagnosis.}

\item{disc.inst.R.S.prob}{Probability that an HIV-infected person in race group R, 
where R is a charater (B, BI, H, HI or W), and sex group S, where s is a charter
(f, msf, msm, msmf), will disclose ther status to a one-off partner.}

\item{circ.R.prob}{Probablity that a male in race group R, where R is a 
charater (B, BI, H, HI or W), newly arriving in the population
will be circumcised.}

\item{ccr5.R.S.prob}{Vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respectively) in the CCR5 gene
among people in race group R, where R is a charater (B, BI, H, HI or W), 
and sex group S, where s is a charter (f or m).}

\item{base.ai.main.R.rate}{Expected coital frequency (for anal intercourse) for males 
(msm and msmf) in race group R, where R is a charater (B, BI, H, HI or W), 
in main partnerships (acts per day).}

\item{base.vi.main.R.rate}{Expected coital frequency (for vaginal intercourse) for 
males and females in race group R, where R is a charater (B, BI, H, HI or W), 
in main partnerships (acts per day).}

\item{base.ai.pers.R.rate}{Expected coital frequency (for anal intercourse) for males 
(msm and msmf) in race group R, where R is a charater (B, BI, H, HI or W), 
in casual partnerships (acts per day).}

\item{base.vi.pers.R.rate}{Expected coital frequency (for vaginal intercourse) for 
males and females in race group R, where R is a charater (B, BI, H, HI or W), 
in casual partnerships (acts per day).}

\item{cond.main.R.prob.msm}{Probability of condom use by an male of race group R, 
where R is a charater (B, BI, H, HI or W), in a male-male main partnership.}

\item{cond.pers.R.prob.msm}{Of men who are not consistent condom users, per-act
probability of condom use in a casual male-male partnership by 
males of race group R, where R is a charater (B, BI, H, HI or W).}

\item{cond.inst.R.prob.msm}{Of men who are not consistent condom users, per-act
probability of condom use in a one-time male-male partnership by 
males of race group R, where R is a charater (B, BI, H, HI or W).}

\item{cond.main.R.prob.het}{Probability of condom use by an person of race group R, 
where R is a charater (B, BI, H, HI or W), in a heterosexual main partnership.}

\item{cond.pers.R.prob.het}{Of people who are not consistent condom users, per-act
probability of condom use in a casual heterosexual partnership by 
people of race group R, where R is a charater (B, BI, H, HI or W).}

\item{cond.inst.R.prob.het}{Of people who are not consistent condom users, per-act
probability of condom use in a one-time heterosexual partnership by 
males of race group R, where R is a charater (B, BI, H, HI or W).}

\item{vv.iev.R.prob}{Probability that a person of group R, where R is a charater (B, BI, H, HI or W),
in a male male relationship will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{immig.depart.R.S}{The probability per time step that a person of group R,
where R is a charater (BI or HI), and sex S, were S is a charater (f or m), leaves the population
being simulated by returning to their home country.}

\item{immig.return.R.S}{The probability per time step that a person of group R,
where R is a charater (BI or HI), and sex S, were S is a charater (f or m), who has left
the population being simulated by returning to their home country returns to the population
being simulated.}

\item{immig.aq.prob.R.S}{The probability per time step that a person of group R,
where R is a charater (BI or HI), and sex S, were S is a charater (f or m), who has left
the population being simulated aquires HIV while in their home country.}
}
\value{
A list object of class \code{param_shamp}, which can be passed to
EpiModel function \code{netsim}.
}
\description{
Sets the epidemic parameters for stochastic network models
             simulated with \code{\link{netsim}} for EpiModelHIV
}
\keyword{shamp}
