% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.dat_updater.R
\name{param_updater}
\alias{param_updater}
\title{Module to modify the `param` list at specified time steps during the simulation}
\usage{
param_updater(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module to modify the `param` list at specified time steps during the simulation
}
\details{
if a list `dat$param$param_updaters` is present, this module will update the
`param` list with new values at given timesteps.
`dat$control$param_updaters` is a list containing `updaters`. An updater is a
list containing an `at` element telling when the changes will happend, an
optional `verbose` boolean controlling whether to output a message when a
change is made (default = TRUE) and a `param` list similar
to the simulation's `dat$param` list.
if the updater is a function but not the value to replace, the
function will be applied to the current element (see example) .
}
\examples{
 ## example of a `param_updaters` list
 list(
   list(
     at = 4860,
     param = list(
       hiv.test.rate = rep(0.0128, 3),
       trans.scale = c(1.61, 0.836, 0.622)
     )
   ),
   list(
     at = 5160,
     verbose = FALSE,
     param = list(
       hiv.test.rate = function(x) x * 3,
       trans.scale = function(x) x^2 / 3
     )
   )
 )

}
