% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.part.R
\name{part_msm}
\alias{part_msm}
\title{Partnership tracking Module}
\usage{
part_msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with the updated master partnership list,
on \code{temp$part.list}.
}
\description{
Module function for tracking partnerships for STD testing and EPT.
}
\details{
Partnerships are tracked in a persistent edge list that allows for easy
reference to determine if a participant has been in a particular type of
relationship within a defined time frame
infected; or post diagnosis for one recently infected. The rates of disclosure
vary at these three points, and also by the partnership type.
}
\keyword{module}
\keyword{msm}
