% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.position.shamp.R
\name{position_shamp}
\alias{position_shamp}
\title{Position Module for msm and heterosexual contacts.}
\usage{
position_shamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the updated discordant edgelist with a \code{ins}
attribute for values of whether the infected node is insertive or the
susceptible node is insertive for that act.
}
\description{
Module function for establishing sexual role or position in each
             act on the discordant edgelist.
}
\details{
The sexual role within each act is determined by each nodes "role identity"
as exclusively receptive, exclusively insertive, or versatile. For heterosexual dyads
the females node is always recpetive and the male is always incertive.  This function
determines whether the infected or the susceptible partner is the insertive
partner for that act. For male-male dyads the first two role identity types, that is
deterministic based on identity. For male-male versatile-versatile pairs, this is
determined stochastically for each act.
}
\keyword{SHAMP}
\keyword{het}
\keyword{module}
\keyword{msm}
