% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.prep.R
\name{prep.mard}
\alias{prep.mard}
\title{PrEP Module}
\usage{
prep.mard(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function for implementation and uptake of pre-exposure
             prophylaxis (PrEP) to prevent HIV infection.
}

