% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.update.PrEP.adherence.R
\name{prep_adherence_trans}
\alias{prep_adherence_trans}
\title{Update PrEP adherence from asmm to MSM values at age 19}
\usage{
prep_adherence_trans(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function updates the individual-level attribute \code{prepClass} on
\code{dat$attr}.
}
\description{
Module function for updating act class in main and casual
             partnerships based on probabilities of transition.
}
\keyword{module}
\keyword{msm}
