% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.prep.adol.R
\name{prep_adol}
\alias{prep_adol}
\title{PrEP ASMM Module}
\usage{
prep_adol(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function for implementation and uptake of pre-exposure
             prophylaxis (PrEP) among ASMM to prevent HIV infection.
}
