% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.prep.KTM.R
\name{prep_KTM}
\alias{prep_KTM}
\title{PrEP Module for TMP using 4 indications.}
\usage{
prep_KTM(dat, at)
}
\description{
Module function for implementation and uptake of pre-exposure
             prophylaxis (PrEP) to prevent HIV infection.  Indications are based on Partner services, 
             disclosure, risk(partners) and all negatives identified through TMP.
}
\keyword{module}
\keyword{shamp}
