% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.prevalence.R
\name{prevalence_het}
\alias{prevalence_het}
\title{Prevalence Module}
\usage{
prevalence_het(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function to calculate and store summary statistics for
             disease prevalence, demographics, and other epidemiological
             outcomes.
}
\keyword{het}
\keyword{module}
