% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.prevalence.shamp.R
\name{prevalence_shamp}
\alias{prevalence_shamp}
\title{Prevalence Calculations within Time Steps for up to 5 race groups heterosexuals and MSM.}
\usage{
prevalence_shamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with an updated summary of current
attributes stored in \code{dat$epi}.
}
\description{
This module calculates demographic, transmission, and clinical
             statistics at each time step within the simulation.
}
\details{
Summary statistic calculations are of two broad forms: prevalence and
incidence. This function establishes the summary statistic vectors for both
prevalence and incidence at time 1, and then calculates the prevalence
statistics for times 2 onward. Incidence statistics (e.g., number of new
infections or deaths) are calculated within the modules as they depend on
vectors that are not stored external to the module.
}
\keyword{SHAMP}
\keyword{het}
\keyword{module}
\keyword{msm}
