% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.progress.R
\name{progress.mard}
\alias{progress.mard}
\title{Disease Progression Module}
\usage{
progress.mard(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object after updating the disease stage
of infected individuals.
}
\description{
Module function for HIV disease progression through acute, chronic
             and AIDS stages.
}
\details{
HIV disease is divided into four stages: acute rising, acute falling, chronic
and AIDS. Acute rising is the time from infection to peak viremia, while
acute falling is the time from peak viremia to chronic stage infection with
an established set-point HIV viral load.

The time spent in chronic stage infection, and thus the time from infection to
AIDS, depends on ART history. For ART-naive persons, time to AIDS is established
by the \code{vl.aids.onset} parameter. For persons ever on ART who fall into
the partially suppressed category (the \code{tt.traj} attribute is \code{"YP"}),
time to AIDS depends on the sum of two ratios: time on treatment over maximum
time on treatment plus time off treatment over maximum time off treatment.
For persons ever on ART who fall into the fully suppressed cateogry
(\code{tt.traj="YF"}), time to AIDS depends on whether the cumulative time
off treatment exceeds a time threshold specified in the \code{max.time.off.tx.full}
parameter.
}
\keyword{module}

