% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.progress.WHAMP.R
\name{progress_msm_whamp}
\alias{progress_msm_whamp}
\title{Disease Progression Module}
\usage{
progress_msm_whamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object after updating the disease stage
of infected individuals.
}
\description{
Module function for HIV disease progression through acute, chronic
             and AIDS stages in the WHAMP module.
}
\details{
HIV disease is divided into four stages: acute rising, acute falling, chronic
and AIDS. Acute rising is the time from infection to peak viremia, while
acute falling is the time from peak viremia to chronic stage infection with
an established set-point HIV viral load.

The time spent in chronic stage infection, and thus the time from infection to
AIDS, depends on ART history. For ART-naive persons, time to AIDS is established
by the \code{vl.aids.onset} parameter. For persons ever on ART, time to AIDS depends 
on whether the cumulative time off treatment exceeds a time threshold specified in 
the \code{max.time.off.tx.int} parameter.
}
\keyword{module}
\keyword{msm}
