% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.partnerservices.KTM.R
\name{pservices_KTM}
\alias{pservices_KTM}
\title{Partner Services Module}
\usage{
pservices_KTM(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function for tracking partnerships over a predetermined time window.
}
\keyword{module}
\keyword{shamp}
