% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.partnerservices.shamp.R
\name{pservices_shamp}
\alias{pservices_shamp}
\title{Partner Services Module}
\usage{
pservices_shamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function for tracking partnerships over a predetermined time window.
}
\keyword{module}
\keyword{shamp}
