% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reallocate_pcp}
\alias{reallocate_pcp}
\title{Proportionally Reallocate PrEP Adherence Class Probability}
\usage{
reallocate_pcp(in.pcp = c(0.211, 0.07, 0.1, 0.619), reall = 0)
}
\arguments{
\item{in.pcp}{Input vector of length four for the \code{prep.class.prob}
parameter.}

\item{reall}{The pure percentage points to shift from the high adherence
group to the lower three groups.}
}
\description{
Shifts probabilities from the high-adherence category to the lower
             three adherence categories while maintaining the proportional
             distribution of those lower categories.
}
