% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.initialize.R
\name{reinit_het}
\alias{reinit_het}
\title{Reinitialization Module}
\usage{
reinit_het(x, param, init, control, s)
}
\arguments{
\item{x}{An \code{EpiModel} object of class \code{\link{netest}}.}

\item{param}{An \code{EpiModel} object of class \code{\link{param_het}}.}

\item{init}{An \code{EpiModel} object of class \code{\link{init_het}}.}

\item{control}{An \code{EpiModel} object of class \code{\link{control_het}}.}

\item{s}{Simulation number, used for restarting dependent simulations.}
}
\value{
This function returns the updated \code{dat} object with the initialized values
for demographics and disease-related variables.
}
\description{
This function reinitializes the master \code{dat} object on which
             data are stored, simulates the initial state of the network, and
             simulates disease status and other attributes.
}
\keyword{het}
\keyword{module}
