% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.init.R
\name{remove_bad_roles}
\alias{remove_bad_roles}
\title{Removes any sexual partnerships prohibited by sexual role mixing}
\usage{
remove_bad_roles(nw)
}
\arguments{
\item{nw}{An object of class \code{network}.}
}
\description{
Due to occassional issues in ERGM fitting, it is possible to
             have initial simulations in which there are pairings between
             exclusively insertive/insertive or receptive/receptive.
}
\keyword{initiation}
\keyword{utility}

