% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.riskhist.R
\name{riskhist.msm}
\alias{riskhist.msm}
\title{Risk History Module}
\usage{
riskhist.msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function to track the risk history of uninfected persons
             for purpose of intervention targeting.
}
\keyword{module}

