% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.riskhist.R
\name{riskhist_ept_msm}
\alias{riskhist_ept_msm}
\title{Risk History for EPT Module}
\usage{
riskhist_ept_msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function to track the risk history of uninfected persons
             for purpose of EPT prevention intervention targeting.
}
\keyword{module}
\keyword{msm}
