% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.riskhist.KTM.R
\name{riskhist_KTM}
\alias{riskhist_KTM}
\title{Risk History Module}
\usage{
riskhist_KTM(dat, at)
}
\description{
Module function to track the risk history of uninfected persons
             for purpose of intervention targeting.
}
\keyword{SHAMP}
\keyword{module}
