% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.prep.WHAMP.R, R/mod.riskhist.WHAMP.R
\name{riskhist_msm_whamp}
\alias{riskhist_msm_whamp}
\alias{riskhist_msm_whamp}
\title{Risk History Sub-Module}
\usage{
riskhist_msm_whamp(dat, at)

riskhist_msm_whamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Sub-Module function to track the risk history of uninfected persons
             for purpose of PrEP targeting in the WHAMP model.

Module function to track the risk history of uninfected persons
             for purpose of intervention targeting for the WHAMP model.
}
\keyword{module}
\keyword{msm}
