% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.riskhist_risktest.R
\name{riskhist_risktest_msm}
\alias{riskhist_risktest_msm}
\title{Risk History for Risk Based Testers Module}
\usage{
riskhist_risktest_msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function to track the risk history persons
             who test selectively based on recent risk.
}
\keyword{module}
\keyword{msm}
