% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mod.births.R
\name{setBirthAttr.mard}
\alias{setBirthAttr.mard}
\title{Set Attributes for Incoming Nodes}
\usage{
setBirthAttr.mard(dat, at, nBirths.B, nBirths.W)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}

\item{nBirths.B}{Number of new entries among black MSM.}

\item{nBirths.W}{Number of new entries among white MSM.}
}
\description{
Updates the attr list for new entries into the population as
             set by the births module.
}
\keyword{submodule}

