% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.births.R
\name{setBirthAttr_het}
\alias{setBirthAttr_het}
\title{Assign Vertex Attributes at Network Entry}
\usage{
setBirthAttr_het(dat, at, nBirths)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}

\item{nBirths}{Number of new births as determined by \code{\link{births_het}}.}
}
\description{
Assigns vertex attributes to incoming nodes at birth/entry into
             the network.
}
\keyword{het}
