% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.simnet.R
\name{simnet_het}
\alias{simnet_het}
\title{Network Resimulation Module}
\usage{
simnet_het(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function to resimulate the dynamic network forward one
             time step conditional on current network structure and vertex
             attributes.
}
\keyword{het}
\keyword{module}
