% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.simnet.WHAMP.R
\name{simnet_msm_whamp}
\alias{simnet_msm_whamp}
\title{Network Resimulation Module}
\usage{
simnet_msm_whamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function for resimulating the sexual networks for one
             time step in the WHAMP model
}
\keyword{module}
\keyword{msm}
