% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.simnet.shamp.R
\name{simnet_shamp}
\alias{simnet_shamp}
\title{Network Resimulation Module}
\usage{
simnet_shamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function for resimulating the main, casual, and one-off
             networks for one time step.
}
\keyword{module}
\keyword{shamp}
