% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{st}
\alias{st}
\title{Network Statistics Network Models for 10000 Person Network}
\format{A list with 18 elements.}
\description{
A dataset containing the necessary network statistics for efficient
package testing.
}
\keyword{datasets}
