% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stergm_prep}
\alias{stergm_prep}
\title{Prep stergm}
\usage{
stergm_prep(
  nw,
  formation,
  dissolution,
  coef.form,
  coef.diss,
  constraints,
  control = control.simulate.network()
)
}
\arguments{
\item{path}{Directory of files to source.}

\item{verbose}{Print names of sourced files to console.}

\item{...}{Additional arguments passed to \code{source}.}
}
\description{
set up the structure for the stergm function. No longer exported in tegmLite
}
