% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.ept.R
\name{sti_ept_msm}
\alias{sti_ept_msm}
\title{EPT Module}
\usage{
sti_ept_msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function for provision  of expedited
             partner therapy (EPT) from index partner to non-index partner
             and uptake by non-index partner to prevent STI infection.
             Eligibility for index partner is handled in the STI treatment
             module and eligibility for non-index handled in risk history
             module.
}
\keyword{module}
\keyword{msm}
