% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.sti.R
\name{sti_recov}
\alias{sti_recov}
\title{STI Recovery Module}
\usage{
sti_recov(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Stochastically simulates GC/CT recovery.
}
\keyword{module}
\keyword{msm}
