% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.test.R
\name{sti_test_msm}
\alias{sti_test_msm}
\title{STI Testing Module}
\usage{
sti_test_msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with updated
\code{last.neg.test}, \code{diag.status} and \code{diag.time} attributes for
each STI.
}
\description{
Module function for STI screening of asymptomatic persons.
}
\details{
This testing module supports two testing parameterizations, input via the
\code{testing.pattern} parameter: memoryless for stochastic and
geometrically-distributed waiting times to test (constant hazard); and
interval for deterministic tested after defined waiting time intervals.
Symptomatic testing is handled in the STI treatment module.
}
\keyword{module}
\keyword{msm}
