% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.sti.R
\name{sti_tx_msm}
\alias{sti_tx_msm}
\title{STI Treatment Module}
\usage{
sti_tx_msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Stochastically simulates GC/CT and syphilis diagnosis and 
             treatment.
}
\keyword{module}
\keyword{msm}
