% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.stitx.R
\name{stitx_msm}
\alias{stitx_msm}
\title{STI Treatment Module}
\usage{
stitx_msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Stochastically simulates GC/CT diagnosis and treatment.
}
\keyword{module}
\keyword{msm}
