% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.WHAMP.R
\name{sumto1}
\alias{sumto1}
\title{Make sure specified proportions sum to 1}
\usage{
sumto1(values)
}
\arguments{
\item{values}{Values of the input vector.}
}
\description{
Checks that specified proportions sum to 1 to account for rounding
             in calculation of the input parameters. If the sum of a vector of 
             values is within 0.001 of 1, this function adds or subtracts from 
              each of the set of values such that the total sums to 1. If the sum 
              of the values is more than 0.001 greater or smaller than 1, it 
              returns an error.
}
