% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.test.WHAMP.R
\name{test_msm_whamp}
\alias{test_msm_whamp}
\title{HIV Testing Module}
\usage{
test_msm_whamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with updated \code{last.neg.test},
\code{diag.status} and \code{diag.time} attributes.
}
\description{
Module function for HIV diagnostic testing of infected persons for the WHAMP model.
}
\details{
This testing module supports testing as an interval process and diagnostic testing 
upon onset of AIDS-associated symptoms
}
\keyword{module}
\keyword{msm}
