% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.test.shamp.R
\name{test_shamp}
\alias{test_shamp}
\title{HIV Testing Module for up to 5 race groups heterosexuals and MSM.}
\usage{
test_shamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with updated \code{last.neg.test},
\code{diag.status} and \code{diag.time} attributes.
}
\description{
Module function for HIV diagnostic testing of infected persons.
}
\details{
This testing module supports two testing parameterizations, input via the
\code{testing.pattern} parameter: memoryless for stochastic and
geometrically-distributed waiting times to test (constant hazard); and interval
for deterministic tested after defined waiting time intervals.
If immigration is used those with dat$attr$immig.loc=1 do not test.
}
\keyword{het}
\keyword{module}
\keyword{msm}
\keyword{shamp}
