% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.hometesting.interventions.R
\name{testing_interventions}
\alias{testing_interventions}
\title{HIV Testing Module for hometesting project:Testing interventions}
\usage{
testing_interventions(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with updated \code{last.neg.test},
\code{diag.status} and \code{diag.time} attributes.
}
\description{
Module function for HIV diagnostic testing of infected persons using clinic and hometests following 
four testing types (None testers (can test under supplimentations interventions), opportunistic only testers
regular testers and risk based testers. All tester use opportunistic tests and there replacement or 
supplementation effects opportunistic tests for all tester types.
}
\details{
This testing module supports two testing interventions, replacement of exsisting test with home based tests
and supplimentation of existing tests with homebased tests.
}
\keyword{module}
