% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.trans.R
\name{trans_het}
\alias{trans_het}
\title{Infection Module}
\usage{
trans_het(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function to simulate transmission over an active
             discordant edgelist.
}
\keyword{het}
\keyword{module}
