% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.trans_shamp.R
\name{trans_shamp}
\alias{trans_shamp}
\title{Transmission Module for AI and VI}
\usage{
trans_shamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
For each new infection, the disease status, infection time, and related
HIV attributes are updated for the infected node. Summary statistics for
disease incidence overall, and by race and age groups are calculated and
stored on \code{dat$epi}.
}
\description{
Stochastically simulates disease transmission given the current
             state of the discordand edgelist.
}
\details{
This is the final substantive function that occurs within the time loop at
each time step. This function takes the discordant edgelist and calculates a
transmission probability for each row (one sexual act) between dyads on the
network. After transmission events, individual-level attributes for the infected
persons are updated and summary statistics for incidence calculated.

The per-act transmission probability depends on the following elements:
insertive versus receptive role, viral load of the infected partner, an
acute stage infection excess risk, condom use, and the CCR5 genetic allele.
Given these transmission probabilities, transmission is stochastically
simulating by drawing from a binomial distribution for each act conditional
on the per-act probability.
}
\keyword{SHAMP}
\keyword{het}
\keyword{module}
\keyword{msm}
