% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.tx.R
\name{tx_het}
\alias{tx_het}
\title{HIV Anti-Retroviral Treatment Module}
\usage{
tx_het(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function for simulating HIV therapy after diagnosis,
             including adherence and non-adherence to ART.
}
\keyword{het}
\keyword{module}
