% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.tx.WHAMP.R
\name{tx_msm_whamp}
\alias{tx_msm_whamp}
\title{Treatment Module}
\usage{
tx_msm_whamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with updated \code{tx.status},
\code{tx.init.time}, \code{cum.time.on.tx}, \code{cum.time.off.tx} attributes.
}
\description{
Module function for anti-retroviral treatment initiation and
             adherence over time for the WHAMP model.
}
\details{
Persons enter into the simulation with one of four ART "patterns": non-screeners
who are diagnosed upon progression to AIDS-associated symptoms and treat to partial 
HIV viral suppression, non-screeners who acheive full viral suppression, regular
screeners who treat to partial viral suppression, and regular screeners who treat
to full viral suppression. These types are stored as individual-level attributes 
in \code{tt.traj}). This module initiates ART for treatment-naive persons and 
cycles them on and off treatment conditional on overall discontinuation rates for
full and partial suppessors and reinitiation rates specific to each group defined 
by race/ethnicity, region, and full vs. partial adherence. ART initiation is deterministic,
whereas discontinuation and reinitiation are stochastically simulated based on binomial 
statistical models.
}
\keyword{module}
\keyword{msm}
