% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{update_degree}
\alias{update_degree}
\title{Update Degree Vertex Attribute on Network Objects}
\usage{
update_degree(nw, nw.source, deg.type, at)
}
\arguments{
\item{nw}{Object of class \code{network} that is the target for updated
degree vertex attribute.}

\item{nw.source}{Object of class \code{network} that is the source of the
updated degree vertex attribute.}

\item{deg.type}{Type of degree to assign to \code{nw}, with options of
\code{"pers"} to assign casual degree onto main network and
\code{"main"} to assign main degree to casual network.}

\item{at}{Current time step.}
}
\description{
Updates the degree vertex attributes on network objects
             conditional on their values from the other networks.
}
\details{
This function queries the degree of each node within a network using the
sociality term in a summary function call to the network, and then sets this
as a race-based vertex attribute on the output network.
}
\seealso{
\code{\link{assign_degree}}
}

