% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.dat_updater.R
\name{update_list}
\alias{update_list}
\title{Update list `x` using the elements of list `new_x`}
\usage{
update_list(x, new_x)
}
\arguments{
\item{x}{a list}

\item{new_x}{a list}
}
\value{
the full `x` list with the modifications added by `new_x`
}
\description{
Update list `x` using the elements of list `new_x`
}
\details{
This function updates list `x` by name. If `x` and `new_x` elements are not
named, the function will not work properly.
If a function is provided to replace an element that was originaly not a
function, this function will be applied to the original value.
}
