% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.update.roleclass.R
\name{update_roleclass_msm}
\alias{update_roleclass_msm}
\title{Update Role Class in Main and Casual Partnerships}
\usage{
update_roleclass_msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function updates the individual-level attribute \code{role.class} on
\code{dat$attr}.
}
\description{
Module function for updating act class in main and casual
             partnerships based on probabilities of transition.
}
\keyword{module}
\keyword{msm}
