% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.simnet.R
\name{updateInputs}
\alias{updateInputs}
\title{Alternate Methods for Computing Ergm Term Inputs}
\usage{
updateInputs(dat, network = 1)
}
\arguments{
\item{dat}{EpiModel dat object tracking simulation state}

\item{network}{Numberic number of network location for multi-network simulations.}
}
\description{
Function to appropriately update model params based on ergm model
             terms when using 'fast_edgelist' representations.
}
\details{
Contains hard-coded implementations of some of the most commonly used
ergm terms called instead of InitErgmTerm.x, checkErgmTerm, etc.

Implemented terms are:
 \itemize{
   \item edges
   \item nodematch
   \item nodefactor
   \item concurrent
   \item absdiff
 }
 All other terms will return errors.
}

