% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.updatenwp.R
\name{updatenwp.msm}
\alias{updatenwp.msm}
\title{Update Network Data Structure and Parameters}
\usage{
updatenwp.msm(dat, network)
}
\arguments{
\item{dat}{Data object created in initialization module.}

\item{network}{Integer value for network number}
}
\description{
Updates the internal data structure containing the main data
             passed into the TERGM resimulation algorithm. This step is
             necessary with the new tergmLite approach.
}
\keyword{module}

