% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.verbose.R
\name{verbose_het}
\alias{verbose_het}
\title{Verbose Module}
\usage{
verbose_het(x, type, s, at)
}
\arguments{
\item{x}{If the \code{type} is "startup", then an object of class
\code{control.net}, otherwise the all master data object in \code{netsim}
simulations.}

\item{type}{Progress type, either of "startup" for starting messages before
all simulations, or "progress" for time step specific messages.}

\item{s}{Current simulation number, if type is "progress"}

\item{at}{Current time step, if type is "progress"}
}
\description{
Prints simulation model progress within the time loop.
}
\details{
In interactive mode, this module function prints out a standard set of
demographic and epidemiologic summary statistics to the console window. In
non-interactive, batch mode these are saved onto \code{.txt} files in a
\code{verb/} subdirectory. This subdirectory will be created if it does not
exist.
}
\keyword{het}
\keyword{module}
