% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.vl.R
\name{vl_het}
\alias{vl_het}
\title{Viral Load Module}
\usage{
vl_het(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function for simulating progression of HIV viral load in
             natural disease dynamics and in the presence of ART.
}
\keyword{het}
\keyword{module}
