% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.vl.WHAMP.R
\name{vl_msm_whamp}
\alias{vl_msm_whamp}
\title{Viral Load Module}
\usage{
vl_msm_whamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with updated \code{vl} attribute.
}
\description{
Module function for updating HIV viral load for the WHAMP model.
}
\details{
HIV viral load varies over time as a function of time since infection and ART
history. In the absence of ART, VL rises during the acute rising stage and
falls during the acute falling stage, until it reaches a set-point value in
chronic stage infection. VL again rises during untreated AIDS stage disease 
until the point of death.

For persons who have ever initated treatment, VL changes depending on current ART
use, time since ART (re)initiation, or time since ART cessation. Current use is 
associated with a reduction in VL, with the rates of decline and nadirs dependent 
on partial or full suppression levels. Treatment cessation is associated with an 
equal rate of increase in VL up to set point. For individuals who have progressed to
AIDS, VL rises up to set point at the same rate as individuals in chronic stage infection
and then rises beyond that at the rate that results in VL of 7 log10 after 2 years. 
Because HIV-related mortality is modeled as a constant increased hazard of death rather
than upon progression to the end of AIDS, VL is constrained not to exceed 7 log10.
}
\keyword{module}
\keyword{msm}
