% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aging_het}
\alias{aging_het}
\title{Aging Module}
\usage{
aging_het(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
This module ages all nodes in the population by one time
             unit at each time step.
}
\keyword{het}
\keyword{module}
