% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimation.R
\name{calc_nwstats.mard}
\alias{calc_nwstats.mard}
\title{Calculate Target Statistics for Network Model Estimation}
\usage{
calc_nwstats.mard(tUnit = 7, num.B, num.W, deg.mp.B, deg.mp.W, mdeg.inst.B,
  mdeg.inst.W, prop.hom.mpi.B, prop.hom.mpi.W, balance = "mean",
  sqrt.adiff.BB, sqrt.adiff.WW, sqrt.adiff.BW, age.method = "heterogeneous",
  diss.main, diss.pers, durs.main, durs.pers, ages, asmr.B, asmr.W, role.freq.B,
  role.freq.W)
}
\arguments{
\item{tUnit}{Time unit relative to 1 for daily.}

\item{num.B}{Population size of black MSM.}

\item{num.W}{Population size of white MSM.}

\item{deg.mp.B}{Degree distribution matrix for main and casual partners for
black MSM.}

\item{deg.mp.W}{Degree distribution matrix for main and causal partners for
white MSM.}

\item{mdeg.inst.B}{Mean degree or rate of instant partnerships per time step
for black MSM.}

\item{mdeg.inst.W}{Mean degree or rate of instant partnerships per time step
for white MSM.}

\item{prop.hom.mpi.B}{Proportion of main, casual, and instant partnerships
in same race for black MSM.}

\item{prop.hom.mpi.W}{Proportion of main, casual, and instant partnerships
in same race for white MSM.}

\item{balance}{Balancing of edges by race for number of mixed-race partnerships,
with options of \code{"black"} to apply black MSM counts, \code{"white"}
to apply white MSM counts, and \code{"mean"} to average over them.}

\item{sqrt.adiff.BB}{Mean absolute differences in the square root of ages in
main, casual, and instant partnerships between two black MSM.}

\item{sqrt.adiff.WW}{Mean absolute differences in the square root of ages in
main, casual, and instant partnerships between two white MSM.}

\item{sqrt.adiff.BW}{Mean absolute differences in the square root of ages in
main, casual, and instant partnerships in mixed-raced couples.}

\item{age.method}{Method for calculating the square root of age differences,
with \code{"heterogeneous"} using the individual \code{sqrt.adiff.XX}
values and \code{"homogeneous"} using a weighted average of them.}

\item{diss.main}{Dissolution model formula for main partnerships.}

\item{diss.pers}{Dissolution model formula for casual partnerships.}

\item{durs.main}{Duration of BB, BW, and WW main partnerships.}

\item{durs.pers}{Duration of BB, BW, and WW casual partnerships.}

\item{ages}{Vector of ages in years for population.}

\item{asmr.B}{Age-sex-specific mortality rate for black MSM.}

\item{asmr.W}{Age-sex-specific mortality rate for white MSM.}

\item{role.freq.B}{Frequency of black MSM as insertive, receptive, and vers.}

\item{role.freq.W}{Frequency of white MSM as insertive, receptive, and vers.}
}
\description{
Translates the raw parameters for the network model to target
             statistics used in the estimation and simulation of the network
             model.
}

