% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.WHAMP.R
\name{calc_nwstats_msm_whamp}
\alias{calc_nwstats_msm_whamp}
\title{Calculate Target Statistics for Network Model Estimation for the WHAMP model}
\usage{
calc_nwstats_msm_whamp(time.unit = 7, num.H.KC, num.B.KC, num.O.KC,
  num.H.OW, num.B.OW, num.O.OW, num.H.EW, num.B.EW, num.O.EW, agestr,
  deg.mp, deg.mp.H, deg.mp.B, deg.mp.O, deg.m.region, prop.deg.p.EW,
  prop.deg.p.KC, prop.deg.p.OW, mdeg.inst, qnts.18to49, qnts.50to59,
  inst.bho, inst.region, prop.hom.mpi.H, prop.hom.mpi.B, prop.hom.mpi.O,
  sqrt.adiff.mpi, prop.hom.region.mpi, diss.main, diss.pers, durs.main,
  durs.pers, ages, asmr.H..wa, asmr.B..wa, asmr.O..wa, role.prob)
}
\arguments{
\item{time.unit}{Time unit relative to 1 for daily.}

\item{agestr}{Vector with the proportion of MSM in each age group 18-24, 25-29...50-59}

\item{deg.mp}{Degree distribution matrix for main and casual partners overall,
as a 2 by 3 matrix.}

\item{deg.mp.H}{Degree distribution matrix for main and causal partners for
Hispanic MSM, as a 2 by 3 matrix.}

\item{deg.mp.B}{Degree distribution matrix for main and causal partners for
Black MSM, as a 2 by 3 matrix.}

\item{deg.mp.O}{Degree distribution matrix for main and causal partners for
Other race/ethnicity MSM, as a 2 by 3 matrix.}

\item{deg.m.region}{Main degree distribution by region (EW, KC, OW)}

\item{prop.deg.p.EW}{Proportion of men with persistent degree 0, 1, and 2+ who live in eastern WA}

\item{prop.deg.p.KC}{Proportion of men with persistent degree 0, 1, and 2+ who live in King County}

\item{prop.deg.p.OW}{Proportion of men with persistent degree 0, 1, and 2+ who live in western WA}

\item{mdeg.inst}{Mean degree, or rate, of one-off partnerships per day.}

\item{qnts.18to49}{Means rates withing quartiles of the distribution of instantaneous partnerships 
for MSM ages 18-49. Use \code{NA} to ignore these quantiles in the target statistics.}

\item{qnts.50to59}{Means rates withing quartiles of the distribution of instantaneous partnerships
for MSM ages 50-59. Use \code{NA} to ignore these quantiles in the target statistics.}

\item{inst.bho}{Mean rate of instantaneous partnerships by racial/ethnic group}

\item{inst.region}{Distribution of instantaneous partnerships by region (EW, KC, OW)}

\item{prop.hom.mpi.H}{A vector of length 3 for the proportion of main, casual,
and one-off partnerships in same race for Hispanic MSM.}

\item{prop.hom.mpi.B}{A vector of length 3 for the proportion of main, casual,
and instantaneous partnerships in same race for black MSM.}

\item{prop.hom.mpi.O}{A vector of length 3 for the proportion of main, casual,
and instantaneous partnerships in same race for other race/ethnicity MSM.}

\item{sqrt.adiff.mpi}{Vector of length 3 with the mean absolute differences
in the square root of ages in main, casual, and instantaneous partnerships.}

\item{prop.hom.region.mpi}{A vector of length 3 for the proportion of main, casual,
and instantaneous partnerships that are within-region.}

\item{diss.main}{Dissolution model formula for main partnerships.}

\item{diss.pers}{Dissolution model formula for casual partnerships.}

\item{durs.main}{Duration of main partnerships in days.}

\item{durs.pers}{Duration of persistent partnerships in days.}

\item{ages}{Integer vector of ages in years that defines range of possible
initial ages in the population.}

\item{asmr.H..wa}{Vector of length 60 defining the age-specific
mortality rates for for Hispanic MSM.}

\item{asmr.B..wa}{Vector of length 60 defining the age-specific
mortality rates for for black MSM.}

\item{asmr.O..wa}{Vector of length 60 defining the age-specific
mortality rates for for other race/ethnicity MSM.}

\item{role.prob}{Vector of length 3 for the probability of sexual role as
insertive, receptive, and versatile.}

\item{num.H..wa}{Population size of Hispanic MSM.}

\item{num.B..wa}{Population size of non-Hispanic black MSM.}

\item{num.O..wa}{Population size of non-Hispanic other MSM.}

\item{num.KC}{Population size of MSM in other counties in western Washington.}

\item{num.OW}{Population size of MSM in other counties in western Washington.}

\item{num.EW}{Population size of MSM in other counties in western Washington.}
}
\description{
Calculates the target statistics for the formation and dissolution
             components of the network model to be estimated with \code{netest}. 
             For the WHAMP model, the base EpiModelHIV code was modified to define 
             three racial/ethnic groups (Hispanic, black, and other), age structure,
             add region as an attribute, and change the way risk group quantiles are 
             defined.
}
\details{
This function performs basic calculations to determine the components of the
formation and dissolution models for the network model estimation to be
conducted with \code{\link{netest}}. The inputs inputs for this function are
calculated externally to the package in a setup scenario file.
}
\seealso{
Network statistics calculated here are entered into \code{\link{base_nw_msm}}
to construct the base network, and then into the parameters in
\code{\link{param_msm}}.
}
\keyword{msm}
