% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{control.mard}
\alias{control.mard}
\title{Epidemic Model Control Settings for MARDHAM Models}
\usage{
control.mard(simno = 1, nsims = 1, ncores = 1, nsteps = 100,
  start = 1, resim.int = 1, initialize.FUN = initialize.mard,
  aging.FUN = aging.mard, deaths.FUN = deaths.mard,
  births.FUN = births.mard, test.FUN = test.mard, tx.FUN = tx.mard,
  prep.FUN = prep.mard, progress.FUN = progress.mard,
  vl.FUN = update_vl.mard, aiclass.FUN = update_aiclass.mard,
  roleclass.FUN = update_roleclass.mard, edgescorr.FUN = edges_correct.mard,
  resimnets.FUN = simnet.mard, disclose.FUN = disclose.mard,
  acts.FUN = acts.mard, condoms.FUN = condoms.mard,
  position.FUN = position.mard, trans.FUN = trans.mard,
  getprev.FUN = prevalence.mard, verbose.FUN = verbose.mard,
  delete.nodes = TRUE, save.dal = FALSE, save.nwstats = TRUE,
  save.network = FALSE, save.other = "attr", prevfull = FALSE,
  verbose = TRUE, verbose.int = 1, ...)
}
\arguments{
\item{simno}{Unique ID for the simulation run, used for file naming purposes
if used in conjunction with the \code{EpiModelHPC} package.}

\item{nsims}{Number of simulations.}

\item{ncores}{Number of cores per run, if parallelization is used within the
\code{EpiModelHPC} package.}

\item{nsteps}{Number of time steps per simulation.}

\item{start}{Starting time step for simulation, with default to 1 to run new
simulation. This may also be set to 1 greater than the final time
step of a previous simulation to resume the simulation with different
parameters.}

\item{resim.int}{Interval unit for resimulation of network, relative to the
base time unit in the model.}

\item{initialize.FUN}{Module function to use for initialization of the epidemic
model.}

\item{aging.FUN}{Module function for aging.}

\item{deaths.FUN}{Module function for general and disease-realted deaths.}

\item{births.FUN}{Module function for births or entries into the population.}

\item{test.FUN}{Module function for diagnostic disease testing.}

\item{tx.FUN}{Module function for ART initiation and adherence.}

\item{prep.FUN}{Module function for PrEP initiation and utilization.}

\item{progress.FUN}{Module function for HIV disease progression.}

\item{vl.FUN}{Module function for HIV viral load evolution.}

\item{aiclass.FUN}{Module function for one-off AI risk class transitions.}

\item{roleclass.FUN}{Module function for transitions in sexual roles.}

\item{edgescorr.FUN}{Module function for the edges coefficient adjustment
to preserve mean degree under varying population sizes.}

\item{resimnets.FUN}{Module function for network resimulation at each time
step.}

\item{disclose.FUN}{Module function for HIV status disclosure.}

\item{acts.FUN}{Module function to simulate the number of sexual acts within
partnerships.}

\item{condoms.FUN}{Module function to simulate condom use within acts.}

\item{position.FUN}{Module function to simulate sexual position within acts.}

\item{trans.FUN}{Module function to stochastically simulate disease transmission
over acts given individual and dyadic attributes.}

\item{getprev.FUN}{Module function to calculate prevalence summary statistics.}

\item{verbose.FUN}{Module function to print model progress to the console or
external text files.}

\item{delete.nodes}{If \code{TRUE}, dead nodes will be removed from the network
object and only active nodes will be retained.}

\item{save.dal}{If \code{TRUE}, the discordant act list will be saved at each time
step, otherwise it will be discarded.}

\item{save.nwstats}{If \code{TRUE}, the network statistics will be saved.}

\item{save.network}{If \code{TRUE}, the \code{network} objects will be saved
out at the end of simulation (necessary for restarting a simulation).}

\item{save.other}{Character vector containing other list elements of \code{dat}
to save.}

\item{prevfull}{If \code{TRUE}, save extended summary statistics for prevalence
and incidence (defined in the prevalence module).}

\item{verbose}{If \code{TRUE}, print out simulation progress to the console
if in interactive mode or text files if in batch mode.}

\item{verbose.int}{Integer specifying the interval between time steps at which
progress is printed.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list object of class \code{control.mard}, which can be passed to the
EpiModel function \code{netsim}.
}
\description{
Sets the controls for stochastic network models simulated with
             \code{\link{netsim}}.
}
