% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelHIV-package.R
\docType{package}
\name{EpiModelHIV-package}
\alias{EpiModelHIV-package}
\alias{EpiModelHIV}
\title{HIV Transmission Dynamics among MSM and Heterosexuals}
\description{
\tabular{ll}{
   Package: \tab EpiModelHIV\cr
   Type: \tab Package\cr
   Version: \tab 1.0.0\cr
   Date: \tab 2016-06-25\cr
   License: \tab GPL-3\cr
   LazyLoad: \tab yes\cr
}
}
\details{
EpiModelHIV provides extensions to our general EpiModel package to allow
for simulating HIV transmission dynamics among two populations: men who
have sex with men (MSM) in the United States and heterosexual adults in
sub-Saharan Africa.
}
\keyword{het}
\keyword{msm}
\keyword{package}
