% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{init_msm}
\alias{init_msm}
\title{Epidemic Model Initial Conditions}
\usage{
init_msm(nwstats, prev.B = 0.1, prev.W = 0.1, prev.ugc = 0.015,
  prev.rgc = 0.015, prev.uct = 0.015, prev.rct = 0.015,
  prev.syph.B = 0.015, prev.syph.W = 0.015, stage.syph.B.prob = c(0, 0.2,
  0.077, 0.277, 0.22, 0.22, 0.006), stage.syph.W.prob = c(0, 0.2, 0.077,
  0.277, 0.22, 0.22, 0.006), ...)
}
\arguments{
\item{nwstats}{Target statistics for the network model. An object of class
\code{nwstats} output from \code{\link{calc_nwstats_msm}}.}

\item{prev.B}{Initial disease prevalence among black MSM.}

\item{prev.W}{Initial disease prevalence among white MSM.}

\item{prev.ugc}{Initial prevalence of urethral gonorrhea.}

\item{prev.rgc}{Initial prevalence of rectal gonorrhea.}

\item{prev.uct}{Initial prevalence of urethral chlamydia.}

\item{prev.rct}{Initial prevalence of rectal chlamydia.}

\item{prev.syph.B}{Initial prevalence of syphilis among black MSM}

\item{prev.syph.W}{Initial prevalence of syphilis among white MSM}

\item{stage.syph.B.prob}{Proportion of black MSM who enter one of the seven
active stages of syphilis: incubating, primary, secondary, early
latent, late latent, late late latent, and tertiary.}

\item{stage.syph.W.prob}{Proportion of white MSM who enter one of the seven
active stages of syphilis: incubating, primary, secondary, early
latent, late latent, late late latent, and tertiary.}

\item{...}{Additional arguments passed to function.}
}
\value{
A list object of class \code{init_msm}, which can be passed to EpiModel
function \code{\link{netsim}}.
}
\description{
Sets the initial conditions for a stochastic epidemic models
             simulated with \code{\link{netsim}}.
}
\keyword{msm}
