% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.initialize.WHAMP.R
\name{init_status_msm_whamp}
\alias{init_status_msm_whamp}
\title{Initialize the HIV status of persons in the network for the WHAMP model}
\usage{
init_status_msm_whamp(dat)
}
\arguments{
\item{dat}{Data object created in initialization module.}
}
\description{
Sets the initial individual-level disease status of persons
             in the network, as well as disease-related attributes for
             infected persons. This function works by 1) assigning time since 
             infection as a draw from a uniform distribution with the max being
             the time since sexual debut, 2) assigning diagnosis based on time 
             since last test, which is sampled from a uniform distribution, or, 
             for non-screener types, time to symptomatic infection, 3) assigning treatment
             status, 4) assigning stage of infection based on time since infection and
             expected cumulative time on/off treatment, 5) assigning viral load  based on 
             stage of infection and treatment status
}
\keyword{(assigned}
\keyword{(time}
\keyword{*}
\keyword{+}
\keyword{-}
\keyword{/}
\keyword{2}
\keyword{7}
\keyword{AIDS}
\keyword{AIDS,}
\keyword{ART}
\keyword{Assume}
\keyword{Because}
\keyword{But}
\keyword{For}
\keyword{HIV-related}
\keyword{So}
\keyword{This}
\keyword{Time}
\keyword{To}
\keyword{VL}
\keyword{We}
\keyword{a}
\keyword{account}
\keyword{achieved}
\keyword{ago.}
\keyword{all}
\keyword{and}
\keyword{are}
\keyword{as}
\keyword{as:}
\keyword{assume}
\keyword{at}
\keyword{average}
\keyword{avg}
\keyword{back}
\keyword{balances}
\keyword{based}
\keyword{been}
\keyword{before}
\keyword{calculate}
\keyword{chronic}
\keyword{consecutive}
\keyword{constant}
\keyword{constrained}
\keyword{death}
\keyword{define}
\keyword{determining}
\keyword{diagnosis}
\keyword{disontinued,}
\keyword{doesn't}
\keyword{dx}
\keyword{end}
\keyword{exceed}
\keyword{fact}
\keyword{first}
\keyword{floor((time}
\keyword{for}
\keyword{from}
\keyword{go}
\keyword{have}
\keyword{hazard}
\keyword{if}
\keyword{implement}
\keyword{in}
\keyword{increased}
\keyword{infected.}
\keyword{infection}
\keyword{initiation}
\keyword{inititated}
\keyword{interval}
\keyword{is}
\keyword{it}
\keyword{last}
\keyword{level}
\keyword{levels.}
\keyword{log10.}
\keyword{long}
\keyword{max}
\keyword{may}
\keyword{men}
\keyword{modeled}
\keyword{mortality}
\keyword{msm}
\keyword{not}
\keyword{occurred}
\keyword{of}
\keyword{off}
\keyword{on}
\keyword{on-treatment}
\keyword{out}
\keyword{people}
\keyword{phase}
\keyword{phase,}
\keyword{point}
\keyword{point,}
\keyword{point.}
\keyword{pos}
\keyword{post}
\keyword{progression}
\keyword{rate}
\keyword{rather}
\keyword{recently}
\keyword{reinitiated}
\keyword{reinitiating}
\keyword{rises}
\keyword{same}
\keyword{set}
\keyword{since}
\keyword{slope}
\keyword{so}
\keyword{some}
\keyword{stopping}
\keyword{suppressed}
\keyword{takes}
\keyword{test}
\keyword{test.int)}
\keyword{test.window)}
\keyword{than}
\keyword{that}
\keyword{that,}
\keyword{the}
\keyword{their}
\keyword{them.}
\keyword{then}
\keyword{this,}
\keyword{those}
\keyword{time}
\keyword{time.off.tx}
\keyword{to}
\keyword{tx}
\keyword{tx,}
\keyword{untreated}
\keyword{up}
\keyword{upon}
\keyword{used}
\keyword{utility}
\keyword{variable}
\keyword{we}
\keyword{who}
\keyword{would}
\keyword{years.}
\keyword{yet}
