% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/params.R
\name{param.mard}
\alias{param.mard}
\title{Epidemic Model Parameters for MARDHAM Models}
\usage{
param.mard(nwstats, last.neg.test.time.range.B = round(301/7),
  mean.test.interv.B = round(301/7),
  last.neg.test.time.range.W = round(315/7),
  mean.test.interv.W = round(315/7), testing.pattern = "interval",
  test.window.period = 3, tt.traj.freq.B = c(0.077, 0, 0.356, 0.567),
  tt.traj.freq.W = c(0.052, 0, 0.331, 0.617), prob.tx.init.B = 0.092,
  prob.tx.init.W = 0.127, prob.tx.halt.B = 0.0102,
  prob.tx.halt.W = 0.0071, prob.tx.reinit.B = 0.00066,
  prob.tx.reinit.W = 0.00291, max.time.off.tx.full = 520,
  max.time.on.tx.part = 52 * 15, max.time.off.tx.part = 520,
  vl.acute.rise.dur = 3, vl.acute.peak = 6.886, vl.acute.fall.dur = 3,
  vl.set.point = 4.5, vl.aids.onset = 520, vl.aids.dur = 52 * 2,
  vl.fatal = 7, vl.full.supp = 1.5, vl.part.supp = 3.5,
  full.supp.down.slope = 0.25, full.supp.up.slope = 0.25,
  part.supp.down.slope = 0.25, part.supp.up.slope = 0.25,
  b.rate.B = 0.001, b.rate.W = 0.001, birth.age = 18,
  betabase.URAI = 0.0082 * 1.09, betabase.UIAI = 0.0031 * 1.09,
  betamult.acute = 4, betamult.circ = 0.4, betamult.condom = 0.25,
  disc.main.outset.B = (0.685 + 0.889)/2, disc.main.outset.W = (0.685 +
  0.889)/2, disc.main.at.diag.B = 1, disc.main.at.diag.W = 1,
  disc.main.post.diag.B = 0, disc.main.post.diag.W = 0,
  disc.pers.outset.B = (0.527 + 0.828)/2, disc.pers.outset.W = (0.527 +
  0.828)/2, disc.pers.at.diag.B = 1, disc.pers.at.diag.W = 1,
  disc.pers.post.diag.B = 0, disc.pers.post.diag.W = 0,
  disc.inst.B = (0.445 + 0.691)/2, disc.inst.W = (0.445 + 0.691)/2,
  circ.prev.B = 0.874, circ.prev.W = 0.918, ccr5.freq.B = c(0.0105,
  0.105), ccr5.freq.W = c(0.0105, 0.105), ccr5.heteroz.rr = 0.3,
  num.inst.ai.classes = 1, base.exp.ai.main.BB = 1.19,
  base.exp.ai.main.BW = 1.79, base.exp.ai.main.WW = 1.56,
  base.exp.ai.pers.BB = 0.75, base.exp.ai.pers.BW = 1.13,
  base.exp.ai.pers.WW = 0.98, incr.exp.ai.full.supp.main = 0,
  incr.exp.ai.part.supp.main = 0, redux.exp.ai.diag.main = 0,
  redux.exp.ai.discl.main = 0, incr.exp.ai.full.supp.pers = 0,
  incr.exp.ai.part.supp.pers = 0, redux.exp.ai.diag.pers = 0,
  redux.exp.ai.discl.pers = 0, cprob.main.BB = 0.38, cprob.main.BW = 0.1,
  cprob.main.WW = 0.15, cprob.pers.BB = 0.39, cprob.pers.BW = 0.11,
  cprob.pers.WW = 0.16, cprob.inst.BB = 0.49, cprob.inst.BW = 0.15,
  cprob.inst.WW = 0.22, beta.cond.fsupp.main = 0,
  beta.cond.psupp.main = 0, beta.cond.diag.main = -0.67,
  beta.cond.discl.main = -0.85, beta.cond.fsupp.pers = 0,
  beta.cond.psupp.pers = 0, beta.cond.diag.pers = -0.67,
  beta.cond.discl.pers = -0.85, beta.cond.fsupp.inst = 0,
  beta.cond.psupp.inst = 0, beta.cond.diag.inst = -0.67,
  beta.cond.discl.inst = -0.85, vv.prob.iev.BB = 0.42,
  vv.prob.iev.BW = 0.56, vv.prob.iev.WW = 0.49, ...)
}
\arguments{
\item{nwstats}{Target statistics for the network model. An object of class
\code{nwstats} output from \code{\link{calc_nwstats.mard}}.}

\item{last.neg.test.time.range.B}{Time range for last negative test for
black men.}

\item{mean.test.interv.B}{Mean intertest interval for black MSM who test.}

\item{last.neg.test.time.range.W}{Time range for last negative test for
white men.}

\item{mean.test.interv.W}{Mean intertest interval for white MSM who test}

\item{testing.pattern}{Method for HIV testing, with options \code{"memoryless"}
for constant hazard without regard to time since previous test, or
\code{"interval"} deterministic fixed intervals.}

\item{test.window.period}{Length of the HIV test window period.}

\item{tt.traj.freq.B}{Proportion of black MSM who enter one of four
testing/treatment trajectories: never test or treat, test and never
initiate treatment, test and treated with partial viral suppression,
and test and treated with full suppression.}

\item{tt.traj.freq.W}{Proportion of white MSM who enter into the four
testing/treatment trajectories, as defined above.}

\item{prob.tx.init.B}{Probability per time step that a black MSM who has
tested positive will initiate treatment.}

\item{prob.tx.init.W}{Probability per time step that a white MSM who has
tested positive will initiate treatment.}

\item{prob.tx.halt.B}{Probability per time step that a black MSM who is
currently on treatment will halt treatment.}

\item{prob.tx.halt.W}{Probability per time step that a white MSM who is
currently on treatment will halt treatment.}

\item{prob.tx.reinit.B}{Probability per time step that a black MSM who is
not currently on treatment but who has been in the past will
re-initiate treatment.}

\item{prob.tx.reinit.W}{Probability per time step that a white MSM who is
not currently on treatment but who has been in the past will
re-initiate treatment.}

\item{max.time.off.tx.full}{Number of time steps off treatment for a full
suppressor before onset of AIDS, including time before diagnosis.}

\item{max.time.on.tx.part}{Number of time steps on treatment for a
partial suppressor beofre onset of AIDS.}

\item{max.time.off.tx.part}{Nnumber of time steps off treatment for a
partial suppressor before onset of AIDS, including time before
diagnosis.}

\item{vl.acute.rise.dur}{Number of time steps to peak viremia during acute
infection.}

\item{vl.acute.peak}{Peak viral load (in log10 units) at the height of acute
infection.}

\item{vl.acute.fall.dur}{Number of time steps from peak viremia to set-point
viral load during the acute infection period.}

\item{vl.set.point}{Set point viral load (in log10 units).}

\item{vl.aids.onset}{Number of time steps to AIDS for a treatment-naive
patient.}

\item{vl.aids.dur}{Duration of AIDS stage infection.}

\item{vl.fatal}{Viral load in AIDS at which death occurs.}

\item{vl.full.supp}{Log10 viral load at full suppression on ART.}

\item{vl.part.supp}{Log10 viral load at partial suppression on ART.}

\item{full.supp.down.slope}{For full suppressors, number of log10 units that
viral load falls per time step from treatment initiation or re-initiation
until the level in \code{vl.full.supp}.}

\item{full.supp.up.slope}{For full suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected
value.}

\item{part.supp.down.slope}{For partial suppressors, number of log10 units
that viral load falls per time step from treatment initiation or
re-initiation until the level in \code{vl.part.supp}.}

\item{part.supp.up.slope}{For partial suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected value.}

\item{b.rate.B}{Rate at which black MSM enter the population.}

\item{b.rate.W}{Rate at which white MSM enter the population.}

\item{birth.age}{Age (in years) of new arrivals.}

\item{betabase.URAI}{Transmissibility for a man having unprotected receptive
anal intercourse with an infected man at set point viral load.}

\item{betabase.UIAI}{Transmissibility for an uncircumcised man having
unprotected insertive anal intercourse with an infected man at set
point viral load.}

\item{betamult.acute}{Multiplicative factor by which acute infection heightens
infectiousness, above that predicted by elevated viral load.}

\item{betamult.circ}{Factor by which to multiply infectivity from
insertive anal sex when the negative insertive partner is circumcised.}

\item{betamult.condom}{Factor by which to multiply infectivity for anal
sex when a condom is used.}

\item{disc.main.outset.B}{Probability that an HIV-infected black MSM will
disclose his status at the start of a main partnership.}

\item{disc.main.outset.W}{Probability that an HIV-infected white MSM will
disclose his status at the start of a main partnership.}

\item{disc.main.at.diag.B}{Probability that a black MSM already in a main
partnership will disclose at the time of diagnosis.}

\item{disc.main.at.diag.W}{Probability that a white MSM already in a main
partnership will disclose at the time of diagnosis.}

\item{disc.main.post.diag.B}{Probability that an HIV-infected black MSM
in a main partnership will disclose his status, assuming he didn't
at the start of the partnership or at diagnosis.}

\item{disc.main.post.diag.W}{Probability that an HIV-infected white MSM
in a main partnership will disclose his status, assuming he didn't
at the start of the partnership or at diagnosis.}

\item{disc.pers.outset.B}{Probability that an HIV-infected black MSM will
disclose his status at the start of a casual partnership.}

\item{disc.pers.outset.W}{Probability that an HIV-infected white MSM will
disclose his status at the start of a casual partnership.}

\item{disc.pers.at.diag.B}{Probability that a black MSM already in a
casual partnership will disclose at the time of diagnosis.}

\item{disc.pers.at.diag.W}{Probability that a white MSM already in a
casual partnership will disclose at the time of diagnosis.}

\item{disc.pers.post.diag.B}{Probability that an HIV-infected black MSM
in a casual partnership will disclose his status, assuming he
didn't at the start of the partnership or at diagnosis.}

\item{disc.pers.post.diag.W}{Probability that an HIV-infected white MSM
in a casual partnership will disclose his status, assuming he
didn't at the start of the partnership or at diagnosis.}

\item{disc.inst.B}{Probability that an HIV-infected black MSM will
disclose his status to a one-off partner.}

\item{disc.inst.W}{Probability that an HIV-infected white MSM will
disclose his status to a one-off partner.}

\item{circ.prev.B}{Probablity that a black new arrival in the population
will be circumcised.}

\item{circ.prev.W}{Probablity that a white new arrival in the population
will be circumcised.}

\item{ccr5.freq.B}{Vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respectively) in the CCR5 gene
among black MSM.}

\item{ccr5.freq.W}{Vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respectively) in the CCR5 gene
among white MSM.}

\item{ccr5.heteroz.rr}{Relative risk of infection for men who are heterozygous
in the CCR5 mutation.}

\item{num.inst.ai.classes}{Number of quantiles into which men should be
divided in determining their levels of one-off anal intercourse.}

\item{base.exp.ai.main.BB}{Expected coital frequency in black-black main
partnerships (acts per week).}

\item{base.exp.ai.main.BW}{Expected coital frequency in black-white main
partnerships (acts per week).}

\item{base.exp.ai.main.WW}{Expected coital frequency in white-white main
partnerships (acts per week).}

\item{base.exp.ai.pers.BB}{Expected coital frequency in black-black casual
partnerships (acts per week).}

\item{base.exp.ai.pers.BW}{Expected coital frequency in black-white casual
partnerships (acts per week).}

\item{base.exp.ai.pers.WW}{Expected coital frequency in white-white casual
partnerships (acts per week).}

\item{incr.exp.ai.full.supp.main}{Percent increase in expected coital
frequency in main partnerships after full suppression.}

\item{incr.exp.ai.part.supp.main}{Percent increase in expected coital
frequency in main partnerships after partial suppression.}

\item{redux.exp.ai.diag.main}{Percent reduction in expected coital frequency
in main partnerships after disease diagnosis.}

\item{redux.exp.ai.discl.main}{Percent reduction in expected coital frequency
in main partnerships after disclosure of diseasediagnosis.}

\item{incr.exp.ai.full.supp.pers}{Percent increase in expected coital
frequency in casual partnerships after full suppression.}

\item{incr.exp.ai.part.supp.pers}{Percent increase in expected coital
frequency in casual partnerships after partial suppression.}

\item{redux.exp.ai.diag.pers}{Percent reduction in expected coital frequency
in casual partnerships after diagnosis.}

\item{redux.exp.ai.discl.pers}{Percent reduction in expected coital frequency
in casual partnerships associated after disclosure of disease diagnosis.}

\item{cprob.main.BB}{Probability of condom use in a black-black main
partnership.}

\item{cprob.main.BW}{Probability of condom use in a black-white main
partnership.}

\item{cprob.main.WW}{Probability of condom use in a white-white main
partnership.}

\item{cprob.pers.BB}{Probability of condom use in a black-black casual
partnership.}

\item{cprob.pers.BW}{Probability of condom use in a black-white casual
partnership.}

\item{cprob.pers.WW}{Probability of condom use in a white-white casual
partnership.}

\item{cprob.inst.BB}{Probability of condom use in a black-black one-off
partnership.}

\item{cprob.inst.BW}{Probability of condom use in a black-white one-off
partnership.}

\item{cprob.inst.WW}{Probability of condom use in a white-white one-off
partnership.}

\item{beta.cond.fsupp.main}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected is fully suppressed.}

\item{beta.cond.psupp.main}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected is partially suppressed.}

\item{beta.cond.diag.main}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected man has been
diagnosed.}

\item{beta.cond.discl.main}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected man has disclosed.}

\item{beta.cond.fsupp.pers}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV-infected man is fully
suppressed.}

\item{beta.cond.psupp.pers}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV-infected man is partially
suppressed.}

\item{beta.cond.diag.pers}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV-infected man has been
diagnosed.}

\item{beta.cond.discl.pers}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV-infected man has disclosed
his status.}

\item{beta.cond.fsupp.inst}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV-infected man is fully
suppressed.}

\item{beta.cond.psupp.inst}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV-infected man is partially
suppressed.}

\item{beta.cond.diag.inst}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV-infected man has been
diagnosed.}

\item{beta.cond.discl.inst}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV-infected man has disclosed
his status.}

\item{vv.prob.iev.BB}{Probability that in a black-black partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{vv.prob.iev.BW}{Probability that in a black-white partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{vv.prob.iev.WW}{Probability that in a white-white partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list object of class \code{param.mard}, which can be passed to
EpiModel function \code{netsim}.
}
\description{
Sets the epidemic parameters for stochastic network models
             simulated with \code{\link{netsim}} for MARDHAM.
}

