% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/params.R
\name{param.mard}
\alias{param.mard}
\title{Epidemic Model Parameters for MARDHAM Models}
\usage{
param.mard(nwstats, last.neg.test.B.int = round(301/7),
  mean.test.B.int = round(301/7), last.neg.test.W.int = round(315/7),
  mean.test.W.int = round(315/7), testing.pattern = "interval",
  test.window.int = 3, tt.traj.B.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.W.prob = c(0.052, 0, 0.331, 0.617), tx.init.B.prob = 0.092,
  tx.init.W.prob = 0.127, tx.halt.B.prob = 0.0102,
  tx.halt.W.prob = 0.0071, tx.reinit.B.prob = 0.00066,
  tx.reinit.W.prob = 0.00291, max.time.off.tx.full.int = 520,
  max.time.on.tx.part.int = 52 * 15, max.time.off.tx.part.int = 520,
  vl.acute.rise.int = 3, vl.acute.peak = 6.886, vl.acute.fall.int = 3,
  vl.set.point = 4.5, vl.aids.onset.int = 520, vl.aids.int = 52 * 2,
  vl.fatal = 7, vl.full.supp = 1.5, vl.part.supp = 3.5,
  full.supp.down.slope = 0.25, full.supp.up.slope = 0.25,
  part.supp.down.slope = 0.25, part.supp.up.slope = 0.25,
  b.B.rate = 0.001, b.W.rate = 0.001, birth.age = 18, URAI.beta = 0.0082
  * 1.09, UIAI.beta = 0.0031 * 1.09, acute.beta = 4, circ.beta = 0.4,
  condom.beta = 0.25, disc.outset.main.B.prob = (0.685 + 0.889)/2,
  disc.outset.main.W.prob = (0.685 + 0.889)/2, disc.at.diag.main.B.prob = 1,
  disc.at.diag.main.W.prob = 1, disc.post.diag.main.B.prob = 0,
  disc.post.diag.main.W.prob = 0, disc.outset.pers.B.prob = (0.527 +
  0.828)/2, disc.outset.pers.W.prob = (0.527 + 0.828)/2,
  disc.at.diag.pers.B.prob = 1, disc.at.diag.pers.W.prob = 1,
  disc.post.diag.pers.B.prob = 0, disc.post.diag.pers.W.prob = 0,
  disc.inst.B.prob = (0.445 + 0.691)/2, disc.inst.W.prob = (0.445 +
  0.691)/2, circ.B.prob = 0.874, circ.W.prob = 0.918,
  ccr5.B.prob = c(0.0105, 0.105), ccr5.W.prob = c(0.0105, 0.105),
  ccr5.heteroz.rr = 0.3, num.inst.ai.classes = 1,
  base.ai.main.BB.rate = 1.19, base.ai.main.BW.rate = 1.79,
  base.ai.main.WW.rate = 1.56, base.ai.pers.BB.rate = 0.75,
  base.ai.pers.BW.rate = 1.13, base.ai.pers.WW.rate = 0.98,
  incr.ai.full.supp.main.rr = 0, incr.ai.part.supp.main.rr = 0,
  redux.ai.diag.main.rr = 0, redux.ai.discl.main.rr = 0,
  incr.ai.full.supp.pers.rr = 0, incr.ai.part.supp.pers.rr = 0,
  redux.ai.diag.pers.rr = 0, redux.ai.discl.pers.rr = 0,
  c.main.BB.prob = 0.38, c.main.BW.prob = 0.1, c.main.WW.prob = 0.15,
  c.pers.BB.prob = 0.39, c.pers.BW.prob = 0.11, c.pers.WW.prob = 0.16,
  c.inst.BB.prob = 0.49, c.inst.BW.prob = 0.15, c.inst.WW.prob = 0.22,
  cond.fsupp.main.beta = 0, cond.psupp.main.beta = 0,
  cond.diag.main.beta = -0.67, cond.discl.main.beta = -0.85,
  cond.fsupp.pers.beta = 0, cond.psupp.pers.beta = 0,
  cond.diag.pers.beta = -0.67, cond.discl.pers.beta = -0.85,
  cond.fsupp.inst.beta = 0, cond.psupp.inst.beta = 0,
  cond.diag.inst.beta = -0.67, cond.discl.inst.beta = -0.85,
  vv.iev.BB.prob = 0.42, vv.iev.BW.prob = 0.56, vv.iev.WW.prob = 0.49,
  ...)
}
\arguments{
\item{nwstats}{Target statistics for the network model. An object of class
\code{nwstats} output from \code{\link{calc_nwstats.mard}}.}

\item{last.neg.test.B.int}{Time range for last negative test for
black men.}

\item{mean.test.B.int}{Mean intertest interval for black MSM who test.}

\item{last.neg.test.W.int}{Time range for last negative test for
white men.}

\item{mean.test.W.int}{Mean intertest interval for white MSM who test}

\item{testing.pattern}{Method for HIV testing, with options \code{"memoryless"}
for constant hazard without regard to time since previous test, or
\code{"interval"} deterministic fixed intervals.}

\item{test.window.int}{Length of the HIV test window period.}

\item{tt.traj.B.prob}{Proportion of black MSM who enter one of four
testing/treatment trajectories: never test or treat, test and never
initiate treatment, test and treated with partial viral suppression,
and test and treated with full suppression.}

\item{tt.traj.W.prob}{Proportion of white MSM who enter into the four
testing/treatment trajectories, as defined above.}

\item{tx.init.B.prob}{Probability per time step that a black MSM who has
tested positive will initiate treatment.}

\item{tx.init.W.prob}{Probability per time step that a white MSM who has
tested positive will initiate treatment.}

\item{tx.halt.B.prob}{Probability per time step that a black MSM who is
currently on treatment will halt treatment.}

\item{tx.halt.W.prob}{Probability per time step that a white MSM who is
currently on treatment will halt treatment.}

\item{tx.reinit.B.prob}{Probability per time step that a black MSM who is
not currently on treatment but who has been in the past will
re-initiate treatment.}

\item{tx.reinit.W.prob}{Probability per time step that a white MSM who is
not currently on treatment but who has been in the past will
re-initiate treatment.}

\item{max.time.off.tx.full.int}{Number of time steps off treatment for a full
suppressor before onset of AIDS, including time before diagnosis.}

\item{max.time.on.tx.part.int}{Number of time steps on treatment for a
partial suppressor beofre onset of AIDS.}

\item{max.time.off.tx.part.int}{Nnumber of time steps off treatment for a
partial suppressor before onset of AIDS, including time before
diagnosis.}

\item{vl.acute.rise.int}{Number of time steps to peak viremia during acute
infection.}

\item{vl.acute.peak}{Peak viral load (in log10 units) at the height of acute
infection.}

\item{vl.acute.fall.int}{Number of time steps from peak viremia to set-point
viral load during the acute infection period.}

\item{vl.set.point}{Set point viral load (in log10 units).}

\item{vl.aids.onset.int}{Number of time steps to AIDS for a treatment-naive
patient.}

\item{vl.aids.int}{Duration of AIDS stage infection.}

\item{vl.fatal}{Viral load in AIDS at which death occurs.}

\item{vl.full.supp}{Log10 viral load at full suppression on ART.}

\item{vl.part.supp}{Log10 viral load at partial suppression on ART.}

\item{full.supp.down.slope}{For full suppressors, number of log10 units that
viral load falls per time step from treatment initiation or re-initiation
until the level in \code{vl.full.supp}.}

\item{full.supp.up.slope}{For full suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected
value.}

\item{part.supp.down.slope}{For partial suppressors, number of log10 units
that viral load falls per time step from treatment initiation or
re-initiation until the level in \code{vl.part.supp}.}

\item{part.supp.up.slope}{For partial suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected value.}

\item{b.B.rate}{Rate at which black MSM enter the population.}

\item{b.W.rate}{Rate at which white MSM enter the population.}

\item{birth.age}{Age (in years) of new arrivals.}

\item{URAI.beta}{Transmissibility for a man having unprotected receptive
anal intercourse with an infected man at set point viral load.}

\item{UIAI.beta}{Transmissibility for an uncircumcised man having
unprotected insertive anal intercourse with an infected man at set
point viral load.}

\item{acute.beta}{Multiplicative factor by which acute infection heightens
infectiousness, above that predicted by elevated viral load.}

\item{circ.beta}{Factor by which to multiply infectivity from
insertive anal sex when the negative insertive partner is circumcised.}

\item{condom.beta}{Factor by which to multiply infectivity for anal
sex when a condom is used.}

\item{disc.outset.main.B.prob}{Probability that an HIV-infected black MSM will
disclose his status at the start of a main partnership.}

\item{disc.outset.main.W.prob}{Probability that an HIV-infected white MSM will
disclose his status at the start of a main partnership.}

\item{disc.at.diag.main.B.prob}{Probability that a black MSM already in a main
partnership will disclose at the time of diagnosis.}

\item{disc.at.diag.main.W.prob}{Probability that a white MSM already in a main
partnership will disclose at the time of diagnosis.}

\item{disc.post.diag.main.B.prob}{Probability that an HIV-infected black MSM
in a main partnership will disclose his status, assuming he didn't
at the start of the partnership or at diagnosis.}

\item{disc.post.diag.main.W.prob}{Probability that an HIV-infected white MSM
in a main partnership will disclose his status, assuming he didn't
at the start of the partnership or at diagnosis.}

\item{disc.outset.pers.B.prob}{Probability that an HIV-infected black MSM will
disclose his status at the start of a casual partnership.}

\item{disc.outset.pers.W.prob}{Probability that an HIV-infected white MSM will
disclose his status at the start of a casual partnership.}

\item{disc.at.diag.pers.B.prob}{Probability that a black MSM already in a
casual partnership will disclose at the time of diagnosis.}

\item{disc.at.diag.pers.W.prob}{Probability that a white MSM already in a
casual partnership will disclose at the time of diagnosis.}

\item{disc.post.diag.pers.B.prob}{Probability that an HIV-infected black MSM
in a casual partnership will disclose his status, assuming he
didn't at the start of the partnership or at diagnosis.}

\item{disc.post.diag.pers.W.prob}{Probability that an HIV-infected white MSM
in a casual partnership will disclose his status, assuming he
didn't at the start of the partnership or at diagnosis.}

\item{disc.inst.B.prob}{Probability that an HIV-infected black MSM will
disclose his status to a one-off partner.}

\item{disc.inst.W.prob}{Probability that an HIV-infected white MSM will
disclose his status to a one-off partner.}

\item{circ.B.prob}{Probablity that a black new arrival in the population
will be circumcised.}

\item{circ.W.prob}{Probablity that a white new arrival in the population
will be circumcised.}

\item{ccr5.B.prob}{Vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respectively) in the CCR5 gene
among black MSM.}

\item{ccr5.W.prob}{Vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respectively) in the CCR5 gene
among white MSM.}

\item{ccr5.heteroz.rr}{Relative risk of infection for men who are heterozygous
in the CCR5 mutation.}

\item{num.inst.ai.classes}{Number of quantiles into which men should be
divided in determining their levels of one-off anal intercourse.}

\item{base.ai.main.BB.rate}{Expected coital frequency in black-black main
partnerships (acts per week).}

\item{base.ai.main.BW.rate}{Expected coital frequency in black-white main
partnerships (acts per week).}

\item{base.ai.main.WW.rate}{Expected coital frequency in white-white main
partnerships (acts per week).}

\item{base.ai.pers.BB.rate}{Expected coital frequency in black-black casual
partnerships (acts per week).}

\item{base.ai.pers.BW.rate}{Expected coital frequency in black-white casual
partnerships (acts per week).}

\item{base.ai.pers.WW.rate}{Expected coital frequency in white-white casual
partnerships (acts per week).}

\item{incr.ai.full.supp.main.rr}{Percent increase in expected coital
frequency in main partnerships after full suppression.}

\item{incr.ai.part.supp.main.rr}{Percent increase in expected coital
frequency in main partnerships after partial suppression.}

\item{redux.ai.diag.main.rr}{Percent reduction in expected coital frequency
in main partnerships after disease diagnosis.}

\item{redux.ai.discl.main.rr}{Percent reduction in expected coital frequency
in main partnerships after disclosure of diseasediagnosis.}

\item{incr.ai.full.supp.pers.rr}{Percent increase in expected coital
frequency in casual partnerships after full suppression.}

\item{incr.ai.part.supp.pers.rr}{Percent increase in expected coital
frequency in casual partnerships after partial suppression.}

\item{redux.ai.diag.pers.rr}{Percent reduction in expected coital frequency
in casual partnerships after diagnosis.}

\item{redux.ai.discl.pers.rr}{Percent reduction in expected coital frequency
in casual partnerships associated after disclosure of disease diagnosis.}

\item{c.main.BB.prob}{Probability of condom use in a black-black main
partnership.}

\item{c.main.BW.prob}{Probability of condom use in a black-white main
partnership.}

\item{c.main.WW.prob}{Probability of condom use in a white-white main
partnership.}

\item{c.pers.BB.prob}{Probability of condom use in a black-black casual
partnership.}

\item{c.pers.BW.prob}{Probability of condom use in a black-white casual
partnership.}

\item{c.pers.WW.prob}{Probability of condom use in a white-white casual
partnership.}

\item{c.inst.BB.prob}{Probability of condom use in a black-black one-off
partnership.}

\item{c.inst.BW.prob}{Probability of condom use in a black-white one-off
partnership.}

\item{c.inst.WW.prob}{Probability of condom use in a white-white one-off
partnership.}

\item{cond.fsupp.main.beta}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected is fully suppressed.}

\item{cond.psupp.main.beta}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected is partially suppressed.}

\item{cond.diag.main.beta}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.main.beta}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected man has disclosed.}

\item{cond.fsupp.pers.beta}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV-infected man is fully
suppressed.}

\item{cond.psupp.pers.beta}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV-infected man is partially
suppressed.}

\item{cond.diag.pers.beta}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.pers.beta}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV-infected man has disclosed
his status.}

\item{cond.fsupp.inst.beta}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV-infected man is fully
suppressed.}

\item{cond.psupp.inst.beta}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV-infected man is partially
suppressed.}

\item{cond.diag.inst.beta}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.inst.beta}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV-infected man has disclosed
his status.}

\item{vv.iev.BB.prob}{Probability that in a black-black partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{vv.iev.BW.prob}{Probability that in a black-white partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{vv.iev.WW.prob}{Probability that in a white-white partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list object of class \code{param.mard}, which can be passed to
EpiModel function \code{netsim}.
}
\description{
Sets the epidemic parameters for stochastic network models
             simulated with \code{\link{netsim}} for MARDHAM.
}

