% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/params.R
\name{param.mard}
\alias{param.mard}
\title{Epidemic Model Parameters for MARDHAM Models}
\usage{
param.mard(nwstats, last.neg.test.B.int = 301, mean.test.B.int = 301,
  last.neg.test.W.int = 315, mean.test.W.int = 315,
  testing.pattern = "interval", test.window.int = 21,
  tt.traj.B.prob = c(0.077, 0, 0.356, 0.567), tt.traj.W.prob = c(0.052, 0,
  0.331, 0.617), tx.init.B.prob = 0.092, tx.init.W.prob = 0.127,
  tx.halt.B.prob = 0.0102, tx.halt.W.prob = 0.0071,
  tx.reinit.B.prob = 0.00066, tx.reinit.W.prob = 0.00291,
  max.time.off.tx.full.int = 520 * 7, max.time.on.tx.part.int = 52 * 15 * 7,
  max.time.off.tx.part.int = 520 * 7, vl.acute.rise.int = 21,
  vl.acute.peak = 6.886, vl.acute.fall.int = 21, vl.set.point = 4.5,
  vl.aids.onset.int = 520 * 7, vl.aids.int = 52 * 2 * 7, vl.fatal = 7,
  vl.full.supp = 1.5, vl.part.supp = 3.5, full.supp.down.slope = 0.25,
  full.supp.up.slope = 0.25, part.supp.down.slope = 0.25,
  part.supp.up.slope = 0.25, b.B.rate = 0.001/7, b.W.rate = 0.001/7,
  birth.age = 18, b.method = "fixed", URAI.prob = 0.0082 * 1.09,
  UIAI.prob = 0.0031 * 1.09, acute.rr = 4, circ.rr = 0.4,
  condom.rr = 0.25, disc.outset.main.B.prob = 0.685,
  disc.outset.main.W.prob = 0.889, disc.at.diag.main.B.prob = 1,
  disc.at.diag.main.W.prob = 1, disc.post.diag.main.B.prob = 0,
  disc.post.diag.main.W.prob = 0, disc.outset.pers.B.prob = 0.527,
  disc.outset.pers.W.prob = 0.828, disc.at.diag.pers.B.prob = 1,
  disc.at.diag.pers.W.prob = 1, disc.post.diag.pers.B.prob = 0,
  disc.post.diag.pers.W.prob = 0, disc.inst.B.prob = 0.445,
  disc.inst.W.prob = 0.691, circ.B.prob = 0.874, circ.W.prob = 0.918,
  ccr5.B.prob = c(0, 0.034), ccr5.W.prob = c(0.021, 0.176),
  ccr5.heteroz.rr = 0.3, num.inst.ai.classes = 1,
  base.ai.main.BB.rate = 1.19/7, base.ai.main.BW.rate = 1.79/7,
  base.ai.main.WW.rate = 1.56/7, base.ai.pers.BB.rate = 0.75/7,
  base.ai.pers.BW.rate = 1.13/7, base.ai.pers.WW.rate = 0.98/7,
  ai.full.supp.main.rr = 1, ai.part.supp.main.rr = 1, ai.diag.main.rr = 1,
  ai.discl.main.rr = 1, ai.full.supp.pers.rr = 1,
  ai.part.supp.pers.rr = 1, ai.diag.pers.rr = 1, ai.discl.pers.rr = 1,
  cond.main.BB.prob = 0.38, cond.main.BW.prob = 0.1,
  cond.main.WW.prob = 0.15, cond.pers.BB.prob = 0.39,
  cond.pers.BW.prob = 0.11, cond.pers.WW.prob = 0.16,
  cond.inst.BB.prob = 0.49, cond.inst.BW.prob = 0.15,
  cond.inst.WW.prob = 0.22, cond.fsupp.main.beta = 0,
  cond.psupp.main.beta = 0, cond.diag.main.beta = -0.67,
  cond.discl.main.beta = -0.85, cond.fsupp.pers.beta = 0,
  cond.psupp.pers.beta = 0, cond.diag.pers.beta = -0.67,
  cond.discl.pers.beta = -0.85, cond.fsupp.inst.beta = 0,
  cond.psupp.inst.beta = 0, cond.diag.inst.beta = -0.67,
  cond.discl.inst.beta = -0.85, vv.iev.BB.prob = 0.42,
  vv.iev.BW.prob = 0.56, vv.iev.WW.prob = 0.49, prep.start = 1,
  prep.elig.model = "all", prep.efficacy = 0.92, prep.class.prob = c(0.5,
  0.25, 0.25), prep.class.effect = c(0, 0.75, 0.9), prep.coverage = 0,
  prep.cov.method = "curr", prep.cov.rate = 1, prep.rcomp = 1, ...)
}
\arguments{
\item{nwstats}{Target statistics for the network model. An object of class
\code{nwstats} output from \code{\link{calc_nwstats.mard}}.}

\item{last.neg.test.B.int}{Time range in days for last negative test for
black men.}

\item{mean.test.B.int}{Mean intertest interval in days for black MSM who test.}

\item{last.neg.test.W.int}{Time range in days for last negative test for
white men.}

\item{mean.test.W.int}{Mean intertest interval in days for white MSM who test.}

\item{testing.pattern}{Method for HIV testing, with options \code{"memoryless"}
for constant hazard without regard to time since previous test, or
\code{"interval"} deterministic fixed intervals.}

\item{test.window.int}{Length of the HIV test window period in days.}

\item{tt.traj.B.prob}{Proportion of black MSM who enter one of four
testing/treatment trajectories: never test or treat, test and never
initiate treatment, test and treated with partial viral suppression,
and test and treated with full suppression.}

\item{tt.traj.W.prob}{Proportion of white MSM who enter into the four
testing/treatment trajectories, as defined above.}

\item{tx.init.B.prob}{Probability per time step that a black MSM who has
tested positive will initiate treatment.}

\item{tx.init.W.prob}{Probability per time step that a white MSM who has
tested positive will initiate treatment.}

\item{tx.halt.B.prob}{Probability per time step that a black MSM who is
currently on treatment will halt treatment.}

\item{tx.halt.W.prob}{Probability per time step that a white MSM who is
currently on treatment will halt treatment.}

\item{tx.reinit.B.prob}{Probability per time step that a black MSM who is
not currently on treatment but who has been in the past will
re-initiate treatment.}

\item{tx.reinit.W.prob}{Probability per time step that a white MSM who is
not currently on treatment but who has been in the past will
re-initiate treatment.}

\item{max.time.off.tx.full.int}{Number of days off treatment for a full
suppressor before onset of AIDS, including time before diagnosis.}

\item{max.time.on.tx.part.int}{Number of days on treatment for a
partial suppressor beofre onset of AIDS.}

\item{max.time.off.tx.part.int}{Nnumber of days off treatment for a
partial suppressor before onset of AIDS, including time before
diagnosis.}

\item{vl.acute.rise.int}{Number of days to peak viremia during acute
infection.}

\item{vl.acute.peak}{Peak viral load (in log10 units) at the height of acute
infection.}

\item{vl.acute.fall.int}{Number of days from peak viremia to set-point
viral load during the acute infection period.}

\item{vl.set.point}{Set point viral load (in log10 units).}

\item{vl.aids.onset.int}{Number of days to AIDS for a treatment-naive
patient.}

\item{vl.aids.int}{Duration of AIDS stage infection in days.}

\item{vl.fatal}{Viral load in AIDS at which death occurs.}

\item{vl.full.supp}{Log10 viral load at full suppression on ART.}

\item{vl.part.supp}{Log10 viral load at partial suppression on ART.}

\item{full.supp.down.slope}{For full suppressors, number of log10 units that
viral load falls per time step from treatment initiation or re-initiation
until the level in \code{vl.full.supp}.}

\item{full.supp.up.slope}{For full suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected
value.}

\item{part.supp.down.slope}{For partial suppressors, number of log10 units
that viral load falls per time step from treatment initiation or
re-initiation until the level in \code{vl.part.supp}.}

\item{part.supp.up.slope}{For partial suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected value.}

\item{b.B.rate}{Rate at which black MSM enter the population.}

\item{b.W.rate}{Rate at which white MSM enter the population.}

\item{birth.age}{Age (in years) of new arrivals.}

\item{b.method}{Method for calculating the number of expected births at each
time step, with \code{"fixed"} based on the number of persons at the
initial time step and \code{"varying"} based on the current time step.
The former method is compatible with \code{mardham1}.}

\item{URAI.prob}{Probability of transmission for a man having unprotected
receptive anal intercourse with an infected man at set point viral
load.}

\item{UIAI.prob}{Probability of transmission for an uncircumcised man having
unprotected insertive anal intercourse with an infected man at set
point viral load.}

\item{acute.rr}{Relative risk of infection (compared to that predicted by
elevated viral load) when positive partner is in the acute stage.}

\item{circ.rr}{Relative risk of infection from insertive anal sex when the
negative insertive partner is circumcised.}

\item{condom.rr}{Relative risk of infection from anal sex when a condom is
used.}

\item{disc.outset.main.B.prob}{Probability that an HIV-infected black MSM will
disclose his status at the start of a main partnership.}

\item{disc.outset.main.W.prob}{Probability that an HIV-infected white MSM will
disclose his status at the start of a main partnership.}

\item{disc.at.diag.main.B.prob}{Probability that a black MSM already in a main
partnership will disclose at the time of diagnosis.}

\item{disc.at.diag.main.W.prob}{Probability that a white MSM already in a main
partnership will disclose at the time of diagnosis.}

\item{disc.post.diag.main.B.prob}{Probability that an HIV-infected black MSM
in a main partnership will disclose his status, assuming he didn't
at the start of the partnership or at diagnosis.}

\item{disc.post.diag.main.W.prob}{Probability that an HIV-infected white MSM
in a main partnership will disclose his status, assuming he didn't
at the start of the partnership or at diagnosis.}

\item{disc.outset.pers.B.prob}{Probability that an HIV-infected black MSM will
disclose his status at the start of a casual partnership.}

\item{disc.outset.pers.W.prob}{Probability that an HIV-infected white MSM will
disclose his status at the start of a casual partnership.}

\item{disc.at.diag.pers.B.prob}{Probability that a black MSM already in a
casual partnership will disclose at the time of diagnosis.}

\item{disc.at.diag.pers.W.prob}{Probability that a white MSM already in a
casual partnership will disclose at the time of diagnosis.}

\item{disc.post.diag.pers.B.prob}{Probability that an HIV-infected black MSM
in a casual partnership will disclose his status, assuming he
didn't at the start of the partnership or at diagnosis.}

\item{disc.post.diag.pers.W.prob}{Probability that an HIV-infected white MSM
in a casual partnership will disclose his status, assuming he
didn't at the start of the partnership or at diagnosis.}

\item{disc.inst.B.prob}{Probability that an HIV-infected black MSM will
disclose his status to a one-off partner.}

\item{disc.inst.W.prob}{Probability that an HIV-infected white MSM will
disclose his status to a one-off partner.}

\item{circ.B.prob}{Probablity that a black new arrival in the population
will be circumcised.}

\item{circ.W.prob}{Probablity that a white new arrival in the population
will be circumcised.}

\item{ccr5.B.prob}{Vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respectively) in the CCR5 gene
among black MSM.}

\item{ccr5.W.prob}{Vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respectively) in the CCR5 gene
among white MSM.}

\item{ccr5.heteroz.rr}{Relative risk of infection for men who are heterozygous
in the CCR5 mutation.}

\item{num.inst.ai.classes}{Number of quantiles into which men should be
divided in determining their levels of one-off anal intercourse.}

\item{base.ai.main.BB.rate}{Expected coital frequency in black-black main
partnerships (acts per day).}

\item{base.ai.main.BW.rate}{Expected coital frequency in black-white main
partnerships (acts per day).}

\item{base.ai.main.WW.rate}{Expected coital frequency in white-white main
partnerships (acts per day).}

\item{base.ai.pers.BB.rate}{Expected coital frequency in black-black casual
partnerships (acts per day).}

\item{base.ai.pers.BW.rate}{Expected coital frequency in black-white casual
partnerships (acts per day).}

\item{base.ai.pers.WW.rate}{Expected coital frequency in white-white casual
partnerships (acts per day).}

\item{ai.full.supp.main.rr}{Relative expected coital frequency in main
partnerships after full suppression (value of 1 indicates no change).}

\item{ai.part.supp.main.rr}{Relative expected coital frequency in main
partnerships after partial suppression (value of 1 indicates no
change).}

\item{ai.diag.main.rr}{Relative expected coital frequency in main
partnerships after disease diagnosis (value of 1 indicates no change).}

\item{ai.discl.main.rr}{Relative expected coital frequency in main
partnerships after disclosure of disease diagnosis (value of 1
indicates no change).}

\item{ai.full.supp.pers.rr}{Relative expected coital frequency in casual
partnerships after full suppression (value of 1 indicates no change).}

\item{ai.part.supp.pers.rr}{Relative expected coital frequency in casual
partnerships after partial suppression (value of 1 indicates no
change).}

\item{ai.diag.pers.rr}{Relative expected coital frequency in casual
partnerships after diagnosis (value of 1 indicates no change).}

\item{ai.discl.pers.rr}{Relative expected coital frequency in casual
partnerships associated after disclosure of disease diagnosis (value
of 1 indicates no change).}

\item{cond.main.BB.prob}{Probability of condom use in a black-black main
partnership.}

\item{cond.main.BW.prob}{Probability of condom use in a black-white main
partnership.}

\item{cond.main.WW.prob}{Probability of condom use in a white-white main
partnership.}

\item{cond.pers.BB.prob}{Probability of condom use in a black-black casual
partnership.}

\item{cond.pers.BW.prob}{Probability of condom use in a black-white casual
partnership.}

\item{cond.pers.WW.prob}{Probability of condom use in a white-white casual
partnership.}

\item{cond.inst.BB.prob}{Probability of condom use in a black-black one-off
partnership.}

\item{cond.inst.BW.prob}{Probability of condom use in a black-white one-off
partnership.}

\item{cond.inst.WW.prob}{Probability of condom use in a white-white one-off
partnership.}

\item{cond.fsupp.main.beta}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected is fully suppressed.}

\item{cond.psupp.main.beta}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected is partially suppressed.}

\item{cond.diag.main.beta}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.main.beta}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected man has disclosed.}

\item{cond.fsupp.pers.beta}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV-infected man is fully
suppressed.}

\item{cond.psupp.pers.beta}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV-infected man is partially
suppressed.}

\item{cond.diag.pers.beta}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.pers.beta}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV-infected man has disclosed
his status.}

\item{cond.fsupp.inst.beta}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV-infected man is fully
suppressed.}

\item{cond.psupp.inst.beta}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV-infected man is partially
suppressed.}

\item{cond.diag.inst.beta}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.inst.beta}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV-infected man has disclosed
his status.}

\item{vv.iev.BB.prob}{Probability that in a black-black partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{vv.iev.BW.prob}{Probability that in a black-white partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{vv.iev.WW.prob}{Probability that in a white-white partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{prep.start}{Time step at which the PrEP intervention should start.}

\item{prep.elig.model}{Modeling approach for determining who is eligible for
PrEP. Current options are limited to: \code{"all"} for all persons who
have never been on PrEP and are disease-susceptible.}

\item{prep.efficacy}{The per-contact efficacy of PrEP to prevent infection if
used (parameter not currently used).}

\item{prep.class.prob}{The frequency of being in a low, medium, or high class
of adherence to PrEP.}

\item{prep.class.effect}{The functional effectiveness of PrEP conditional on
PrEP class.}

\item{prep.coverage}{The proportion of the eligible population who are start
PrEP once they become eligible.}

\item{prep.cov.method}{The method for calculating PrEP coverage, with options
of \code{"curr"} to base the numerator on the number of people currently
on PrEP and \code{"ever"} to base it on the number of people ever on
PrEP.}

\item{prep.cov.rate}{The rate at which persons initiate PrEP conditional on
their eligibility, with 1 equal to instant start.}

\item{prep.rcomp}{The relative change in rate of UAI across all partnership
types given current PrEP use, where 1 is no risk compensation.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list object of class \code{param.mard}, which can be passed to
EpiModel function \code{netsim}.
}
\description{
Sets the epidemic parameters for stochastic network models
             simulated with \code{\link{netsim}} for MARDHAM.
}

