% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_msm}
\alias{param_msm}
\title{Epidemic Model Parameters}
\usage{
param_msm(nwstats, race.method = 1, last.neg.test.B.int = 301,
  last.neg.test.W.int = 315, mean.test.B.int = 301, mean.test.W.int = 315,
  testing.pattern = "memoryless", testing.pattern.sti = "memoryless",
  test.window.int = 21, tt.traj.B.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.W.prob = c(0.052, 0, 0.331, 0.617), stage.syph.B.prob = c(0, 0.2,
  0.077, 0.277, 0.22, 0.22, 0.006), stage.syph.W.prob = c(0, 0.2, 0.077,
  0.277, 0.22, 0.22, 0.006), tx.init.B.prob = 0.092, tx.init.W.prob = 0.127,
  tx.halt.B.prob = 0.0102, tx.halt.W.prob = 0.0071,
  tx.reinit.B.prob = 0.00066, tx.reinit.W.prob = 0.00291,
  max.time.off.tx.full.int = 520 * 7, max.time.on.tx.part.int = 52 * 15 * 7,
  max.time.off.tx.part.int = 520 * 7, vl.acute.rise.int = 45,
  vl.acute.peak = 6.886, vl.acute.fall.int = 45, vl.set.point = 4.5,
  vl.aids.onset.int = 520 * 7, vl.aids.int = 52 * 2 * 7, vl.fatal = 7,
  vl.full.supp = 1.5, vl.part.supp = 3.5, full.supp.down.slope = 0.25,
  full.supp.up.slope = 0.25, part.supp.down.slope = 0.25,
  part.supp.up.slope = 0.25, incu.syph.int = 27, prim.syph.int = 60,
  seco.syph.int = 120, earlat.syph.int = 365 - 27 - 60 - 120,
  latelat.syph.int = 9 * 52 * 7, latelatelat.syph.int = 20 * 52 * 7,
  tert.syph.int = 20 * 52 * 7, syph.tert.prog.prob = 0.15/(52 * 7 * 20),
  b.B.rate = 0.001/7, b.W.rate = 0.001/7, birth.age = 18,
  b.method = "fixed", URAI.prob = 0.0082 * 1.09, UIAI.prob = 0.0031 *
  1.09, acute.rr = 6, circ.rr = 0.4, condom.rr = 0.295,
  disc.outset.main.B.prob = 0.685, disc.outset.main.W.prob = 0.889,
  disc.at.diag.main.B.prob = 1, disc.at.diag.main.W.prob = 1,
  disc.post.diag.main.B.prob = 0, disc.post.diag.main.W.prob = 0,
  disc.outset.pers.B.prob = 0.527, disc.outset.pers.W.prob = 0.828,
  disc.at.diag.pers.B.prob = 1, disc.at.diag.pers.W.prob = 1,
  disc.post.diag.pers.B.prob = 0, disc.post.diag.pers.W.prob = 0,
  disc.inst.B.prob = 0.445, disc.inst.W.prob = 0.691, circ.B.prob = 0.874,
  circ.W.prob = 0.918, ccr5.B.prob = c(0, 0.034), ccr5.W.prob = c(0.021,
  0.176), ccr5.heteroz.rr = 0.3, num.inst.ai.classes = 1,
  base.ai.main.BB.rate = 0.17, base.ai.main.BW.rate = 0.26,
  base.ai.main.WW.rate = 0.23, base.ai.pers.BB.rate = 0.11,
  base.ai.pers.BW.rate = 0.16, base.ai.pers.WW.rate = 0.14,
  ai.scale = 1.12, cond.main.BB.prob = 0.38, cond.main.BW.prob = 0.1,
  cond.main.WW.prob = 0.15, cond.pers.always.prob = 0.216,
  cond.pers.BB.prob = 0.26, cond.pers.BW.prob = 0.26,
  cond.pers.WW.prob = 0.26, cond.inst.always.prob = 0.326,
  cond.inst.BB.prob = 0.27, cond.inst.BW.prob = 0.27,
  cond.inst.WW.prob = 0.27, cond.always.prob.corr = 0.5, cond.rr.BB = 1,
  cond.rr.BW = 1, cond.rr.WW = 1, cond.diag.main.beta = -0.67,
  cond.discl.main.beta = -0.85, cond.diag.pers.beta = -0.67,
  cond.discl.pers.beta = -0.85, cond.diag.inst.beta = -0.67,
  cond.discl.inst.beta = -0.85, vv.iev.BB.prob = 0.42,
  vv.iev.BW.prob = 0.56, vv.iev.WW.prob = 0.49, prep.start = 2601,
  prep.elig.model = "base", prep.class.prob = c(0.211, 0.07, 0.1, 0.619),
  prep.class.hr = c(1, 0.69, 0.19, 0.05), prep.coverage = 0,
  prep.cov.method = "curr", prep.cov.rate = 1, prep.tst.int = 90,
  prep.risk.int = 182, prep.risk.reassess = TRUE,
  stitest.active.int = 364, stitest.start = 5000, ept.start = Inf,
  ept.risk.int = 60, ept.coverage = 0, ept.cov.method = "curr",
  ept.cov.rate = 1, rcomp.prob = 0, rcomp.adh.groups = 0:3,
  rcomp.main.only = FALSE, rcomp.discl.only = FALSE, rgc.tprob = 0.39,
  ugc.tprob = 0.29, rct.tprob = 0.19, uct.tprob = 0.14,
  usyph.tprob = 0.04883012, rsyph.tprob = 0.03829557,
  syph.earlat.rr = 0.5, syph.late.rr = 0, rgc.sympt.prob = 0.16,
  ugc.sympt.prob = 0.9, rct.sympt.prob = 0.14, uct.sympt.prob = 0.58,
  syph.prim.sympt.prob = 0.5, syph.seco.sympt.prob = 0.85,
  syph.earlat.sympt.prob = 0, syph.latelat.sympt.prob = 0,
  syph.tert.sympt.prob = 0.9, rgc.asympt.int = 35.11851 * 7,
  ugc.asympt.int = 35.11851 * 7, gc.tx.int = 2 * 7, gc.ntx.int = NA,
  rct.asympt.int = 44.24538 * 7, uct.asympt.int = 44.24538 * 7,
  ct.tx.int = 2 * 7, ct.ntx.int = NA, syph.early.tx.int = 7,
  syph.late.tx.int = 15, gc.prob.cease = 0, ct.prob.cease = 0,
  syph.prob.cease = 0, gc.sympt.prob.tx = 0.9, ct.sympt.prob.tx = 0.85,
  gc.asympt.prob.tx = 0.9, ct.asympt.prob.tx = 0.85,
  syph.prim.sympt.prob.tx = 0.35, syph.prim.asympt.prob.tx = 0,
  syph.seco.sympt.prob.tx = 0.6, syph.seco.asympt.prob.tx = 0,
  syph.earlat.prob.tx = 0.15, syph.latelat.prob.tx = 0.1,
  syph.tert.sympt.prob.tx = 0.9, syph.tert.asympt.prob.tx = 0,
  tst.rect.sti.rr = 1, sti.highrisktest.int = 182,
  stianntest.coverage = 1, stianntest.cov.method = "curr",
  stianntest.cov.rate = 1, stihighrisktest.coverage = 1,
  stihighrisktest.cov.method = "curr", stihighrisktest.cov.rate = 1,
  prep.sti.screen.int = 182, prep.sti.prob.tx = 1,
  prep.continue.stand.tx = TRUE, sti.cond.rr = 0.3,
  hiv.rgc.rr = 2.64869327, hiv.ugc.rr = 1.67105474,
  hiv.rct.rr = 2.64869327, hiv.uct.rr = 1.67105474, hiv.dual.rr = 0.2,
  hiv.rsyph.rr = 2.7, hiv.usyph.rr = 1.7, syph.rhiv.rr = 2.6,
  syph.uhiv.rr = 1.7, hivdx.syph.sympt.tx.rr = 2,
  partlist.start = stitest.start - sti.highrisktest.int - 1, ...)
}
\arguments{
\item{nwstats}{Target statistics for the network model. An object of class
\code{nwstats} output from \code{\link{calc_nwstats_msm}}.}

\item{race.method}{Number of races in the model, with options of 1 or 2. If
1, then race-specific parameters will be averaged.}

\item{last.neg.test.B.int}{Time range in days for last negative test for
black men.}

\item{last.neg.test.W.int}{Time range in days for last negative test for
white men.}

\item{mean.test.B.int}{Mean intertest interval in days for black MSM who test.}

\item{mean.test.W.int}{Mean intertest interval in days for white MSM who test.}

\item{testing.pattern}{Method for HIV testing, with options \code{"memoryless"}
for constant hazard without regard to time since previous test, or
\code{"interval"} deterministic fixed intervals.}

\item{testing.pattern.sti}{Method for STI testing, with options 
\code{"memoryless"} for constant hazard without regard to time since    
previous test, or \code{"interval"} deterministic fixed intervals.}

\item{test.window.int}{Length of the HIV test window period in days.}

\item{tt.traj.B.prob}{Proportion of black MSM who enter one of four
testing/treatment trajectories: never test or treat, test and never
initiate treatment, test and treated with partial viral suppression,
and test and treated with full suppression.}

\item{tt.traj.W.prob}{Proportion of white MSM who enter into the four
testing/treatment trajectories, as defined above.}

\item{stage.syph.B.prob}{Proportion of black MSM who enter one of the seven
active stages of syphilis: incubating, primary, secondary, early 
latent, late latent, late late latent, and tertiary.}

\item{stage.syph.W.prob}{Proportion of white MSM who enter each of the seven
active stages of syphilis, as defined above.}

\item{tx.init.B.prob}{Probability per time step that a black MSM who has
tested positive will initiate treatment.}

\item{tx.init.W.prob}{Probability per time step that a white MSM who has
tested positive will initiate treatment.}

\item{tx.halt.B.prob}{Probability per time step that a black MSM who is
currently on treatment will halt treatment.}

\item{tx.halt.W.prob}{Probability per time step that a white MSM who is
currently on treatment will halt treatment.}

\item{tx.reinit.B.prob}{Probability per time step that a black MSM who is
not currently on treatment but who has been in the past will
re-initiate treatment.}

\item{tx.reinit.W.prob}{Probability per time step that a white MSM who is
not currently on treatment but who has been in the past will
re-initiate treatment.}

\item{max.time.off.tx.full.int}{Number of days off treatment for a full
suppressor before onset of AIDS, including time before diagnosis.}

\item{max.time.on.tx.part.int}{Number of days on treatment for a
partial suppressor beofre onset of AIDS.}

\item{max.time.off.tx.part.int}{Nnumber of days off treatment for a
partial suppressor before onset of AIDS, including time before
diagnosis.}

\item{vl.acute.rise.int}{Number of days to peak viremia during acute
infection.}

\item{vl.acute.peak}{Peak viral load (in log10 units) at the height of acute
infection.}

\item{vl.acute.fall.int}{Number of days from peak viremia to set-point
viral load during the acute infection period.}

\item{vl.set.point}{Set point viral load (in log10 units).}

\item{vl.aids.onset.int}{Number of days to AIDS for a treatment-naive
patient.}

\item{vl.aids.int}{Duration of AIDS stage infection in days.}

\item{vl.fatal}{Viral load in AIDS at which death occurs.}

\item{vl.full.supp}{Log10 viral load at full suppression on ART.}

\item{vl.part.supp}{Log10 viral load at partial suppression on ART.}

\item{full.supp.down.slope}{For full suppressors, number of log10 units that
viral load falls per time step from treatment initiation or re-initiation
until the level in \code{vl.full.supp}.}

\item{full.supp.up.slope}{For full suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected
value.}

\item{part.supp.down.slope}{For partial suppressors, number of log10 units
that viral load falls per time step from treatment initiation or
re-initiation until the level in \code{vl.part.supp}.}

\item{part.supp.up.slope}{For partial suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected value.}

\item{incu.syph.int}{Number of days in incubation stage of syphilis.}

\item{prim.syph.int}{Number of days in primary stage of syphilis.}

\item{seco.syph.int}{Number of days in secondary stage of syphilis.}

\item{earlat.syph.int}{= Number of days in early latent stage of syphilis.}

\item{latelat.syph.int}{Number of days in first late latent stage of syphilis.}

\item{latelatelat.syph.int}{Number of days in second stage of late latent 
syphilis for those who will not progress to tertiary infection.}

\item{tert.syph.int}{Number of days in tertiary stage of syphilis.}

\item{syph.tert.prog.prob}{Probability of progression from late latent stage of
syphilis to tertiary stage at any point in time during late latent.}

\item{b.B.rate}{Rate at which black MSM enter the population.}

\item{b.W.rate}{Rate at which white MSM enter the population.}

\item{birth.age}{Age (in years) of new arrivals.}

\item{b.method}{Method for calculating the number of expected births at each
time step, with \code{"fixed"} based on the number of persons at the
initial time step and \code{"varying"} based on the current time step.}

\item{URAI.prob}{Probability of transmission for a man having unprotected
receptive anal intercourse with an infected man at set point viral
load.}

\item{UIAI.prob}{Probability of transmission for an uncircumcised man having
unprotected insertive anal intercourse with an infected man at set
point viral load.}

\item{acute.rr}{Relative risk of infection (compared to that predicted by
elevated viral load) when positive partner is in the acute stage.}

\item{circ.rr}{Relative risk of infection from insertive anal sex when the
negative insertive partner is circumcised.}

\item{condom.rr}{Relative risk of infection from anal sex when a condom is
used.}

\item{disc.outset.main.B.prob}{Probability that an HIV-infected black MSM will
disclose his status at the start of a main partnership.}

\item{disc.outset.main.W.prob}{Probability that an HIV-infected white MSM will
disclose his status at the start of a main partnership.}

\item{disc.at.diag.main.B.prob}{Probability that a black MSM already in a main
partnership will disclose at the time of diagnosis.}

\item{disc.at.diag.main.W.prob}{Probability that a white MSM already in a main
partnership will disclose at the time of diagnosis.}

\item{disc.post.diag.main.B.prob}{Probability that an HIV-infected black MSM
in a main partnership will disclose his status, assuming he didn't
at the start of the partnership or at diagnosis.}

\item{disc.post.diag.main.W.prob}{Probability that an HIV-infected white MSM
in a main partnership will disclose his status, assuming he didn't
at the start of the partnership or at diagnosis.}

\item{disc.outset.pers.B.prob}{Probability that an HIV-infected black MSM will
disclose his status at the start of a casual partnership.}

\item{disc.outset.pers.W.prob}{Probability that an HIV-infected white MSM will
disclose his status at the start of a casual partnership.}

\item{disc.at.diag.pers.B.prob}{Probability that a black MSM already in a
casual partnership will disclose at the time of diagnosis.}

\item{disc.at.diag.pers.W.prob}{Probability that a white MSM already in a
casual partnership will disclose at the time of diagnosis.}

\item{disc.post.diag.pers.B.prob}{Probability that an HIV-infected black MSM
in a casual partnership will disclose his status, assuming he
didn't at the start of the partnership or at diagnosis.}

\item{disc.post.diag.pers.W.prob}{Probability that an HIV-infected white MSM
in a casual partnership will disclose his status, assuming he
didn't at the start of the partnership or at diagnosis.}

\item{disc.inst.B.prob}{Probability that an HIV-infected black MSM will
disclose his status to a one-off partner.}

\item{disc.inst.W.prob}{Probability that an HIV-infected white MSM will
disclose his status to a one-off partner.}

\item{circ.B.prob}{Probablity that a black new arrival in the population
will be circumcised.}

\item{circ.W.prob}{Probablity that a white new arrival in the population
will be circumcised.}

\item{ccr5.B.prob}{Vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respectively) in the CCR5 gene
among black MSM.}

\item{ccr5.W.prob}{Vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respectively) in the CCR5 gene
among white MSM.}

\item{ccr5.heteroz.rr}{Relative risk of infection for men who are heterozygous
in the CCR5 mutation.}

\item{num.inst.ai.classes}{Number of quantiles into which men should be
divided in determining their levels of one-off anal intercourse.}

\item{base.ai.main.BB.rate}{Expected coital frequency in black-black main
partnerships (acts per day).}

\item{base.ai.main.BW.rate}{Expected coital frequency in black-white main
partnerships (acts per day).}

\item{base.ai.main.WW.rate}{Expected coital frequency in white-white main
partnerships (acts per day).}

\item{base.ai.pers.BB.rate}{Expected coital frequency in black-black casual
partnerships (acts per day).}

\item{base.ai.pers.BW.rate}{Expected coital frequency in black-white casual
partnerships (acts per day).}

\item{base.ai.pers.WW.rate}{Expected coital frequency in white-white casual
partnerships (acts per day).}

\item{ai.scale}{General relative scaler for all act rates for model
calibration.}

\item{cond.main.BB.prob}{Probability of condom use in a black-black main
partnership.}

\item{cond.main.BW.prob}{Probability of condom use in a black-white main
partnership.}

\item{cond.main.WW.prob}{Probability of condom use in a white-white main
partnership.}

\item{cond.pers.always.prob}{Fraction of men in casual partnerships who always
use condoms in those partnerships.}

\item{cond.pers.BB.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a black-black casual partnerships.}

\item{cond.pers.BW.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a black-white casual partnerships.}

\item{cond.pers.WW.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a white-white casual partnerships.}

\item{cond.inst.always.prob}{Fraction of men in instant partnerships who always
use condoms in those partnerships.}

\item{cond.inst.BB.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a black-black one-off partnerships.}

\item{cond.inst.BW.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a black-white one-off partnerships.}

\item{cond.inst.WW.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a white-white one-off partnerships.}

\item{cond.always.prob.corr}{Correlation coefficient for probability of always
using condoms in both casual and one-off}

\item{cond.rr.BB}{Condom probability scaler for black-black partnerships for
model calibration purposes.}

\item{cond.rr.BW}{Condom probability scaler for black-white partnerships for
model calibration purposes.}

\item{cond.rr.WW}{Condom probability scaler for white-white partnerships for
model calibration purposes.}

\item{cond.diag.main.beta}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.main.beta}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected man has disclosed.}

\item{cond.diag.pers.beta}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.pers.beta}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV-infected man has disclosed
his status.}

\item{cond.diag.inst.beta}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.inst.beta}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV-infected man has disclosed
his status.}

\item{vv.iev.BB.prob}{Probability that in a black-black partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{vv.iev.BW.prob}{Probability that in a black-white partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{vv.iev.WW.prob}{Probability that in a white-white partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{prep.start}{Time step at which the PrEP intervention should start.}

\item{prep.elig.model}{Modeling approach for determining who is eligible for
PrEP. Current options are limited to: \code{"all"} for all persons who
have never been on PrEP and are disease-susceptible.}

\item{prep.class.prob}{The probability of adherence class in non-adherent,
low adherence, medium adherence, or high adherence groups (from Liu).}

\item{prep.class.hr}{The hazard ratio for infection per act associated with each
level of adherence (from Grant).}

\item{prep.coverage}{The proportion of the eligible population who are start
PrEP once they become eligible.}

\item{prep.cov.method}{The method for calculating PrEP coverage, with options
of \code{"curr"} to base the numerator on the number of people currently
on PrEP and \code{"ever"} to base it on the number of people ever on
PrEP.}

\item{prep.cov.rate}{The rate at which persons initiate PrEP conditional on
their eligibility, with 1 equal to instant start.}

\item{prep.tst.int}{Testing interval for those who are actively on PrEP. This
overrides the mean testing interval parameters.}

\item{prep.risk.int}{Time window for assessment of risk eligibility for PrEP
in days.}

\item{prep.risk.reassess}{If \code{TRUE}, reassess eligibility for PrEP at
each testing visit.}

\item{stitest.active.int}{Interval in which individuals need to be sexually active
for STI Testing Guidelines}

\item{stitest.start}{Time step at which the STI testing guidelines intervention
should start.}

\item{ept.start}{Time step at which the EPT intervention should start}

\item{ept.risk.int}{Time window for assessment of risk eligibility for EPT in days}

\item{ept.coverage}{The proportion of the eligible population who are starting
EPT once they become eligible.}

\item{ept.cov.method}{The method for calculating EPT coverage, with options
of \code{"curr"} to base the numerator on the number of people currently
on EPT and \code{"ever"} to base it on the number of people ever on
EPT.}

\item{ept.cov.rate}{The rate at which persons initiate EPT conditional on
their eligibility, with 1 equal to instant start.}

\item{rcomp.prob}{Level of risk compensation from 0 to 1, where 0 is no risk
compensation, 0.5 is a 50% reduction in the probability of condom use
per act, and 1 is a complete cessation of condom use following PrEP
initiation.}

\item{rcomp.adh.groups}{PrEP adherence groups for whom risk compensation
occurs, as a vector with values 0, 1, 2, 3 corresponding to non-adherent,
low adherence, medium adherence, and high adherence to PrEP.}

\item{rcomp.main.only}{Logical, if risk compensation is limited to main
partnerships only, versus all partnerships.}

\item{rcomp.discl.only}{Logical, if risk compensation is limited known-discordant
partnerships only, versus all partnerships.}

\item{rgc.tprob}{Probability of rectal gonorrhea infection per act.}

\item{ugc.tprob}{Probability of urethral gonorrhea infection per act.}

\item{rct.tprob}{Probability of rectal chlamydia infection per act.}

\item{uct.tprob}{Probability of urethral chlamydia infection per act.}

\item{usyph.tprob}{Base probability of urethral syphilis infection per act.}

\item{rsyph.tprob}{Base probability of rectal syphilis infection per act.}

\item{syph.earlat.rr}{Multiplier for reduced infection probability in early latent 
stage of syphilis infection.}

\item{syph.late.rr}{Multiplier for reduced infection probability in late stages 
of syphilis infection.}

\item{rgc.sympt.prob}{Probability of symptoms given infection with rectal
gonorrhea.}

\item{ugc.sympt.prob}{Probability of symptoms given infection with urethral
gonorrhea.}

\item{rct.sympt.prob}{Probability of symptoms given infection with rectal
chlamydia.}

\item{uct.sympt.prob}{Probability of symptoms given infection with urethral
chlamydia.}

\item{syph.prim.sympt.prob}{Probability of symptoms given primary stage syphilis 
infection.}

\item{syph.seco.sympt.prob}{Probability of symptoms given secondary stage syphilis 
infection.}

\item{syph.earlat.sympt.prob}{Probability of symptoms given early latent stage 
syphilis infection.}

\item{syph.latelat.sympt.prob}{Probability of symptoms given late latent stage 
syphilis infection.}

\item{syph.tert.sympt.prob}{Probability of symptoms given tertiary stage syphilis 
infection.}

\item{rgc.asympt.int}{Average duration in days of asymptomatic rectal gonorrhea.}

\item{ugc.asympt.int}{Average duration in days of asymptomatic urethral gonorrhea.}

\item{gc.tx.int}{Average duration in days of treated gonorrhea (both sites).}

\item{gc.ntx.int}{Average duration in days of untreated, symptomatic gonorrhea (both sites).
If \code{NA}, uses site-specific durations for asymptomatic infections.}

\item{rct.asympt.int}{Average in days duration of asymptomatic rectal chlamydia.}

\item{uct.asympt.int}{Average in days duration of asymptomatic urethral chlamydia.}

\item{ct.tx.int}{Average in days duration of treated chlamydia (both sites).}

\item{ct.ntx.int}{Average in days duration of untreated, symptomatic chlamydia (both sites).
If \code{NA}, uses site-specific durations for asymptomatic infections.}

\item{syph.early.tx.int}{Average in days duration of treatment for early syphilis.}

\item{syph.late.tx.int}{Average in days duration of treatment for late syphilis.}

\item{gc.prob.cease}{Probability of ceasing sexual activity during symptomatic
infection with gonorrhea.}

\item{ct.prob.cease}{Probability of ceasing sexual activity during symptomatic
infection with chlamydia.}

\item{syph.prob.cease}{Probability of ceasing sexual activity during symptomatic
infection with syphilis.}

\item{gc.sympt.prob.tx}{Probability of treatment for symptomatic gonorrhea.}

\item{ct.sympt.prob.tx}{Probability of treatment for symptomatic chlamydia.}

\item{gc.asympt.prob.tx}{Probability of treatment, given diagnosis, for asymptomatic gonorrhea.}

\item{ct.asympt.prob.tx}{Probability of treatment, given diagnosis, for asymptomatic chlamydia.}

\item{syph.prim.sympt.prob.tx}{Probability of treatment for symptomatic primary
stage syphilis infection.}

\item{syph.prim.asympt.prob.tx}{Probability of treatment, given diagnosis, for asymptomatic primary
stage syphilis infection.}

\item{syph.seco.sympt.prob.tx}{Probability of treatment for symptomatic secondary
stage syphilis infection.}

\item{syph.seco.asympt.prob.tx}{Probability of treatment, given diagnosis, for asymptomatic secondary
stage syphilis infection.}

\item{syph.earlat.prob.tx}{Probability of treatment for symptomatic or
asymptomatic early latent stage syphilis infection.}

\item{syph.latelat.prob.tx}{Probability of treatment for symptomatic or
asymptomatic late latent stage syphilis infection.}

\item{syph.tert.sympt.prob.tx}{Probability of treatment for symptomatic tertiary
stage syphilis infection.}

\item{syph.tert.asympt.prob.tx}{Probability of treatment, given diagnosis, for asymptomatic tertiary
stage syphilis infection.}

\item{tst.rect.sti.rr}{Relative likelihood of rectal STI testing compared to urethral testing among those selected to be tested}

\item{sti.highrisktest.int}{Interval in which individuals need to be engaged in high-risk behavior for STI screening for high-risk testers}

\item{stianntest.coverage}{The proportion of the eligible population who are starting annual STI testing once they become eligible.
This is not inclusive of those who are simultaneously indicated for more frequent testing.}

\item{stianntest.cov.method}{The method for calculating STI annual testing, with options of \code{"curr"} to base the numerator 
on the number of people currently annually testing for STI and \code{"ever"} to base it on the number of people who have 
ever been annually tested for STI. This is not inclusive of those who are simultaneously indicated for more frequent testing.}

\item{stianntest.cov.rate}{The rate at which persons initiate annual STI testing conditional on their eligibility, with 1 equal to instant start.}

\item{stihighrisktest.coverage}{The proportion of the eligible population who are starting high-risk STI testing once they become eligible.}

\item{stihighrisktest.cov.method}{The method for calculating STI high-risk testing, with options of \code{"curr"} to base the numerator 
on the number of people currently high-risk testing for STI and \code{"ever"} to base it on the number of people who have 
ever been high-risk tested for STI.}

\item{stihighrisktest.cov.rate}{The rate at which persons initiate high-risk STI testing conditional on their eligibility, with 1 equal to instant start.}

\item{prep.sti.screen.int}{Interval in days between STI screening at PrEP visits.}

\item{prep.sti.prob.tx}{Probability of treatment given positive screening during
PrEP visit.}

\item{prep.continue.stand.tx}{Logical, if \code{TRUE} will continue standard
STI treatment of symptomatic cases even after PrEP initiation.}

\item{sti.cond.rr}{Relative risk of STI infection (in either direction) given
a condom used by the insertive partner.}

\item{hiv.rgc.rr}{Relative risk of HIV infection given current rectal gonorrhea in the HIV-negative partner.}

\item{hiv.ugc.rr}{Relative risk of HIV infection given current urethral gonorrhea in the HIV-negative partner.}

\item{hiv.rct.rr}{Relative risk of HIV infection given current rectal chlamydia in the HIV-negative partner.}

\item{hiv.uct.rr}{Relative risk of HIV infection given current urethral chlamydia in the HIV-negative partner.}

\item{hiv.dual.rr}{Additive proportional risk, from 0 to 1, for HIV infection
given dual infection with both gonorrhea and chlamydia.}

\item{hiv.rsyph.rr}{Relative risk of HIV infection given current rectal syphilis infection in the HIV-negative partner.}

\item{hiv.usyph.rr}{Relative risk of HIV infection given current urethral syphilis infection in the HIV-negative partner.}

\item{syph.rhiv.rr}{Relative risk of syphilis infection given current HIV infection in the receptive syphilis-negative partner.}

\item{syph.uhiv.rr}{Relative risk of syphilis infection given current HIV infection in the insertive syphilis-negative partner.}

\item{hivdx.syph.sympt.tx.rr}{Relative risk of treatment for symptomatic syphilis given diagnosed HIV infection. 
This is implemented as an inverse, 1 / this value for those who do not have diagnosed infection, so it is a boost to get to 0.35.}

\item{partlist.start}{Time step at which persisting edge list should begin
accumulating for eventual STI testing indications.}

\item{...}{Additional arguments passed to the function.}

\item{immune.syph.B.prob}{Prevalence of immunity (e.g. recovering from treated late syphilis) 
among black MSM not actively infected with syphilis.}

\item{immune.syph.W.prob}{Prevalence of immunity among black MSM not actively infected with
syphilis.}

\item{immune.syph.int}{Number of days that temporary immunity from syphilis
infection lasts after treatment and recovery from late syphilis infection.}

\item{syph.immune.rr}{Multiplier for reduced infection probability given an 
immune partner.}
}
\value{
A list object of class \code{param_msm}, which can be passed to
EpiModel function \code{netsim}.
}
\description{
Sets the epidemic parameters for stochastic network models
             simulated with \code{\link{netsim}} for EpiModelHIV
}
\keyword{msm}
