% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.prep.adol.R
\name{prep_adol}
\alias{prep_adol}
\title{PrEP ASMM Module}
\usage{
prep_adol(dat, at)
}
\description{
Module function for implementation and uptake of pre-exposure
             prophylaxis (PrEP) among ASMM to prevent HIV infection.
}
