% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.initialize.R
\name{reinit_msm}
\alias{reinit_msm}
\title{Re-Initialization Module}
\usage{
reinit_msm(x, param, init, control, s)
}
\arguments{
\item{x}{An \code{EpiModel} object of class \code{\link{netsim}}.}

\item{param}{An \code{EpiModel} object of class \code{\link{param_msm}}.}

\item{init}{An \code{EpiModel} object of class \code{\link{init_msm}}.}

\item{control}{An \code{EpiModel} object of class \code{\link{control_msm}}.}

\item{s}{Simulation number, used for restarting dependent simulations.}
}
\value{
This function resets the data elements on the \code{dat} master data object
in the needed ways for the time loop to function.
}
\description{
This function reinitializes an epidemic model to restart at a
             specified time step given an input \code{netsim} object.
}
\details{
Currently, the necessary components that must be on \code{x} for a simulation
to be restarted must be: param, control, nwparam, epi, attr, temp, el, p.
TODO: describe this more.
}
\keyword{module}
\keyword{msm}
