% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.ept.R
\name{sti_ept_msm}
\alias{sti_ept_msm}
\title{EPT Module}
\usage{
sti_ept_msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function for implementation of expedited
             partner therapy (EPT) in index partner to prevent 
             STI infection. Eligbility is handled in the risk history
             module.
}
\keyword{module}
\keyword{msm}
